Hydra.API.APIServerLog

Documentation

data APIServerLog

Constructors

APIServerStarted

Fields

listeningPort :: PortNumber

NewAPIConnection

APIOutputSent

Fields

sentOutput :: Value

APIInputReceived

Fields

receivedInput :: Value

APIInvalidInput

Fields

reason :: String

inputReceived :: Text

APIConnectionError

Fields

reason :: String

APIHTTPRequestReceived

Fields

method :: Method

path :: PathInfo

APITransactionSubmitted

Fields

submittedTxId :: String

newtype PathInfo

New type wrapper to define JSON instances.

Constructors

PathInfo ByteString

newtype Method

New type wrapper to define JSON instances.

NOTE: We are not using http-types StdMethod as we do not want to be constrained in terms of logging and accept any method in a Request.

Constructors

Method ByteString