Hydra.API.Server

Documentation

data APIServerLog

Constructors

APIServerStarted

Fields

listeningPort :: PortNumber

NewAPIConnection

APIOutputSent

Fields

sentOutput :: Value

APIInputReceived

Fields

receivedInput :: Value

APIInvalidInput

Fields

reason :: String

inputReceived :: Text

APIConnectionError

Fields

reason :: String

APIHandshakeError

Fields

reason :: String

APIRestInputReceived

Fields

method :: Text

paths :: [Text]

requestInputBody :: Maybe Value

newtype Server tx m

Handle to provide a means for sending server outputs to clients.

Constructors

Server

Fields

sendOutput :: ServerOutput tx -> m ()

Send some output to all connected clients.

type ServerCallback tx m = ClientInput tx -> m ()

Callback for receiving client inputs.

type ServerComponent tx m a = ServerCallback tx m -> (Server tx m -> m a) -> m a

A type tying both receiving input and sending output into a Component.

withAPIServer :: forall tx. IsChainState tx => IP -> PortNumber -> Party -> PersistenceIncremental (TimedServerOutput tx) IO -> Tracer IO APIServerLog -> Chain tx IO -> ServerComponent tx IO ()

nextSequenceNumber :: TVar [TimedServerOutput tx] -> STM Natural

type NotifyServerRunning = IO ()

type WaitForServer = IO ()

setupServerNotification :: IO (NotifyServerRunning, WaitForServer)

Setup notification and waiter to ensure that something only runs after the server is actually listening.

runAPIServer

Arguments

:: forall tx. IsChainState tx

=> IP

-> PortNumber

-> Party

-> Tracer IO APIServerLog

-> TVar [TimedServerOutput tx]

-> Chain tx IO

-> (ClientInput tx -> IO ())

-> Projection STM (ServerOutput tx) HeadStatus
Read model to enhance Greetings messages with HeadStatus.

-> Projection STM (ServerOutput tx) (Maybe (UTxOType tx))
Read model to enhance Greetings messages with snapshot UTxO.

-> TChan (TimedServerOutput tx)

-> NotifyServerRunning
Called when the server is listening before entering the main loop.

-> IO ()

data RunServerException

Constructors

RunServerException

Fields

ioException :: IOException

host :: IP

port :: PortNumber

handleDraftCommitUtxo :: Chain tx IO -> Tracer IO APIServerLog -> ByteString -> Method -> [Text] -> (Response -> IO ResponseReceived) -> IO ResponseReceived