Hydra.API.ServerOutput

Documentation

data TimedServerOutput tx

The type of messages sent to clients by the Server.

Constructors

TimedServerOutput

Fields

output :: ServerOutput tx

seq :: Natural

time :: UTCTime

data ServerOutput tx

Individual server output messages as produced by the HeadLogic in the ClientEffect.

Constructors

PeerConnected

Fields

peer :: NodeId

PeerDisconnected

Fields

peer :: NodeId

ReadyToCommit

Fields

parties :: Set Party

Committed

Fields

party :: Party

utxo :: UTxOType tx

HeadIsOpen

Fields

utxo :: UTxOType tx

HeadIsClosed

Fields

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

Nominal deadline until which contest can be submitted and after which fanout is possible. NOTE: Use this only for informational purpose and wait for ReadyToFanout instead before sending Fanout as the ledger of our cardano-node might not have progressed sufficiently in time yet and we do not re-submit transactions (yet).

HeadIsContested

Fields

snapshotNumber :: SnapshotNumber

ReadyToFanout

HeadIsAborted

Fields

utxo :: UTxOType tx

HeadIsFinalized

Fields

utxo :: UTxOType tx

CommandFailed

Fields

clientInput :: ClientInput tx

TxSeen

Fields

transaction :: tx

TxValid

Fields

transaction :: tx

TxInvalid

Fields

utxo :: UTxOType tx

transaction :: tx

validationError :: ValidationError

TxExpired

Fields

transaction :: tx

SnapshotConfirmed

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

GetUTxOResponse

Fields

utxo :: UTxOType tx

InvalidInput

Fields

reason :: String

input :: Text

Greetings
A friendly welcome message which tells a client something about the node. Currently used for knowing what signing key the server uses (it only knows one).

Fields

me :: Party

PostTxOnChainFailed

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

RolledBack