Hydra.Chain

Description

Specifies the Head-Chain Interaction part of the protocol

Incoming and outgoing on-chain transactions are modelled respectively as OnChainTx and PostChainTx which are data type that abstracts away the details of the structure of the transaction.

Documentation

data HeadParameters

Contains the head's parameters as established in the initial transaction.

Constructors

HeadParameters

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

data PostChainTx tx

Data type used to post transactions on chain. It holds everything to construct corresponding Head protocol transactions.

Constructors

InitTx

Fields

headParameters :: HeadParameters

CommitTx

Fields

party :: Party

committed :: UTxOType tx

AbortTx

Fields

utxo :: UTxOType tx

CollectComTx

Fields

utxo :: UTxOType tx

CloseTx

Fields

confirmedSnapshot :: ConfirmedSnapshot tx

ContestTx

Fields

confirmedSnapshot :: ConfirmedSnapshot tx

FanoutTx

Fields

utxo :: UTxOType tx

contestationDeadline :: UTCTime

newtype HeadId

Uniquely identifies a Hydra Head.

Constructors

HeadId ByteString

data OnChainTx tx

Describes transactions as seen on chain. Holds as minimal information as possible to simplify observing the chain.

Constructors

OnInitTx

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

OnCommitTx

Fields

party :: Party

committed :: UTxOType tx

OnAbortTx

OnCollectComTx

OnCloseTx

Fields

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

OnContestTx

Fields

snapshotNumber :: SnapshotNumber

OnFanoutTx

data PostTxError tx

Exceptions thrown by $sel:postTx:Chain.

Constructors

MoreThanOneUTxOCommitted

InternalWalletError

Fields

headUTxO :: UTxOType tx

reason :: Text

tx :: tx

NotEnoughFuel

CannotFindOwnInitial

Fields

knownUTxO :: UTxOType tx

FailedToPostTx

Fields

failureReason :: Text

PlutusValidationFailed

Fields

plutusFailure :: Text

plutusDebugInfo :: Text

NoSeedInput

NoPaymentInput

InvalidStateToPost

Fields

txTried :: PostChainTx tx

chainState :: ChainStateType tx

UnsupportedLegacyOutput

Fields

byronAddress :: Address ByronAddr

class (IsTx tx, Eq (ChainStateType tx), Show (ChainStateType tx), FromJSON (ChainStateType tx), ToJSON (ChainStateType tx)) => IsChainState tx where

Interface available from a chain state. Expected to be instantiated by all 'ChainStateType tx'.

Associated Types

type ChainStateType tx = c | c -> tx

Types of what to keep as L1 chain state.

Methods

chainStateSlot :: ChainStateType tx -> ChainSlot

Get the chain slot for a chain state. NOTE: For any sequence of a encountered, we assume monotonically increasing slots.

newtype ChainSlot

A generic description for a chain slot all implementions need to use.

Constructors

ChainSlot Natural

nextChainSlot :: ChainSlot -> ChainSlot

Get the next chain slot. Use this instead of giving Enum or Num instances to ChainSlot.

newtype Chain tx m

Handle to interface with the main chain network

Constructors

Chain

Fields

postTx :: (IsChainState tx, MonadThrow m) => ChainStateType tx -> PostChainTx tx -> m ()

Construct and send a transaction to the main chain corresponding to the given PostChainTx description and the current ChainState. This function is not expected to block, so it is only responsible for submitting, but it should validate the created transaction against a reasonable local view of the chain and throw an exception when invalid.

Does at least throw PostTxError.

data ChainEvent tx

Constructors

Observation

Fields

observedTx :: OnChainTx tx

newChainState :: ChainStateType tx

Rollback ChainSlot

Tick UTCTime

type ChainCallback tx m = (ChainStateType tx -> Maybe (ChainEvent tx)) -> m ()

A callback indicating receival of a potential Hydra transaction which is Maybe observing a relevant 'ChainEvent tx'.

type ChainComponent tx m a = ChainCallback tx m -> (Chain tx m -> m a) -> m a

A type tying both posting and observing transactions into a single Component.