Hydra.Chain.Blockfrost.Client

Documentation

data APIBlockfrostError

Constructors

BlockfrostError Text

DecodeError Text

runBlockfrostM :: (MonadIO m, MonadThrow m) => Project -> BlockfrostClientT IO a -> m a

publishHydraScripts

Arguments

:: FilePath
The path where the Blockfrost project token hash is stored.

-> SigningKey PaymentKey
Keys assumed to hold funds to pay for the publishing transaction.

-> IO [TxId]

scriptTypeToPlutusVersion :: ScriptType -> Maybe Language

toCardanoPoolId :: PoolId -> Hash StakePoolKey

toCardanoUTxO :: [AddressUtxo] -> AddressInEra -> UTxO' (TxOut CtxUTxO)

toCardanoTxIn :: AddressUtxo -> TxIn

toCardanoTxOut :: AddressUtxo -> AddressInEra -> TxOut CtxUTxO

toCardanoPolicyId :: Text -> PolicyId

toCardanoAssetName :: Text -> AssetName

toCardanoValue :: [Amount] -> Value

Helpers

unwrapAddress :: AddressInEra -> Text

textAddrOf :: NetworkId -> VerificationKey PaymentKey -> Text

toCardanoNetworkMagic :: Integer -> NetworkId

data BlockfrostConversion

Constructors

BlockfrostConversion

Fields

a0 :: NonNegativeInterval

rho :: UnitInterval

tau :: UnitInterval

priceMemory :: NonNegativeInterval

priceSteps :: NonNegativeInterval

pvtMotionNoConfidence :: UnitInterval

pvtCommitteeNormal :: UnitInterval

pvtCommitteeNoConfidence :: UnitInterval

pvtHardForkInitiation :: UnitInterval

pvtPPSecurityGroup :: UnitInterval

dvtMotionNoConfidence :: UnitInterval

dvtCommitteeNormal :: UnitInterval

dvtCommitteeNoConfidence :: UnitInterval

dvtUpdateToConstitution :: UnitInterval

dvtHardForkInitiation :: UnitInterval

dvtPPNetworkGroup :: UnitInterval

dvtPPEconomicGroup :: UnitInterval

dvtPPTechnicalGroup :: UnitInterval

dvtPPGovGroup :: UnitInterval

dvtTreasuryWithdrawal :: UnitInterval

committeeMinSize :: Quantity

committeeMaxTermLength :: Quantity

govActionLifetime :: Quantity

govActionDeposit :: Coin

drepDeposit :: Integer

drepActivity :: Quantity

minFeeRefScriptCostPerByte :: NonNegativeInterval

toCardanoPParams :: MonadIO m => BlockfrostClientT m (PParams LedgerEra)

toCardanoGenesisParameters :: Genesis -> GenesisParameters ShelleyEra

mkEraHistory :: Genesis -> EraHistory