Hydra.Chain.CardanoClient

Description

A basic cardano-node client that can talk to a local cardano-node.

The idea of this module is to provide a Haskell interface on top of cardano-cli's API, using cardano-api types.

Documentation

type NodeSocket = FilePath

data QueryException

Constructors

QueryAcquireException AcquireFailure

QueryEraMismatchException EraMismatch

CardanoClient handle

data CardanoClient

Handle interface for abstract querying of a cardano node.

Constructors

CardanoClient

Fields

queryUTxOByAddress :: [Address ShelleyAddr] -> IO UTxO

networkId :: NetworkId

mkCardanoClient :: NetworkId -> NodeSocket -> CardanoClient

Construct a CardanoClient handle.

Tx Construction / Submission

buildTransaction

Arguments

:: NetworkId
Current network identifier

-> FilePath
Filepath to the cardano-node's domain socket

-> AddressInEra
Change address to send

-> UTxO
Unspent transaction outputs to spend.

-> [TxIn]
Collateral inputs.

-> [TxOut CtxTx]
Outputs to create.

-> IO (Either TxBodyErrorAutoBalance TxBody)

Construct a simple payment consuming some inputs and producing some outputs (no certificates or withdrawals involved).

On success, the returned transaction is fully balanced. On error, return TxBodyErrorAutoBalance.

submitTransaction

Arguments

:: NetworkId
Current network discriminant

-> FilePath
Filepath to the cardano-node's domain socket

-> Tx
A signed transaction.

-> IO ()

Submit a (signed) transaction to the node.

Throws SubmitTransactionException if submission fails.

data SubmitTransactionException

Exceptions that can occur during a transaction submission.

In principle, we can only encounter an EraMismatch at era boundaries, when we try to submit a "next era" transaction as a "current era" transaction, or vice-versa. Similarly, TxValidationError shouldn't occur given that the transaction was safely constructed through buildTransaction.

Constructors

SubmitEraMismatch EraMismatch

SubmitTxValidationError (TxValidationErrorInMode CardanoMode)

awaitTransaction

Arguments

:: NetworkId
Current network discriminant

-> FilePath
Filepath to the cardano-node's domain socket

-> Tx
The transaction to watch / await

-> IO UTxO

Await until the given transaction is visible on-chain. Returns the UTxO set produced by that transaction.

Note that this function loops forever; hence, one probably wants to couple it with a surrounding timeout.

Local state query

data QueryPoint

Describes whether to query at the tip or at a specific point.

Constructors

QueryTip

QueryAt ChainPoint

queryTip :: NetworkId -> FilePath -> IO ChainPoint

Query the latest chain point aka "the tip".

queryTipSlotNo :: NetworkId -> FilePath -> IO SlotNo

Query the latest chain point just for the slot number.

querySystemStart :: NetworkId -> FilePath -> QueryPoint -> IO SystemStart

Query the system start parameter at given point.

Throws at least QueryException if query fails.

queryEraHistory :: NetworkId -> FilePath -> QueryPoint -> IO (EraHistory CardanoMode)

Query the era history at given point.

Throws at least QueryException if query fails.

queryProtocolParameters :: NetworkId -> FilePath -> QueryPoint -> IO ProtocolParameters

Query the protocol parameters at given point.

Throws at least QueryException if query fails.

queryUTxO :: NetworkId -> FilePath -> QueryPoint -> [Address ShelleyAddr] -> IO UTxO

Query UTxO for all given addresses at given point.

Throws at least QueryException if query fails.

queryUTxOByTxIn :: NetworkId -> FilePath -> QueryPoint -> [TxIn] -> IO UTxO

Query UTxO for given tx inputs at given point.

Throws at least QueryException if query fails.

queryUTxOWhole :: NetworkId -> FilePath -> QueryPoint -> IO UTxO

Query the whole UTxO from node at given point. Useful for debugging, but should obviously not be used in production code.

Throws at least QueryException if query fails.

queryUTxOFor :: NetworkId -> FilePath -> QueryPoint -> VerificationKey PaymentKey -> IO UTxO

Query UTxO for the address of given verification key at point.

Throws at least QueryException if query fails.

queryStakePools :: NetworkId -> FilePath -> QueryPoint -> IO (Set PoolId)

Query the current set of registered stake pools.

Throws at least QueryException if query fails.

runQuery :: NetworkId -> FilePath -> QueryPoint -> QueryInMode CardanoMode a -> IO a

Throws at least QueryException if query fails.

Helpers

throwOnEraMismatch :: MonadThrow m => Either EraMismatch a -> m a

localNodeConnectInfo :: NetworkId -> FilePath -> LocalNodeConnectInfo CardanoMode

cardanoModeParams :: ConsensusModeParams CardanoMode