Hydra.Chain.Direct

Description

Chain component implementation which uses directly the Node-to-Client protocols to submit "hand-rolled" transactions including Plutus validators and observing the chain using it as well.

Documentation

newtype NetworkMagic

Constructors

NetworkMagic Word32

withIOManager :: WithIOManager

data DirectChainLog tx

Constructors

PostTx

Fields

toPost :: PostChainTx tx

postedTx :: ValidatedTx Era

ReceiveTxs

Fields

receivedTxs :: [ValidatedTx Era]

onChainTxs :: [OnChainTx tx]

RolledBackward

Fields

point :: Point Block

Wallet TinyWalletLog

data ConnectException

Constructors

ConnectException

Fields

ioException :: IOException

socketPath :: FilePath

networkMagic :: NetworkMagic

withDirectChain

Arguments

:: Tracer IO (DirectChainLog tx)
Tracer for logging

-> NetworkMagic
Network identifer to which we expect to connect.

-> IOManager
A cross-platform abstraction for managing I/O operations on local sockets

-> FilePath
Path to a domain socket used to connect to the server.

-> (VerificationKey, SigningKey)
Key pair for the wallet

-> ChainComponent tx IO ()

client :: (MonadST m, MonadTimer m) => Tracer m (DirectChainLog tx) -> TQueue m (PostChainTx tx) -> TVar m OnChainHeadState -> TinyWallet m -> ChainCallback tx m -> NodeToClientVersion -> OuroborosApplication 'InitiatorMode LocalAddress LByteString m () Void

chainSyncClient :: forall m tx. MonadSTM m => Tracer m (DirectChainLog tx) -> ChainCallback tx m -> TVar m OnChainHeadState -> ChainSyncClient Block (Point Block) (Tip Block) m ()

txSubmissionClient :: forall m tx. MonadSTM m => Tracer m (DirectChainLog tx) -> TQueue m (PostChainTx tx) -> TVar m OnChainHeadState -> TinyWallet m -> LocalTxSubmissionClient (GenTx Block) (ApplyTxErr Block) m ()

getAlonzoTxs :: Block -> StrictSeq (ValidatedTx Era)

This extract Alonzo transactions from a block. If the block wasn't produced in the Alonzo era, it returns a empty sequence.