Hydra.Chain.Direct.Context

Documentation

data HydraContext

Define some global context from which generators can pick values for generation. This allows to write fairly independent generators which however still make sense with one another within the context of a head.

For example, one can generate a head's _party_ from that global list, whereas other functions may rely on all parties and thus, we need both generation to be coherent.

Constructors

HydraContext

Fields

ctxVerificationKeys :: [VerificationKey PaymentKey]

ctxHydraSigningKeys :: [SigningKey]

ctxNetworkId :: NetworkId

ctxContestationPeriod :: ContestationPeriod

ctxParties :: HydraContext -> [Party]

ctxHeadParameters :: HydraContext -> HeadParameters

genHydraContext :: Int -> Gen HydraContext

Generate a HydraContext for a bounded arbitrary number of parties.

maxParties sets the upper bound in the number of parties in the Head.

genHydraContextFor :: Int -> Gen HydraContext

Generate a HydraContext for a given number of parties.

genStIdle :: HydraContext -> Gen (OnChainHeadState 'StIdle)

genStInitialized :: HydraContext -> Gen (OnChainHeadState 'StInitialized)

genInitTx :: HydraContext -> Gen Tx

genCommits :: HydraContext -> Tx -> Gen [Tx]

genCommit :: Gen UTxO

genCloseTx :: Int -> Gen (OnChainHeadState 'StOpen, Tx, ConfirmedSnapshot Tx)

genFanoutTx :: Int -> Int -> Gen (OnChainHeadState 'StClosed, Tx)

genStOpen :: HydraContext -> Gen (UTxO, OnChainHeadState 'StOpen)

genStClosed :: HydraContext -> UTxO -> Gen (SnapshotNumber, UTxO, OnChainHeadState 'StClosed)

unsafeObserveTx :: forall st st'. (ObserveTx st st', HasCallStack) => Tx -> OnChainHeadState st -> (OnChainTx Tx, OnChainHeadState st')

unsafeCommit :: HasCallStack => UTxO -> OnChainHeadState 'StInitialized -> Tx

executeCommits :: Tx -> [Tx] -> OnChainHeadState 'StIdle -> ([UTxO], OnChainHeadState 'StInitialized)