Hydra.Chain.Direct.Fixture

Description

Generic Cardano constants for use in testing.

Documentation

testNetworkId :: NetworkId

testPolicyId :: PolicyId

testSeedInput :: TxIn

defaultLedgerEnv :: LedgerEnv LedgerEra

Default environment for the L2 ledger using the fixed L1 pparams with zeroed fees and prices. NOTE: This is using still a constant SlotNo = 1.

defaultPParams :: ProtocolParameters

defaultGlobals :: Globals

pparams :: ProtocolParameters

Current (2023-04-12) mainchain protocol parameters. XXX: Avoid specifiying not required parameters here (e.g. max block units should not matter). XXX: Load and use mainnet parameters from a file which we can easily review to be in sync with mainnet.

systemStart :: SystemStart

epochInfo :: Monad m => EpochInfo m

An artifical EpochInfo comprised by a single never ending (forking) era, with fixed epochSize and slotLength.