Hydra.Chain.Direct.Handlers

Description

Provide infrastructure-independent "handlers" for posting transactions and following the chain.

This module encapsulates the transformation logic between cardano transactions and HydraNode abstractions PostChainTx and OnChainTx, and maintainance of on-chain relevant state.

Posting Transactions

type SubmitTx m = ValidatedTx LedgerEra -> m ()

A callback used to actually submit a transaction to the chain.

mkChain :: (MonadSTM m, MonadTimer m, MonadThrow (STM m)) => Tracer m DirectChainLog -> m TimeHandle -> [VerificationKey PaymentKey] -> TinyWallet m -> TVar m SomeOnChainHeadStateAt -> SubmitTx m -> Chain Tx m

Create a Chain component for posting "real" cardano transactions.

This component does not actually interact with a cardano-node, but creates cardano transactions from PostChainTx transactions emitted by a HydraNode, balancing them using given TinyWallet, while maintaining some state in the headState variable and before handing it off to the given SubmitTx callback.

NOTE: Given the constraints on m this function should work within IOSim and does not require any actual IO to happen which makes it highly suitable for simulations and testing.

data TimeHandle

Constructors

TimeHandle

Fields

currentPointInTime :: Either Text PointInTime

Get the current PointInTime

adjustPointInTime :: SlotNo -> PointInTime -> Either Text PointInTime

Adjust a PointInTime by some number of slots, positively or negatively.

finalizeTx :: (MonadSTM m, MonadThrow (STM m)) => TinyWallet m -> TVar m SomeOnChainHeadStateAt -> ValidatedTx LedgerEra -> STM m (ValidatedTx LedgerEra)

Balance and sign the given partial transaction.

Following the Chain

data SomeOnChainHeadStateAt

The on-chain head state is maintained in a mutually-recursive data-structure, pointing to the previous chain state; This allows to easily rewind the state to a point in the past when rolling backward due to a change of chain fork by our local cardano-node.

Constructors

SomeOnChainHeadStateAt

Fields

currentOnChainHeadState :: SomeOnChainHeadState

recordedAt :: RecordedAt

data RecordedAt

Records when a state was seen on-chain. AtStart is used for states that simply exist out of any chain events (e.g. the Idle state).

Constructors

AtStart

AtPoint ChainPoint SomeOnChainHeadStateAt

data ChainSyncHandler m

A handler that takes care of following the chain. TODO: replace use of Block with Tx

Constructors

ChainSyncHandler

Fields

onRollForward :: Block -> m ()

onRollBackward :: Point Block -> m ()

chainSyncHandler

Arguments

:: forall m. (MonadSTM m, MonadTime m)

=> Tracer m DirectChainLog
Tracer for logging

-> (ChainEvent Tx -> m ())
Chain callback

-> TVar m SomeOnChainHeadStateAt
On-chain head-state.

-> ChainSyncHandler m
A chain-sync handler to use in a local-chain-sync client.

Creates a ChainSyncHandler that can notify the given callback of events happening on-chain.

This forms the other half of a ChainComponent along with mkChain but is decoupled from actual interactions with the chain.

rollback :: ChainPoint -> SomeOnChainHeadStateAt -> (SomeOnChainHeadStateAt, Word)

Rewind some head state back to the first known state that is strictly before the provided ChainPoint.

This function also computes the rollback depth, e.g the number of observed states that needs to be discarded given some rollback chainPoint. This depth is used in rollback to discard corresponding off-chain state. Consequently the two states must be kept in sync in order for rollbacks to be handled properly.

closeGraceTime :: SlotNo

Hardcoded grace time for close transaction to be valid. TODO: make it a node configuration parameter

fromPostChainTx :: (MonadSTM m, MonadThrow (STM m)) => TimeHandle -> [VerificationKey PaymentKey] -> TinyWallet m -> TVar m SomeOnChainHeadStateAt -> PostChainTx Tx -> STM m Tx

getBabbageTxs :: Block -> StrictSeq (ValidatedTx LedgerEra)

This extract Babbage transactions from a block. If the block wasn't produced in the Babbage era, it returns an empty sequence.

data DirectChainLog

Constructors

ToPost

Fields

toPost :: PostChainTx Tx

PostingTx

Fields

postedTx :: (TxId StandardCrypto, ValidatedTx LedgerEra)

PostedTx

Fields

postedTxId :: TxId StandardCrypto

ReceivedTxs

Fields

onChainTxs :: [OnChainTx Tx]

receivedTxs :: [(TxId StandardCrypto, ValidatedTx LedgerEra)]

RolledBackward

Fields

point :: SomePoint

Wallet TinyWalletLog