Hydra.Chain.Direct.ScriptRegistry

Description

A data-type to keep track of reference Hydra scripts published on-chain, and needed to construct transactions leveraging reference inputs.

Documentation

data ScriptRegistry

Hydra scripts published as reference scripts at these UTxO.

Constructors

ScriptRegistry

Fields

initialReference :: (TxIn, TxOut CtxUTxO)

commitReference :: (TxIn, TxOut CtxUTxO)

data NewScriptRegistryException

Constructors

MissingScript

Fields

scriptName :: Text

scriptHash :: ScriptHash

discoveredScripts :: Set ScriptHash

newScriptRegistry :: UTxO -> Either NewScriptRegistryException ScriptRegistry

Create a script registry from a UTxO containing outputs with reference scripts. This will return Nothing if one or all of the references could not be found.

registryUTxO :: ScriptRegistry -> UTxO

Get the UTxO that corresponds to a script registry.

*Property**:

newScriptRegistry (registryUTxO r) === Just r

queryScriptRegistry :: (MonadIO m, MonadThrow m) => NetworkId -> FilePath -> TxId -> m ScriptRegistry

Query for TxIns in the search for outputs containing all the reference scripts of the ScriptRegistry.

This is implemented by repeated querying until we have all necessary reference scripts as we do only know the transaction id, not the indices.

NOTE: This is limited to an upper bound of 10 to not query too much before providing an error.

NOTE: If this should change, make sure to update the command line help.

Can throw at least NewScriptRegistryException on failure.

genScriptRegistry :: Gen ScriptRegistry

publishHydraScripts

Arguments

:: NetworkId
Expected network discriminant.

-> FilePath
Path to the cardano-node's domain socket

-> SigningKey PaymentKey
Keys assumed to hold funds to pay for the publishing transaction.

-> IO TxId