Hydra.Chain.Direct.State

OnChainHeadState

data OnChainHeadState (st :: HeadStateKind)

An opaque on-chain head state, which records information and events happening on the layer-1 for a given Hydra head.

data HeadStateKind

Some Kind for witnessing Hydra state-machine's states at the type-level.

This is useful to

(a) Reads code evolving the state machine, as it makes transition more obvious from type-signatures; (b) Pattern-match on the HydraStateMachine without having to bother with non-reachable cases.

Constructors

StIdle

StInitialized

StOpen

StClosed

class HeadStateKindVal (st :: HeadStateKind) where

Methods

headStateKindVal :: Proxy st -> HeadStateKind

getKnownUTxO :: OnChainHeadState st -> UTxO

Working with opaque states

data SomeOnChainHeadState where

An existential wrapping some OnChainHeadState into a value that carry no type-level information about the state.

Constructors

SomeOnChainHeadState :: forall st. HasTransition st => OnChainHeadState st -> SomeOnChainHeadState

data TokHeadState (st :: HeadStateKind) where

A token witnessing the state's type of an OnChainHeadState. See reifyState

Constructors

TkIdle :: TokHeadState 'StIdle

TkInitialized :: TokHeadState 'StInitialized

TkOpen :: TokHeadState 'StOpen

TkClosed :: TokHeadState 'StClosed

reifyState :: forall st. OnChainHeadState st -> TokHeadState st

Reify a HeadStateKind kind into a value to enable pattern-matching on existentials.

Initializing

idleOnChainHeadState :: NetworkId -> [VerificationKey PaymentKey] -> VerificationKey PaymentKey -> Party -> OnChainHeadState 'StIdle

Initialize a new OnChainHeadState.

Constructing transitions

initialize :: HeadParameters -> [VerificationKey PaymentKey] -> TxIn -> OnChainHeadState 'StIdle -> Tx

Initialize a head from an StIdle state. This does not change the state but produces a transaction which, if observed, does apply the transition.

abort :: HasCallStack => OnChainHeadState 'StInitialized -> Tx

commit :: UTxO -> OnChainHeadState 'StInitialized -> Either (PostTxError Tx) Tx

collect :: OnChainHeadState 'StInitialized -> Tx

close :: ConfirmedSnapshot Tx -> PointInTime -> OnChainHeadState 'StOpen -> Tx

contest :: ConfirmedSnapshot Tx -> PointInTime -> OnChainHeadState 'StClosed -> Tx

fanout :: UTxO -> PointInTime -> OnChainHeadState 'StClosed -> Tx

Observing transitions

observeSomeTx :: Tx -> SomeOnChainHeadState -> Maybe (OnChainTx Tx, SomeOnChainHeadState)

A convenient way to apply transition to SomeOnChainHeadState without bothering about the internal details.

Internal API

class HasTransition st => ObserveTx (st :: HeadStateKind) (st' :: HeadStateKind) where

A class for describing a Hydra transition from a state to another.

The transition is encoded at the type-level through the HeadStateKind and the function transition overloaded for all transitions.

Methods

observeTx :: Tx -> OnChainHeadState st -> Maybe (OnChainTx Tx, OnChainHeadState st')

class HasTransition (st :: HeadStateKind) where

A convenient class to declare all possible transitions from a given starting state st. This is useful to embed OnChainHeadState with an existential that carries some capabilities in the form of transitions (e.g. SomeOnChainHeadState).

Methods

transitions :: Proxy st -> [TransitionFrom st]

data TransitionFrom st where

Constructors

TransitionTo :: forall st st'. (ObserveTx st st', HasTransition st', HeadStateKindVal st, HeadStateKindVal st') => Proxy st' -> TransitionFrom st

getContestationDeadline :: OnChainHeadState 'StClosed -> POSIXTime