Hydra.Chain.Direct.TimeHandle

Description

Module to deal with time in direct cardano chain layer. Defines the type for a PointInTime and a means to acquire one via a TimeHandle and queryTimeHandle.

Documentation

type PointInTime = (SlotNo, UTCTime)

data TimeHandle

Constructors

TimeHandle

Fields

currentPointInTime :: Either Text PointInTime

Get the current PointInTime

slotFromUTCTime :: UTCTime -> Either Text SlotNo

Lookup slot number given a UTCTime. This will fail if the time is outside the "safe zone".

slotToUTCTime :: SlotNo -> Either Text UTCTime

Convert a slot number to a UTCTime using the stored epoch info. This will fail if the slot is outside the "safe zone".

adjustPointInTime :: SlotNo -> PointInTime -> Either Text PointInTime

Adjust a PointInTime by some number of slots, positively or negatively.

genTimeParams :: Gen (SystemStart, EraHistory CardanoMode, SlotNo, UTCTime)

Generate consistent values for SystemStart and EraHistory which has a horizon at the returned SlotNo as well as some UTCTime before that

mkTimeHandle :: UTCTime -> SystemStart -> EraHistory CardanoMode -> TimeHandle

Construct a time handle using current time and given chain parameters. See queryTimeHandle to create one by querying a cardano-node.

queryTimeHandle :: NetworkId -> FilePath -> IO TimeHandle

Query node for system start and era history before constructing a TimeHandle using getCurrentTime.