Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Documentation

network :: Network

type Era = AlonzoEra StandardCrypto

Post Hydra Head transactions

data OnChainHeadState

Maintains information needed to construct on-chain transactions depending on the current state of the head.

Constructors

Closed

Initial

Fields

threadOutput :: (TxIn StandardCrypto, TxOut Era, AssetClass, HeadParameters)

The state machine UTxO produced by the Init transaction This output should always be present and threaded across all transactions.

initials :: [(TxIn StandardCrypto, PubKeyHash)]

Final

threadToken :: AssetClass

initTx :: HeadParameters -> TxIn StandardCrypto -> ValidatedTx Era

Create the init transaction from some HeadParameters and a single TxIn which will be used as unique parameter for minting NFTs.

abortTx

Arguments

:: (TxIn StandardCrypto, AssetClass, HeadParameters)
The data needed to consume the state-machine output

-> [(TxIn StandardCrypto, PubKeyHash)]
Data needed to consume the inital output sent to each party to the Head which should contain the PT and is locked by initial script

-> ValidatedTx Era

Create transaction which aborts by spending one input. This is currently only possible if this is governed by the initial script and only for a single input. Of course, the Head protocol specifies we need to spend ALL the Utxo containing PTs.

Observe Hydra Head transactions

runOnChainTxs :: forall m tx. MonadSTM m => TVar m OnChainHeadState -> [ValidatedTx Era] -> m [OnChainTx tx]

Update observable on-chain head state from on-chain transactions. NOTE(AB): I tried to separate the 2 functions, the one working on list of txs and the one working on single tx but I keep getting failed unification between m and m0 which is puzzling...

observeInitTx :: ValidatedTx Era -> Maybe (OnChainTx tx, OnChainHeadState)

observeAbortTx :: ValidatedTx Era -> Maybe (OnChainTx tx, OnChainHeadState)

Identify an abort tx by trying to decode all redeemers to the right type. This is a very weak observation and should be more concretized. TODO(SN): make sure this is aborting "the right head / your head"

knownUtxo :: OnChainHeadState -> Map (TxIn StandardCrypto) (TxOut Era)

Provide a UTXO map for some given OnChainHeadState. At least used by the TinyWallet to lookup inputs.

Helpers

mkUnsignedTx :: TxBody Era -> TxDats Era -> Redeemers Era -> Map (ScriptHash StandardCrypto) (Script Era) -> ValidatedTx Era

scriptAddr :: Script Era -> Addr StandardCrypto

Get the ledger address for a given plutus script.

plutusScript :: Script -> Script Era

withDataHash :: Data Era -> (DataHash StandardCrypto, Data Era)

withScriptHash :: Script Era -> (ScriptHash StandardCrypto, Script Era)

datumsFromList :: [Data Era] -> TxDats Era

redeemersFromList :: [(StrictMaybe RdmrPtr, (Data Era, ExUnits))] -> Redeemers Era

Slightly unsafe, as it drops SNothing values from the list silently.