Hydra.Chain.Direct.Wallet

Description

Companion tiny-wallet for the direct chain component. This module provide some useful utilities to tracking the wallet's UTXO, and accessing it

Documentation

withTinyWallet :: NetworkMagic -> (VerificationKey, SigningKey) -> FilePath -> (TinyWallet IO -> IO ()) -> IO ()

data TinyWallet m

A TinyWallet is a small abstraction of a wallet with basic UTXO management. The wallet is assumed to have only one address, and only one UTXO at that address.

It can sign transactions and keeps track of its UTXO behind the scene.

Constructors

TinyWallet

Fields

getUtxo :: m (Maybe (TxIn, TxOut))

getAddress :: Address

sign :: TxBody -> VkWitness

newtype NetworkMagic

Constructors

NetworkMagic

Fields

unNetworkMagic :: Word32

applyBlock :: Block -> Address -> Map TxIn TxOut -> Map TxIn TxOut

Apply a block to our wallet. Does nothing if the transaction does not modify the UTXO set, or else, remove consumed utxos and add produced ones.

To determine whether a produced output is ours, we compare it to our unique address.