Hydra.Crypto

Description

Hydra multi-signature credentials and cryptographic primitives used to sign and verify snapshots (or any messages) within the Hydra protocol.

We are re-using the Key interface of 'cardano-api' for a consistent representation. For example: Cardano credentials are 'VerificationKey PaymentKey', Hydra credentials are 'VerificationKey HydraKey'.

Currently MultiSignature interface is only supporting naiive, concatenated multi-signatures and will change when we adopt aggregated multi-signatures including aggregate keys.

Cardano Key interface

class (Eq (VerificationKey keyrole), Show (VerificationKey keyrole), SerialiseAsRawBytes (Hash keyrole), HasTextEnvelope (VerificationKey keyrole), HasTextEnvelope (SigningKey keyrole)) => Key keyrole where

Associated Types

data VerificationKey keyrole

data SigningKey keyrole

Methods

getVerificationKey :: SigningKey keyrole -> VerificationKey keyrole

deterministicSigningKey :: AsType keyrole -> Seed -> SigningKey keyrole

deterministicSigningKeySeedSize :: AsType keyrole -> Word

verificationKeyHash :: VerificationKey keyrole -> Hash keyrole

Hydra specifics

data family Hash keyrole

data family AsType t

pattern HydraVerificationKey :: VerKeyDSIGN Ed25519DSIGN -> VerificationKey HydraKey

pattern HydraSigningKey :: SignKeyDSIGN Ed25519DSIGN -> SigningKey HydraKey

pattern HydraKeyHash :: Hash Blake2b_256 (VerificationKey HydraKey) -> Hash HydraKey

pattern AsHydraKey :: AsType HydraKey

newtype MultiSignature a

Naiively aggregated multi-signatures.

Constructors

HydraMultiSignature

Fields

multiSignature :: [Signature a]

newtype Signature a

Signature of a, not containing the actual payload.

Constructors

HydraSignature (SigDSIGN Ed25519DSIGN)

data HydraKey

Hydra keys (keyrole) which can be used to sign and verify messages, as well as aggregate multi-signatures.

generateSigningKey :: ByteString -> SigningKey HydraKey

Create a new SigningKey from a ByteString seed. The created keys are not random and insecure, so don't use this in production code!

sign :: SignableRepresentation a => SigningKey HydraKey -> a -> Signature a

Sign some value a with the provided SigningKey.

verify :: SignableRepresentation a => VerificationKey HydraKey -> Signature a -> a -> Bool

Verify a given 'Signature a' and value a using provided VerificationKey.

aggregate :: [Signature a] -> MultiSignature a

Combine multiple signatures of a into a 'MultiSignature a'.

aggregateInOrder :: Ord k => Map k (Signature a) -> [k] -> MultiSignature a

toPlutusSignatures :: MultiSignature a -> [Signature]