Hydra.Events.Rotation

Documentation

newtype RotationConfig

Constructors

RotateAfter Natural

type LogId = EventId

data EventStore e m

An EventSource and EventSink combined

Constructors

EventStore

Fields

eventSource :: EventSource e m

eventSink :: EventSink e m

rotate :: LogId -> e -> m ()

Rotate existing events into a given log id and start a new log from given e.

newRotatedEventStore

Arguments

:: (HasEventId e, MonadSTM m, MonadUnliftIO m, MonadTime m)

=> RotationConfig

-> s
Starting state of aggregate

-> (s -> e -> s)
Update aggregate state

-> (s -> EventId -> UTCTime -> e)
Create a checkpoint event

-> EventStore e m

-> m (EventStore e m)

Creates an event store that rotates according to given config and StateAggregate.