Hydra.HeadLogic

Description

Implements the Head Protocol's state machine as pure functions in an event sourced manner.

More specifically, the update will handle incoming Event (or rather "commands" in event sourcing speak) and convert that into a set of side-Effects and internal StateChanged events, which in turn are aggregated into a single HeadState.

As the specification is using a more imperative way of specifying the protocl behavior, one would find the decision logic in update while state updates can be found in the corresponding aggregate branch.

Documentation

defaultTTL :: TTL

onIdleClientInit :: Environment -> Outcome tx

Client request to init the head. This leads to an init transaction on chain, containing the head parameters.

Transition: IdleState → IdleState

onIdleChainInitTx

Arguments

:: ChainStateType tx
New chain state.

-> [Party]

-> ContestationPeriod

-> HeadId

-> Outcome tx

Observe an init transaction, initialize parameters in an InitialState and notify clients that they can now commit.

Transition: IdleState → InitialState

onInitialChainCommitTx

Arguments

:: Monoid (UTxOType tx)

=> InitialState tx

-> ChainStateType tx
New chain state

-> Party
Comitting party

-> UTxOType tx
Committed UTxO

-> Outcome tx

Observe a commit transaction and record the committed UTxO in the state. Also, if this is the last commit to be observed, post a collect-com transaction on-chain.

Transition: InitialState → InitialState

onInitialClientAbort :: Monoid (UTxOType tx) => InitialState tx -> Outcome tx

Client request to abort the head. This leads to an abort transaction on chain, reimbursing already committed UTxOs.

Transition: InitialState → InitialState

onInitialChainAbortTx

Arguments

:: Monoid (UTxOType tx)

=> ChainStateType tx
New chain state

-> Committed tx

-> HeadId

-> Outcome tx

Observe an abort transaction by switching the state and notifying clients about it.

Transition: InitialState → IdleState

onInitialChainCollectTx

Arguments

:: IsChainState tx

=> InitialState tx

-> ChainStateType tx
New chain state

-> Outcome tx

Observe a collectCom transaction. We initialize the OpenState using the head parameters from IdleState and construct an InitialSnapshot holding u0 from the committed UTxOs.

Transition: InitialState → OpenState

onOpenClientNewTx

Arguments

:: tx
The transaction to be submitted to the head.

-> Outcome tx

Client request to ingest a new transaction into the head.

Transition: OpenState → OpenState

onOpenNetworkReqTx

Arguments

:: IsTx tx

=> Environment

-> Ledger tx

-> OpenState tx

-> TTL

-> tx
The transaction to be submitted to the head.

-> Outcome tx

Process a transaction request (ReqTx) from a party.

We apply this transaction to the seen utxo (ledger state). If not applicable, we wait and retry later. If it applies, this yields an updated seen ledger state. Then, we check whether we are the leader for the next snapshot and emit a snapshot request ReqSn including this transaction if needed.

Transition: OpenState → OpenState

onOpenNetworkReqSn

Arguments

:: IsTx tx

=> Environment

-> Ledger tx

-> OpenState tx

-> Party
Party which sent the ReqSn.

-> SnapshotNumber
Requested snapshot number.

-> [TxIdType tx]
List of transactions to snapshot.

-> Outcome tx

Process a snapshot request (ReqSn) from party.

This checks that s is the next snapshot number and that the party is responsible for leading that snapshot. Then, we potentially wait until the previous snapshot is confirmed (no snapshot is in flight), before we apply (or wait until applicable) the requested transactions to the last confirmed snapshot. Only then, we start tracking this new "seen" snapshot, compute a signature of it and send the corresponding AckSn to all parties. Finally, the pending transaction set gets pruned to only contain still applicable transactions.

Transition: OpenState → OpenState

onOpenNetworkAckSn

Arguments

:: IsTx tx

=> Environment

-> OpenState tx

-> Party
Party which sent the AckSn.

-> Signature (Snapshot tx)
Signature from other party.

-> SnapshotNumber
Snapshot number of this AckSn.

-> Outcome tx

Process a snapshot acknowledgement (AckSn) from a party.

We do require that the is from the last seen or next expected snapshot, and potentially wait wait for the corresponding ReqSn before proceeding. If the party hasn't sent us a signature yet, we store it. Once a signature from each party has been collected, we aggregate a multi-signature and verify it is correct. If everything is fine, the snapshot can be considered as the latest confirmed one. Similar to processing a ReqTx, we check whether we are leading the next snapshot and craft a corresponding ReqSn if needed.

Transition: OpenState → OpenState

onOpenClientClose :: OpenState tx -> Outcome tx

Client request to close the head. This leads to a close transaction on chain using the latest confirmed snaphshot of the OpenState.

Transition: OpenState → OpenState

onOpenChainCloseTx

Arguments

:: OpenState tx

-> ChainStateType tx
New chain state.

-> SnapshotNumber
Closed snapshot number.

-> UTCTime
Contestation deadline.

-> Outcome tx

Observe a close transaction. If the closed snapshot number is smaller than our last confirmed, we post a contest transaction. Also, we do schedule a notification for clients to fanout at the deadline.

Transition: OpenState → ClosedState

onClosedChainContestTx :: ClosedState tx -> SnapshotNumber -> Outcome tx

Observe a contest transaction. If the contested snapshot number is smaller than our last confirmed snapshot, we post a contest transaction.

Transition: ClosedState → ClosedState

onClosedClientFanout :: ClosedState tx -> Outcome tx

Client request to fanout leads to a fanout transaction on chain using the latest confirmed snapshot from ClosedState.

Transition: ClosedState → ClosedState

onClosedChainFanoutTx

Arguments

:: ClosedState tx

-> ChainStateType tx
New chain state

-> Outcome tx

Observe a fanout transaction by finalize the head state and notifying clients about it.

Transition: ClosedState → IdleState

update

Arguments

:: IsChainState tx

=> Environment

-> Ledger tx

-> HeadState tx
Current HeadState to validate the command against.

-> Event tx
Command sent to the HeadLogic to be processed.

-> Outcome tx

Handles commands and converts them into internal StateChanged events along with Effects, in case it is processed succesfully. Later, the Node will aggregate the events, resulting in a new HeadState.

aggregate :: IsChainState tx => HeadState tx -> StateChanged tx -> HeadState tx

Reflect StateChanged events onto the HeadState aggregate.

aggregateState :: IsChainState tx => HeadState tx -> Outcome tx -> HeadState tx

recoverState :: (Foldable t, IsChainState tx) => HeadState tx -> t (StateChanged tx) -> HeadState tx

data Event tx

The different events which are processed by the head logic (the "core"). Corresponding to each of the "shell" layers, we distinguish between events from the client, the network and the chain.

Constructors

ClientEvent
Event received from clients via the Hydra.API.

Fields

clientInput :: ClientInput tx

NetworkEvent
Event received from peers via a Hydra.Network.

ttl is a simple counter that's decreased every time the event is reenqueued due to a wait. It's default value is defaultTTL

Fields

ttl :: TTL

party :: Party

message :: Message tx

OnChainEvent
Event received from the chain via a Hydra.Chain.

Fields

chainEvent :: ChainEvent tx

PostTxError
Event to re-ingest errors from postTx for further processing.

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

type TTL = Natural

data RequirementFailure tx

Constructors

ReqSnNumberInvalid

Fields

requestedSn :: SnapshotNumber

lastSeenSn :: SnapshotNumber

ReqSnNotLeader

Fields

requestedSn :: SnapshotNumber

leader :: Party

InvalidMultisignature

Fields

multisig :: Text

vkeys :: [VerificationKey HydraKey]

SnapshotAlreadySigned

Fields

knownSignatures :: [Party]

receivedSignature :: Party

AckSnNumberInvalid

Fields

requestedSn :: SnapshotNumber

lastSeenSn :: SnapshotNumber

SnapshotDoesNotApply

Fields

requestedSn :: SnapshotNumber

txid :: TxIdType tx

error :: ValidationError

data LogicError tx

Preliminary type for collecting errors occurring during update. TODO: Try to merge this (back) into Outcome.

Constructors

InvalidEvent (Event tx) (HeadState tx)

RequireFailed (RequirementFailure tx)

NotOurHead

Fields

ourHeadId :: HeadId

otherHeadId :: HeadId

data ClosedState tx

An Closed head with an current candidate ConfirmedSnapshot, which may be contested before the $sel:contestationDeadline:ClosedState.

Constructors

ClosedState

Fields

parameters :: HeadParameters

confirmedSnapshot :: ConfirmedSnapshot tx

contestationDeadline :: UTCTime

readyToFanoutSent :: Bool

Tracks whether we have informed clients already about being ReadyToFanout.

chainState :: ChainStateType tx

headId :: HeadId

data SeenSnapshot tx

Data structure to help in tracking whether we have seen or requested a ReqSn already and if seen, the signatures we collected already.

Constructors

NoSeenSnapshot
Never saw a ReqSn.

LastSeenSnapshot
No snapshot in flight with last seen snapshot number as given.

Fields

lastSeen :: SnapshotNumber

RequestedSnapshot
ReqSn was sent out and it should be considered already in flight.

Fields

lastSeen :: SnapshotNumber

requested :: SnapshotNumber

SeenSnapshot
ReqSn for given snapshot was received.

Fields

snapshot :: Snapshot tx

signatories :: Map Party (Signature (Snapshot tx))

Collected signatures and so far.

data CoordinatedHeadState tx

Off-chain state of the Coordinated Head protocol.

Constructors

CoordinatedHeadState

Fields

localUTxO :: UTxOType tx

The latest UTxO resulting from applying $sel:localTxs:CoordinatedHeadState to CoordinatedHeadState. Spec: L̂

localTxs :: [tx]

List of transactions applied locally and pending inclusion in a snapshot. Spec: T̂

allTxs :: Map (TxIdType tx) tx

Map containing all the transactions ever seen by this node and not yet included in a snapshot. Spec: Tall

confirmedSnapshot :: ConfirmedSnapshot tx

The latest confirmed snapshot. Spec: U̅, s̅ and σ̅

seenSnapshot :: SeenSnapshot tx

Last seen snapshot and signatures accumulator. Spec: Û, ŝ and Σ̂

data OpenState tx

An Open head with a CoordinatedHeadState tracking off-chain transactions.

Constructors

OpenState

Fields

parameters :: HeadParameters

coordinatedHeadState :: CoordinatedHeadState tx

chainState :: ChainStateType tx

headId :: HeadId

currentSlot :: ChainSlot

type Committed tx = Map Party (UTxOType tx)

type PendingCommits = Set Party

data InitialState tx

An Initial head which already has an identity and is collecting commits.

Constructors

InitialState

Fields

parameters :: HeadParameters

pendingCommits :: PendingCommits

committed :: Committed tx

chainState :: ChainStateType tx

headId :: HeadId

newtype IdleState tx

An Idle head only having a chain state with things seen on chain so far.

Constructors

IdleState

Fields

chainState :: ChainStateType tx

data HeadState tx

The main state of the Hydra protocol state machine. It holds both, the overall protocol state, but also the off-chain CoordinatedHeadState.

Each of the sub-types (InitialState, OpenState, etc.) contain black-box IdleState corresponding to OnChainEvent that has been observed leading to the state.

Note that rollbacks are currently not fully handled in the head logic and only this internal chain state gets replaced with the "rolled back to" version.

Constructors

Idle (IdleState tx)

Initial (InitialState tx)

Open (OpenState tx)

Closed (ClosedState tx)

data Environment

Constructors

Environment

Fields

party :: Party

This is the p_i from the paper

signingKey :: SigningKey HydraKey

otherParties :: [Party]

contestationPeriod :: ContestationPeriod

getChainState :: HeadState tx -> ChainStateType tx

Get the chain state in any HeadState.

setChainState :: ChainStateType tx -> HeadState tx -> HeadState tx

Update the chain state in any HeadState.

seenSnapshotNumber :: SeenSnapshot tx -> SnapshotNumber

Get the last seen snapshot number given a SeenSnapshot.

data WaitReason tx

Constructors

WaitOnNotApplicableTx

Fields

validationError :: ValidationError

WaitOnSnapshotNumber

Fields

waitingFor :: SnapshotNumber

WaitOnSeenSnapshot

WaitOnTxs

Fields

waitingForTxIds :: [TxIdType tx]

WaitOnContestationDeadline

data Outcome tx

Constructors

Effects

Fields

effects :: [Effect tx]

StateChanged (StateChanged tx)

Wait

Fields

reason :: WaitReason tx

Error

Fields

error :: LogicError tx

Combined

Fields

left :: Outcome tx

right :: Outcome tx

data StateChanged tx

Head state changed event. These events represent all the internal state changes, get persisted and processed in an event sourcing manner.

Constructors

HeadInitialized

Fields

parameters :: HeadParameters

chainState :: ChainStateType tx

headId :: HeadId

CommittedUTxO

Fields

party :: Party

committedUTxO :: UTxOType tx

chainState :: ChainStateType tx

HeadAborted

Fields

chainState :: ChainStateType tx

HeadOpened

Fields

chainState :: ChainStateType tx

initialUTxO :: UTxOType tx

TransactionAppliedToLocalUTxO

Fields

tx :: tx

newLocalUTxO :: UTxOType tx

SnapshotRequestDecided

Fields

snapshotNumber :: SnapshotNumber

SnapshotRequested
A snapshot was requested by some party. NOTE: We deliberately already include an updated local ledger state to not need a ledger to interpret this event.

Fields

snapshot :: Snapshot tx

requestedTxIds :: [TxIdType tx]

newLocalUTxO :: UTxOType tx

newLocalTxs :: [tx]

TransactionReceived

Fields

tx :: tx

PartySignedSnapshot

Fields

snapshot :: Snapshot tx

party :: Party

signature :: Signature (Snapshot tx)

SnapshotConfirmed

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

HeadClosed

Fields

chainState :: ChainStateType tx

contestationDeadline :: UTCTime

HeadIsReadyToFanout

HeadFannedOut

Fields

chainState :: ChainStateType tx

ChainRolledBack

Fields

chainState :: ChainStateType tx

TickObserved

Fields

chainSlot :: ChainSlot

data Effect tx

Analogous to events, the pure head logic "core" can have effects emited to the "shell" layers and we distinguish the same: effects onto the client, the network and the chain.

Constructors

ClientEffect
Effect to be handled by the Hydra.API, results in sending this ServerOutput.

Fields

serverOutput :: ServerOutput tx

NetworkEffect
Effect to be handled by a Hydra.Network, results in a broadcast.

Fields

message :: Message tx

OnChainEffect
Effect to be handled by a Hydra.Chain, results in a postTx.

Fields

postChainTx :: PostChainTx tx

collectEffects :: Outcome tx -> [Effect tx]

collectWaits :: Outcome tx -> [WaitReason tx]

isLeader :: HeadParameters -> Party -> SnapshotNumber -> Bool