Hydra.HeadLogic.Event

Documentation

type TTL = Natural

data Event tx

The different events which are processed by the head logic (the "core"). Corresponding to each of the "shell" layers, we distinguish between events from the client, the network and the chain.

Constructors

ClientEvent
Event received from clients via the Hydra.API.

Fields

clientInput :: ClientInput tx

NetworkEvent
Event received from peers via a Hydra.Network.

ttl is a simple counter that's decreased every time the event is reenqueued due to a wait. It's default value is defaultTTL

Fields

ttl :: TTL

party :: Party

message :: Message tx

OnChainEvent
Event received from the chain via a Hydra.Chain.

Fields

chainEvent :: ChainEvent tx

PostTxError
Event to re-ingest errors from postTx for further processing.

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx