Hydra.HeadLogic.Outcome

Documentation

data Effect tx

Analogous to events, the pure head logic "core" can have effects emited to the "shell" layers and we distinguish the same: effects onto the client, the network and the chain.

Constructors

ClientEffect
Effect to be handled by the Hydra.API, results in sending this ServerOutput.

Fields

serverOutput :: ServerOutput tx

NetworkEffect
Effect to be handled by a Hydra.Network, results in a broadcast.

Fields

message :: Message tx

OnChainEffect
Effect to be handled by a Hydra.Chain, results in a postTx.

Fields

postChainTx :: PostChainTx tx

data StateChanged tx

Head state changed event. These events represent all the internal state changes, get persisted and processed in an event sourcing manner.

Constructors

HeadInitialized

Fields

parameters :: HeadParameters

chainState :: ChainStateType tx

headId :: HeadId

CommittedUTxO

Fields

party :: Party

committedUTxO :: UTxOType tx

chainState :: ChainStateType tx

HeadAborted

Fields

chainState :: ChainStateType tx

HeadOpened

Fields

chainState :: ChainStateType tx

initialUTxO :: UTxOType tx

TransactionAppliedToLocalUTxO

Fields

tx :: tx

newLocalUTxO :: UTxOType tx

SnapshotRequestDecided

Fields

snapshotNumber :: SnapshotNumber

SnapshotRequested
A snapshot was requested by some party. NOTE: We deliberately already include an updated local ledger state to not need a ledger to interpret this event.

Fields

snapshot :: Snapshot tx

requestedTxIds :: [TxIdType tx]

newLocalUTxO :: UTxOType tx

newLocalTxs :: [tx]

TransactionReceived

Fields

tx :: tx

PartySignedSnapshot

Fields

snapshot :: Snapshot tx

party :: Party

signature :: Signature (Snapshot tx)

SnapshotConfirmed

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

HeadClosed

Fields

chainState :: ChainStateType tx

contestationDeadline :: UTCTime

HeadIsReadyToFanout

HeadFannedOut

Fields

chainState :: ChainStateType tx

ChainRolledBack

Fields

chainState :: ChainStateType tx

TickObserved

Fields

chainSlot :: ChainSlot

data Outcome tx

Constructors

Effects

Fields

effects :: [Effect tx]

StateChanged (StateChanged tx)

Wait

Fields

reason :: WaitReason tx

Error

Fields

error :: LogicError tx

Combined

Fields

left :: Outcome tx

right :: Outcome tx

collectEffects :: Outcome tx -> [Effect tx]

collectWaits :: Outcome tx -> [WaitReason tx]

data WaitReason tx

Constructors

WaitOnNotApplicableTx

Fields

validationError :: ValidationError

WaitOnSnapshotNumber

Fields

waitingFor :: SnapshotNumber

WaitOnSeenSnapshot

WaitOnTxs

Fields

waitingForTxIds :: [TxIdType tx]

WaitOnContestationDeadline