Hydra.HeadLogic.State

Documentation

data HeadState tx

The main state of the Hydra protocol state machine. It holds both, the overall protocol state, but also the off-chain CoordinatedHeadState.

Each of the sub-types (InitialState, OpenState, etc.) contain a black-box IdleState corresponding to the ChainEvent that has been observed leading to the state.

Note that rollbacks are currently not fully handled in the head logic and only this internal chain state gets replaced with the "rolled back to" version.

TODO: chainState would actually not be needed in the HeadState anymore as we do not persist the HeadState and not access it in the HeadLogic either.

Constructors

Idle (IdleState tx)

Initial (InitialState tx)

Open (OpenState tx)

Closed (ClosedState tx)

setChainState :: ChainStateType tx -> HeadState tx -> HeadState tx

Update the chain state in any HeadState.

getChainState :: HeadState tx -> ChainStateType tx

Get the chain state in any HeadState.

getHeadParameters :: HeadState tx -> Maybe HeadParameters

Get the head parameters in any HeadState.

Idle

newtype IdleState tx

An Idle head only having a chain state with things seen on chain so far.

Constructors

IdleState

Fields

chainState :: ChainStateType tx

Initial

data InitialState tx

An Initial head which already has an identity and is collecting commits.

Constructors

InitialState

Fields

parameters :: HeadParameters

pendingCommits :: PendingCommits

committed :: Committed tx

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

type PendingCommits = Set Party

type Committed tx = Map Party (UTxOType tx)

Open

data OpenState tx

An Open head with a CoordinatedHeadState tracking off-chain transactions.

Constructors

OpenState

Fields

parameters :: HeadParameters

coordinatedHeadState :: CoordinatedHeadState tx

chainState :: ChainStateType tx

headId :: HeadId

currentSlot :: ChainSlot

headSeed :: HeadSeed

data CoordinatedHeadState tx

Off-chain state of the Coordinated Head protocol.

Constructors

CoordinatedHeadState

Fields

localUTxO :: UTxOType tx

The latest UTxO resulting from applying $sel:localTxs:CoordinatedHeadState to CoordinatedHeadState. Spec: L̂

localTxs :: [tx]

List of transactions applied locally and pending inclusion in a snapshot. Ordering in this list is important as transactions are added in order of application. Spec: T̂

allTxs :: !(Map (TxIdType tx) tx)

Map containing all the transactions ever seen by this node and not yet included in a snapshot. Spec: Tall

confirmedSnapshot :: ConfirmedSnapshot tx

The latest confirmed snapshot. Spec: S̅

seenSnapshot :: SeenSnapshot tx

Last seen snapshot and signatures accumulator. Spec: Û, ŝ and Σ̂

pendingDeposits :: Map (TxIdType tx) (Deposit tx)

Pending deposits as observed on chain. TODO: These should be actually stored outside of the HeadState to allow recovery when a head is not open. See https://github.com/cardano-scaling/hydra/issues/1812

currentDepositTxId :: Maybe (TxIdType tx)

Current/next deposit to incrementally commit. Spec: Uα TODO: update in spec: Uα -> tx^#α

decommitTx :: Maybe tx

Pending decommit transaction. Spec: txω

version :: SnapshotVersion

Last open state version as observed on chain. Spec: ̂v

data SeenSnapshot tx

Data structure to help in tracking whether we have seen or requested a ReqSn already and if seen, the signatures we collected already.

Constructors

NoSeenSnapshot
Never saw a ReqSn.

LastSeenSnapshot
No snapshot in flight with last seen snapshot number as given.

Fields

lastSeen :: SnapshotNumber

RequestedSnapshot
ReqSn was sent out and it should be considered already in flight.

Fields

lastSeen :: SnapshotNumber

requested :: SnapshotNumber

SeenSnapshot
ReqSn for given snapshot was received.

Fields

snapshot :: Snapshot tx

signatories :: Map Party (Signature (Snapshot tx))

Collected signatures and so far.

seenSnapshotNumber :: SeenSnapshot tx -> SnapshotNumber

Get the last seen snapshot number given a SeenSnapshot.

data Deposit tx

A deposit tracked by the protocol. The DepositStatus determines whether it may be used for an incremental commit or not.

Constructors

Deposit

Fields

headId :: HeadId

deposited :: UTxOType tx

created :: UTCTime

deadline :: UTCTime

status :: DepositStatus

data DepositStatus

Constructors

Inactive

Active

Expired

Closed

data ClosedState tx

An Closed head with an current candidate ConfirmedSnapshot, which may be contested before the $sel:contestationDeadline:ClosedState.

Constructors

ClosedState

Fields

parameters :: HeadParameters

confirmedSnapshot :: ConfirmedSnapshot tx

contestationDeadline :: UTCTime

readyToFanoutSent :: Bool

Tracks whether we have informed clients already about being ReadyToFanout.

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

version :: SnapshotVersion