Hydra.HeadLogic.StateEvent

State change events as used by Hydra.Node

data StateEvent tx

A state change event with an event id that is the common entity to be loaded from an EventSource and sent to EventSinks.

Constructors

StateEvent

Fields

eventId :: EventId

stateChanged :: StateChanged tx

time :: UTCTime

genStateEvent :: StateChanged tx -> Gen (StateEvent tx)

mkCheckpoint :: NodeState tx -> EventId -> UTCTime -> StateEvent tx