Hydra.Ledger.Cardano

Documentation

cardanoLedger :: Globals -> LedgerEnv LedgerEra -> Ledger Tx

Use the cardano-ledger as an in-hydra Ledger.

mkSimpleTx

Arguments

:: (TxIn, TxOut CtxUTxO)

-> (AddressInEra, Value)
Recipient address and amount.

-> SigningKey PaymentKey
Sender's signing key.

-> Either TxBodyError Tx

Create a zero-fee, payment cardano transaction.

mkRangedTx

Arguments

:: (TxIn, TxOut CtxUTxO)

-> (AddressInEra, Value)
Recipient address and amount.

-> SigningKey PaymentKey
Sender's signing key.

-> (Maybe TxValidityLowerBound, Maybe TxValidityUpperBound)

-> Either TxBodyError Tx

Create a zero-fee, payment cardano transaction with validity range.

genSigningKey :: Gen (SigningKey PaymentKey)

genVerificationKey :: Gen (VerificationKey PaymentKey)

genKeyPair :: Gen (VerificationKey PaymentKey, SigningKey PaymentKey)

genSequenceOfSimplePaymentTransactions :: Gen (UTxO, [Tx])

Generates a sequence of simple "transfer" transactions for a single key. The kind of transactions produced by this generator is very limited, see generateOneTransfer.

genFixedSizeSequenceOfSimplePaymentTransactions :: Int -> Gen (UTxO, [Tx])

generateOneTransfer :: NetworkId -> (UTxO, (VerificationKey PaymentKey, SigningKey PaymentKey), [Tx]) -> Int -> Gen (UTxO, (VerificationKey PaymentKey, SigningKey PaymentKey), [Tx])

genOutput :: forall ctx. VerificationKey PaymentKey -> Gen (TxOut ctx)

genTxOutAdaOnly :: VerificationKey PaymentKey -> Gen (TxOut ctx)

Generate an ada-only TxOut payed to an arbitrary public key.

genUTxOAdaOnlyOfSize :: Int -> Gen UTxO

Generate a fixed size UTxO with ada-only outputs.

genUTxOAlonzo :: Gen UTxO

Generate Babbage era UTxO', which may contain arbitrary assets in TxOuts addressed to public keys *and* scripts. NOTE: This is not reducing size when generating assets in TxOuts, so will end up regularly with 300+ assets with generator size 30. NOTE: The Arbitrary TxIn instance from the ledger is producing colliding values, so we replace them.

genUTxOSized :: Int -> Gen UTxO

Generate a Babbage era UTxO' with given number of outputs. See also genTxOut.

genUTxO1 :: Gen (TxOut CtxUTxO) -> Gen UTxO

Genereate a UTxO' with a single entry using given TxOut generator.

genTxOut :: Gen (TxOut ctx)

Generate a Babbage era TxOut, which may contain arbitrary assets addressed to public keys and scripts, as well as datums.

NOTE: This generator does * not produce byron addresses as most of the cardano ecosystem dropped support for that (including plutus), * not produce reference scripts as they are not fully "visible" from plutus, * replace stake pointers with null references as nobody uses that.

genTxOutByron :: Gen (TxOut ctx)

Generate a TxOut with a byron address. This is usually not supported by Hydra or Plutus.

genTxOutWithReferenceScript :: Gen (TxOut ctx)

Generate a TxOut with a reference script. The standard genTxOut is not including reference scripts, use this generator if you are interested in these cases.

genUTxOFor :: VerificationKey PaymentKey -> Gen UTxO

Generate utxos owned by the given cardano key.

genOneUTxOFor :: VerificationKey PaymentKey -> Gen UTxO

Generate a single UTXO owned by vk.

genAddressInEra :: NetworkId -> Gen AddressInEra

NOTE: See note on mkVkAddress about NetworkId.

genValue :: Gen Value

genAdaOnlyUTxO :: Gen UTxO

Generate UTXO entries that do not contain any assets. Useful to test / measure cases where

adaOnly :: TxOut CtxUTxO -> TxOut CtxUTxO

genUTxOWithSimplifiedAddresses :: Gen UTxO

Generate "simplified" UTXO, ie. without some of the complexities required for backward-compatibility and obscure features.

shrinkUTxO :: UTxO -> [UTxO]

shrinkValue :: Value -> [Value]

module Hydra.Ledger.Cardano.Builder

data ShelleyGenesis era

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTime

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(ShelleyPParams era)

sgGenDelegs :: !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds :: ListMap (Addr (Crypto era)) Coin

sgStaking :: ShelleyGenesisStaking (Crypto era)

type Tx = Tx Era

Orphan instances

Arbitrary TxId

Instance details

Methods

arbitrary :: Gen TxId

shrink :: TxId -> [TxId]

Arbitrary AssetName

Instance details

Methods

arbitrary :: Gen AssetName

shrink :: AssetName -> [AssetName]

Arbitrary UTxO

Instance details

Methods

arbitrary :: Gen UTxO

shrink :: UTxO -> [UTxO]

Arbitrary Tx

Instance details

Methods

arbitrary :: Gen Tx

shrink :: Tx -> [Tx]

FromJSON Tx

Instance details

Methods

parseJSON :: Value -> Parser Tx

parseJSONList :: Value -> Parser [Tx]

ToJSON Tx

Instance details

Methods

toJSON :: Tx -> Value

toEncoding :: Tx -> Encoding

toJSONList :: [Tx] -> Value

toEncodingList :: [Tx] -> Encoding

FromCBOR UTxO

Instance details

Methods

fromCBOR :: Decoder s UTxO

label :: Proxy UTxO -> Text

FromCBOR Tx

Instance details

Methods

fromCBOR :: Decoder s Tx

label :: Proxy Tx -> Text

ToCBOR UTxO

Instance details

Methods

toCBOR :: UTxO -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy UTxO -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [UTxO] -> Size

ToCBOR Tx

Instance details

Methods

toCBOR :: Tx -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy Tx -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [Tx] -> Size

IsTx Tx

Instance details

Associated Types

type UTxOType Tx

type TxIdType Tx

type ValueType Tx

Methods

txId :: Tx -> TxIdType Tx

balance :: UTxOType Tx -> ValueType Tx

hashUTxO :: UTxOType Tx -> ByteString

Arbitrary (Hash PaymentKey)

Instance details

Methods

arbitrary :: Gen (Hash PaymentKey)

shrink :: Hash PaymentKey -> [Hash PaymentKey]

Arbitrary (VerificationKey PaymentKey)

Instance details

Methods

arbitrary :: Gen (VerificationKey PaymentKey)

shrink :: VerificationKey PaymentKey -> [VerificationKey PaymentKey]

Arbitrary (TxOut CtxUTxO)

Instance details

Methods

arbitrary :: Gen (TxOut CtxUTxO)

shrink :: TxOut CtxUTxO -> [TxOut CtxUTxO]