Hydra.Ledger.Cardano.Builder

Description

A concise interface for building transactions, built on top of the cardano-api.

Types

type TxBuilder = TxBodyContent BuildTx

Executing

unsafeBuildTransaction :: HasCallStack => TxBuilder -> Tx

Construct a transction from a builder. It is said unsafe because the underlying implementation will perform some sanity check on a transaction; for example, check that it has at least one input, that no outputs are negatives and whatnot.

We use the builder only internally for on-chain transaction crafted in the context of Hydra.

data InvalidTransactionException

A runtime exception to capture (programmer) failures when building transactions. This should never happened in practice (famous last words...)!

Constructors

InvalidTransactionException

Fields

txBodyError :: TxBodyError

builder :: TxBuilder

Constructing

emptyTxBody :: TxBuilder

An empty TxBodyContent with all empty/zero values to be extended using record updates.

FIXME: makeTransactionBody throws when one tries to build a transaction with scripts but no collaterals. This is unfortunate because collaterals are currently added after by out integrated wallet... We may want to revisit our flow to avoid this exception and have the wallet work from a TxBuilder instead of fiddling with a sealed CardanoTx.

Similarly, makeTransactionBody throws when building a transaction with scripts and no protocol parameters (needed to compute the script integrity hash). This is also added by our wallet at the moment so hopefully, this ugly work-around will be removed eventually.

So we currently bypass this by having default but seemingly innofensive values for collaterals and protocol params in the empty value

addInputs :: TxIns BuildTx -> TxBuilder -> TxBuilder

Add new inputs to an ongoing builder.

addVkInputs :: [TxIn] -> TxBuilder -> TxBuilder

Like addInputs but only for vk inputs which requires no additional data.

addOutputs :: [TxOut CtxTx] -> TxBuilder -> TxBuilder

Append new outputs to an ongoing builder.

addExtraRequiredSigners :: [Hash PaymentKey] -> TxBuilder -> TxBuilder

Add extra required key witnesses to a transaction.

mintTokens :: ToScriptData redeemer => PlutusScript -> redeemer -> [(AssetName, Quantity)] -> TxBuilder -> TxBuilder

Mint tokens with given plutus minting script and redeemer.

burnTokens :: ToScriptData redeemer => PlutusScript -> redeemer -> [(AssetName, Quantity)] -> TxBuilder -> TxBuilder

Burn tokens with given plutus minting script and redeemer. This is really just mintTokens with negated Quantity.

setValidityUpperBound :: SlotNo -> TxBuilder -> TxBuilder

Set the upper validity bound for this transaction to some SlotNo.

setValidityLowerBound :: SlotNo -> TxBuilder -> TxBuilder

Set the lower validity bound for this transaction to some SlotNo.