Hydra.Ledger.Mock

Description

A mock implementation of a ledger

Documentation

data MockTx

Simple mock transaction.

NOTE: There's no need to represent a real tx and do any fake ledger validation because we can already represent that via InvalidTx.

In the end, we are really interested in the resulting UTxO which _could_ be constructed from all the valid transactions that have passed through the head. So it suffices to keep a list of all valid transactions in the mock.

Constructors

ValidTx TxId

InvalidTx

type TxId = Integer

mockLedger :: Ledger MockTx