Hydra.Logging

Description

Adapter module to the actual logging framework. For now we are using the iohk-monitoring package, but that might change soon.

Tracer

newtype Tracer (m :: Type -> Type) a

Constructors

Tracer

Fields

runTracer :: a -> m ()

natTracer :: (forall x. m x -> n x) -> Tracer m s -> Tracer n s

nullTracer :: forall (m :: Type -> Type) a. Applicative m => Tracer m a

traceWith :: Tracer m a -> a -> m ()

traceInTVar :: MonadSTM m => TVar m [a] -> Tracer m a

class ToObject a where

Minimal complete definition

Nothing

Methods

toObject :: TracingVerbosity -> a -> Object

textTransformer :: a -> Object -> Text

data TracingVerbosity

Constructors

MinimalVerbosity

NormalVerbosity

MaximalVerbosity

Using it

data Verbosity

Constructors

Quiet

Verbose Text

data Envelope a

Provides logging metadata for entries. NOTE(AB): setting the message to be HydraLog would require to pass along tx and net type arguments which is useless in the context, hence the more generic type.

Constructors

Envelope

Fields

namespace :: Text

timestamp :: UTCTime

threadId :: Int

message :: a

withTracer :: forall m msg a. (MonadIO m, ToJSON msg) => Verbosity -> (Tracer m msg -> IO a) -> IO a

Start logging thread and acquire a Tracer. This tracer will dump all messsages on stdout, one message per line, formatted as JSON.

contramap :: Contravariant f => (a -> b) -> f b -> f a

showLogsOnFailure :: (MonadSTM m, MonadCatch m, Show msg, MonadSay m) => (Tracer m msg -> m a) -> m a