Hydra.Logging.Messages

Description

Aggregates all tracing messages in a single type.

This module provides a central point where top-level traced messages are grouped. This is useful for traces consumers that will need to do something specific depending on various tracing messages, eg. monitoring and metrics collection.

Documentation

data HydraLog tx

Constructors

DirectChain

Fields

directChain :: CardanoChainLog

APIServer

Fields

api :: APIServerLog

Network

Fields

network :: NetworkLog

Node

Fields

node :: HydraNodeLog tx

NodeOptions

Fields

runOptions :: RunOptions

Persistence

Fields

persistenceLog :: PersistenceLog

EnteringMainloop