Hydra.Network

Description

Interface to the Hydra network and base types. Concrete implementations are provided by submodules. Import those instead of this one if interested in actually configuring and running a real network layer.

Types

newtype Network m msg

Handle to interface with the hydra network and send messages "off chain".

Constructors

Network

Fields

broadcast :: msg -> m ()

Send a msg to the whole hydra network.

type NetworkComponent m msg = NetworkCallback msg m -> (Network m msg -> m ()) -> m ()

A type tying both inbound and outbound messages sending in a single Component.

data PortNumber

type Host = (HostName, PortNumber)

type NetworkCallback msg m = msg -> m ()

Handle to interface for inbound messages.

data IP

Parser

readHost :: String -> Maybe Host

readPort :: String -> Maybe PortNumber

Orphan instances

Generic (HydraMessage tx)

Instance details

Associated Types

type Rep (HydraMessage tx) :: Type -> Type

Methods

from :: HydraMessage tx -> Rep (HydraMessage tx) x

to :: Rep (HydraMessage tx) x -> HydraMessage tx

(FromCBOR tx, FromCBOR (UTxO tx)) => FromCBOR (HydraMessage tx)

Instance details

Methods

fromCBOR :: Decoder s (HydraMessage tx)

label :: Proxy (HydraMessage tx) -> Text

(FromCBOR tx, FromCBOR (UTxO tx)) => FromCBOR (Snapshot tx)

Instance details

Methods

fromCBOR :: Decoder s (Snapshot tx)

label :: Proxy (Snapshot tx) -> Text

(ToCBOR tx, ToCBOR (UTxO tx)) => ToCBOR (HydraMessage tx)

Instance details

Methods

toCBOR :: HydraMessage tx -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy (HydraMessage tx) -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [HydraMessage tx] -> Size

(ToCBOR tx, ToCBOR (UTxO tx)) => ToCBOR (Snapshot tx)

Instance details

Methods

toCBOR :: Snapshot tx -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy (Snapshot tx) -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [Snapshot tx] -> Size