Hydra.Network.Ouroboros

Description

Ouroboros-based implementation of Network interface. This implements a dumb FireForget protocol and maintains one connection to each peer. Contrary to other protocols implemented in Ouroboros, this is a push-based protocol.

Documentation

withIOManager :: WithIOManager

data PortNumber

newtype Network m msg

Interface from the application to the network layer.

Constructors

Network

Fields

broadcast :: msg -> m ()

Send a msg to the whole configured hydra network including ourselves.

data Host

Constructors

Host

Fields

hostname :: Text

port :: PortNumber

newtype NetworkCallback msg m

Interface from network layer to the application. XXX: Reliably delivering a message in the crash-recovery fault model is tricky. According to "Introduction to Reliable and Secure Distributed Programming" section "2.2.4 Crashes with recoveries" explains that storing to stable storage and just pointing to stored events is a better way.

Constructors

NetworkCallback

Fields

deliver :: msg -> m ()

The given msg was received from the network.

type NetworkComponent m inbound outbound a = NetworkCallback inbound m -> (Network m outbound -> m a) -> m a

A type tying both inbound and outbound messages sending in a single Component.

A NetworkComponent can have different inbound and outbound message types.

data TraceOuroborosNetwork msg

Constructors

TraceSubscriptions (WithIPList (SubscriptionTrace SockAddr))

TraceErrorPolicy (WithAddr SockAddr ErrorPolicyTrace)

TraceAcceptPolicy AcceptConnectionsPolicyTrace

TraceHandshake (WithMuxBearer (ConnectionId SockAddr) (TraceSendRecv (Handshake HydraVersionedProtocolNumber Term)))

TraceSendRecv (TraceSendRecv (FireForget msg))

data WithHost trace

Constructors

WithHost Host trace

data NetworkServerListenException

Constructors

NetworkServerListenException

Fields

ioException :: IOException

localHost :: Host

tagged :: Text -> [Pair] -> Value

withOuroborosNetwork :: forall inbound outbound. (ToCBOR outbound, FromCBOR outbound) => (ToCBOR inbound, FromCBOR inbound) => Tracer IO (WithHost (TraceOuroborosNetwork outbound)) -> HydraNetworkConfig -> (HydraHandshakeRefused -> IO ()) -> NetworkComponent IO inbound outbound ()

encodeWithIPList :: WithIPList (SubscriptionTrace SockAddr) -> Value

encodeWithAddr :: WithAddr SockAddr ErrorPolicyTrace -> Value

encodeTraceSendRecvHandshake :: WithMuxBearer (ConnectionId SockAddr) (TraceSendRecv (Handshake HydraVersionedProtocolNumber Term)) -> [Pair]

encodeTraceSendRecvFireForget :: forall msg. ToJSON msg => TraceSendRecv (FireForget msg) -> [Pair]

data HydraNetworkConfig

Constructors

HydraNetworkConfig

Fields

protocolVersion :: HydraVersionedProtocolNumber

localHost :: Host

remoteHosts :: [Host]

data HydraVersionedProtocolData

Constructors

MkHydraVersionedProtocolData

hydraVersionedProtocolCodec :: CodecCBORTerm (String, Maybe Int) HydraVersionedProtocolNumber

hydraVersionedProtocolDataCodec :: VersionDataCodec Term HydraVersionedProtocolNumber HydraVersionedProtocolData