Hydra.Network.Ouroboros.Server

Documentation

data FireForgetServer msg m a

Constructors

FireForgetServer

Fields

recvMsg :: msg -> m (FireForgetServer msg m a)

The client sent us a message. There is no response and we must have effects.

recvMsgDone :: m a

The client terminated. Here we have a pure return value, but we could have done another action in m if we wanted to.

fireForgetServerPeer :: Monad m => FireForgetServer msg m a -> Peer (FireForget msg) 'AsServer 'StIdle m a