Hydra.Network.Ouroboros.Type

Documentation

data FireForget msg where

TODO explain Protocol

It is used both as a type level tag for the protocol and as the kind of the types of the states in the protocol state machine. That is FireForget is a kind, and StIdle is a type of that kind.

Constructors

StIdle :: FireForget msg

StDone :: FireForget msg

data SFireForget (st :: FireForget msg) where

Constructors

SingIdle :: SFireForget StIdle

SingDone :: SFireForget StDone