Hydra.Node

Description

Top-level module to run a single Hydra node.

Checkout Hydra Documentation for some details about the overall architecture of the Node.

Environment Handling

initEnvironment :: RunOptions -> IO Environment

Intialize the Environment from command line options.

newtype ParameterMismatch

Exception used to indicate command line options not matching the persisted state.

Constructors

ParameterMismatch [ParamMismatch]

data ParamMismatch

Constructors

ContestationPeriodMismatch

Fields

loadedCp :: ContestationPeriod

configuredCp :: ContestationPeriod

PartiesMismatch

Fields

loadedParties :: [Party]

configuredParties :: [Party]

checkHeadState :: MonadThrow m => Tracer m (HydraNodeLog tx) -> Environment -> HeadState tx -> m ()

Checks that command line options match a given HeadState. This funciton takes Environment because it is derived from RunOptions via initEnvironment.

Throws: ParameterMismatch when state not matching the environment.

Create and run a hydra node

data HydraNode tx m

Main handle of a hydra node where all layers are tied together.

Constructors

HydraNode

Fields

eq :: EventQueue m (Event tx)

hn :: Network m (Message tx)

nodeState :: NodeState tx m

oc :: Chain tx m

server :: Server tx m

ledger :: Ledger tx

env :: Environment

persistence :: PersistenceIncremental (StateChanged tx) m

data HydraNodeLog tx

Constructors

BeginEvent

Fields

by :: Party

eventId :: Word64

event :: Event tx

EndEvent

Fields

by :: Party

eventId :: Word64

BeginEffect

Fields

by :: Party

eventId :: Word64

effectId :: Word32

effect :: Effect tx

EndEffect

Fields

by :: Party

eventId :: Word64

effectId :: Word32

LogicOutcome

Fields

by :: Party

outcome :: Outcome tx

LoadedState

Fields

numberOfEvents :: Word64

Misconfiguration

Fields

misconfigurationErrors :: [ParamMismatch]

runHydraNode :: (MonadCatch m, MonadAsync m, IsChainState tx) => Tracer m (HydraNodeLog tx) -> HydraNode tx m -> m ()

stepHydraNode :: (MonadCatch m, MonadAsync m, IsChainState tx) => Tracer m (HydraNodeLog tx) -> HydraNode tx m -> m ()

waitDelay :: DiffTime

The time to wait between re-enqueuing a Wait outcome from HeadLogic.

processNextEvent :: IsChainState tx => HydraNode tx m -> Event tx -> STM m (Outcome tx)

Monadic interface around update.

processEffects :: (MonadAsync m, MonadCatch m, IsChainState tx) => HydraNode tx m -> Tracer m (HydraNodeLog tx) -> Word64 -> Outcome tx -> m ()

Manage state

data NodeState tx m

Handle to access and modify the state in the Hydra Node.

Constructors

NodeState

Fields

modifyHeadState :: forall a. (HeadState tx -> (a, HeadState tx)) -> STM m a

queryHeadState :: STM m (HeadState tx)

createNodeState :: MonadLabelledSTM m => HeadState tx -> m (NodeState tx m)

Initialize a new NodeState.

loadState :: (MonadThrow m, IsChainState tx) => Tracer m (HydraNodeLog tx) -> PersistenceIncremental (StateChanged tx) m -> ChainStateType tx -> m (HeadState tx)

Load a HeadState from persistence.