Hydra.Node.EventQueue

Description

The general event queue from which the Hydra head is fed with events.

Documentation

data EventQueue m e

The single, required queue in the system from which a hydra head is "fed". NOTE(SN): this probably should be bounded and include proper logging NOTE(SN): handle pattern, but likely not required as there is no need for an alternative implementation

Constructors

EventQueue

Fields

putEvent :: e -> m ()

putEventAfter :: DiffTime -> Queued e -> m ()

nextEvent :: m (Queued e)

isEmpty :: m Bool

data Queued e

Constructors

Queued

Fields

eventId :: Word64

queuedEvent :: e

createEventQueue :: (MonadDelay m, MonadAsync m, MonadLabelledSTM m) => m (EventQueue m e)