Hydra.Node.InputQueue

Description

The general input queue from which the Hydra head is fed with inputs.

Documentation

data InputQueue m e

The single, required queue in the system from which a hydra head is "fed". NOTE(SN): this probably should be bounded and include proper logging NOTE(SN): handle pattern, but likely not required as there is no need for an alternative implementation

Constructors

InputQueue

Fields

enqueue :: e -> m ()

reenqueue :: DiffTime -> Queued e -> m ()

dequeue :: m (Queued e)

isEmpty :: m Bool

data Queued a

Constructors

Queued

Fields

queuedId :: Word64

queuedItem :: a

createInputQueue :: (MonadDelay m, MonadAsync m, MonadLabelledSTM m) => m (InputQueue m e)