Hydra.Node.Network

Description

Concrete Network stack dedicated to running a hydra-node.

This module provides a withNetwork function which is the composition of several layers in order to provide various capabilities:

withHeartbeat maintains knowledge about peers' connectivity,

withReliability deals with connections reliability, handling the case of messages being dropped (but not node crash in general),

withAuthentication handles messages' authentication and signature verification,

withOuroborosNetwork deals with maintaining individual connections to peers and the nitty-gritty details of messages sending and retrieval.

The following diagram details the various types of messages each layer is exchanging with its predecessors and successors.

         ▲                                    │
         │ Authenticate msg                   │ msg
         │                                    │
┌────────┴────────────────────────────────────▼──────┐
│                                                    │
│                   Heartbeat                        │
│        ▲                                           │
└────────┬────────────────────────────────────┼──────┘
         │                                    │
         │ Heartbeat (Authenticate msg)       │ Heartbeat msg
         │                                    │
┌────────┴───────────────┐                    │
│                        │                    │
│    FlipHeartbeats      │                    │
│                        │                    │
└────────▲───────────────┘                    │
         │                                    │
         │ Authenticate (Heartbeat msg)       │
         │                                    │
┌────────┴────────────────────────────────────▼──────┐
│                                                    │
│                   Reliability                      │
│                                                    │
└─────────▲───────────────────────────────────┼──────┘
          │                                   │
     Authenticated (ReliableMsg (Heartbeat msg))    ReliableMsg (Heartbeat msg)
          │                                   │
┌─────────┼───────────────────────────────────▼──────┐
│                                                    │
│                  Authenticate                      │
│                                                    │
└─────────▲───────────────────────────────────┼──────┘
          │                                   │
          │                                   │
      Signed (ReliableMsg (Heartbeat msg))       Signed (ReliableMsg (Heartbeat msg))
          │                                   │
┌─────────┼───────────────────────────────────▼──────┐
│                                                    │
│                  Ouroboros                         │
│                                                    │
└─────────▲───────────────────────────────────┼──────┘
          │                                   │
          │           (bytes)                 │
          │                                   ▼

Documentation

withNetwork

Arguments

:: (ToCBOR msg, ToJSON msg, FromJSON msg, FromCBOR msg)

=> Tracer IO (LogEntry tx msg)
Tracer to use for logging messages.

-> FilePath
Persistence directory

-> ConnectionMessages IO
Callback/observer for connectivity changes in peers.

-> SigningKey HydraKey
This node's signing key. This is used to sign messages sent to peers.

-> [Party]
The list of peers Party known to this node.

-> IP
IP address to listen on for incoming connections.

-> PortNumber
Port to listen on.

-> [Host]
Addresses and ports of remote peers.

-> NodeId
This node's id.

-> NetworkComponent IO (Authenticated msg) msg ()
Produces a NetworkComponent that can send msg and consumes Authenticated msg.

withFlipHeartbeats :: NetworkComponent m (Authenticated (Heartbeat msg)) msg1 a -> NetworkComponent m (Heartbeat (Authenticated msg)) msg1 a