Hydra.Option

Documentation

data Option

Constructors

Option

Fields

verbosity :: Verbosity

nodeId :: Natural

host :: IP

port :: PortNumber

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

monitoringPort :: Maybe PortNumber

parseHydraOptions :: IO Option

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraOptionsFromString :: [String] -> ParserResult Option

Pure parsing of Option from a list of arguments.

getParseResult :: ParserResult a -> Maybe a

defaultOption :: Option

data ParserResult a

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult