Hydra.Options

Documentation

data InvalidOptions

Constructors

MaximumNumberOfPartiesExceeded

CardanoAndHydraKeysMissmatch

data ChainConfig

Constructors

DirectChainConfig

Fields

networkId :: NetworkId

Network identifer to which we expect to connect.

nodeSocket :: FilePath

Path to a domain socket used to connect to the server.

cardanoSigningKey :: FilePath

Path to the cardano signing key of the internal wallet.

cardanoVerificationKeys :: [FilePath]

Paths to other node's verification keys.

startChainFrom :: Maybe ChainPoint

Point at which to start following the chain.

data LedgerConfig

Constructors

CardanoLedgerConfig

Fields

cardanoLedgerGenesisFile :: FilePath

cardanoLedgerProtocolParametersFile :: FilePath

data RunOptions

Constructors

RunOptions

Fields

verbosity :: Verbosity

nodeId :: NodeId

host :: IP

port :: PortNumber

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

monitoringPort :: Maybe PortNumber

hydraSigningKey :: FilePath

hydraVerificationKeys :: [FilePath]

hydraScriptsTxId :: TxId

persistenceDir :: FilePath

chainConfig :: ChainConfig

ledgerConfig :: LedgerConfig

data PublishOptions

Constructors

PublishOptions

Fields

publishNetworkId :: NetworkId

publishNodeSocket :: FilePath

publishSigningKey :: FilePath

data Command

Constructors

Run RunOptions

Publish PublishOptions

commandParser :: Parser Command

publishOptionsParser :: Parser PublishOptions

runOptionsParser :: Parser RunOptions

defaultLedgerConfig :: LedgerConfig

ledgerConfigParser :: Parser LedgerConfig

cardanoLedgerGenesisParser :: Parser FilePath

cardanoLedgerProtocolParametersParser :: Parser FilePath

defaultChainConfig :: ChainConfig

chainConfigParser :: Parser ChainConfig

networkIdParser :: Parser NetworkId

nodeSocketParser :: Parser FilePath

cardanoSigningKeyFileParser :: Parser FilePath

cardanoVerificationKeyFileParser :: Parser FilePath

hydraSigningKeyFileParser :: Parser FilePath

hydraVerificationKeyFileParser :: Parser FilePath

peerParser :: Parser Host

nodeIdParser :: Parser NodeId

verbosityParser :: Parser Verbosity

hostParser :: Parser IP

portParser :: Parser PortNumber

apiHostParser :: Parser IP

apiPortParser :: Parser PortNumber

monitoringPortParser :: Parser PortNumber

startChainFromParser :: Parser ChainPoint

hydraScriptsTxIdParser :: Parser TxId

persistenceDirParser :: Parser FilePath

hydraNodeCommand :: ParserInfo Command

maximumNumberOfParties :: Int

Hardcoded limit for maximum number of parties in a head protocol The value 4 is obtained from calculating the costs of running the scripts and on-chan validators (see computeCollectComCost computeAbortCost)

explain :: InvalidOptions -> String

validateRunOptions :: RunOptions -> Either InvalidOptions ()

Validate cmd line arguments for hydra-node and check if they make sense before actually running the node. Rules we apply: - Check if number of parties is bigger than our hardcoded limit (by looking at loaded hydra or cardano keys and comparing it to the maximumNumberOfParties) - Check that number of loaded hydra keys match with the number of loaded cardano keys (by comparing lengths of the two lists)

parseHydraCommand :: IO Command

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraCommandFromArgs :: [String] -> ParserResult Command

Pure parsing of Option from a list of arguments.

toArgs :: RunOptions -> [String]

Convert an Options instance into the corresponding list of command-line arguments.

This is useful in situations where one wants to programatically define Options, providing some measure of type safety, without having to juggle with strings.

toArgNetworkId :: NetworkId -> String

genFilePath :: String -> Gen FilePath

genDirPath :: Gen FilePath

genChainPoint :: Gen ChainPoint

data ParserResult a

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult