Hydra.ServerOutput

Documentation

data ServerOutput tx

Constructors

PeerConnected

Fields

peer :: Host

PeerDisconnected

Fields

peer :: Host

ReadyToCommit

Fields

parties :: Set Party

Committed

Fields

party :: Party

utxo :: UTxOType tx

HeadIsOpen

Fields

utxo :: UTxOType tx

HeadIsClosed

Fields

snapshotNumber :: SnapshotNumber

remainingContestationPeriod :: NominalDiffTime

HeadIsContested

Fields

snapshotNumber :: SnapshotNumber

ReadyToFanout

HeadIsAborted

Fields

utxo :: UTxOType tx

HeadIsFinalized

Fields

utxo :: UTxOType tx

CommandFailed

Fields

clientInput :: ClientInput tx

TxSeen

Fields

transaction :: tx

TxValid

Fields

transaction :: tx

TxInvalid

Fields

utxo :: UTxOType tx

transaction :: tx

validationError :: ValidationError

SnapshotConfirmed

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

GetUTxOResponse

Fields

utxo :: UTxOType tx

InvalidInput

Fields

reason :: String

input :: Text

Greetings
A friendly welcome message which tells a client something about the node. Currently used for knowing what signing key the server uses (it only knows one).

Fields

me :: Party

PostTxOnChainFailed

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

RolledBack