Hydra.Snapshot

Documentation

type SnapshotNumber = Natural

data Snapshot tx

Constructors

Snapshot

Fields

number :: SnapshotNumber

utxo :: UTxOType tx

confirmed :: [tx]

The set of transactions that lead to utxo

data ConfirmedSnapshot tx

Snapshot when it was signed by all parties, i.e. it is confirmed.

Constructors

InitialSnapshot

Fields

snapshot :: Snapshot tx

ConfirmedSnapshot

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

getSnapshot :: ConfirmedSnapshot tx -> Snapshot tx

Safely get a Snapshot from a confirmed snapshot.

isInitialSnapshot :: ConfirmedSnapshot tx -> Bool

Tell whether a snapshot is the initial snapshot coming from the collect-com transaction.

genConfirmedSnapshot

Arguments

:: IsTx tx

=> SnapshotNumber
The lower bound on snapshot number to generate. If this is 0, then we can generate an InitialSnapshot or a ConfirmedSnapshot. Otherwise we generate only ConfirmedSnapshot with a number strictly superior to this lower bound.

-> UTxOType tx

-> [SigningKey]

-> Gen (ConfirmedSnapshot tx)