Hydra.API.Server

Documentation

newtype Server tx m

Handle to provide a means for sending server outputs to clients.

Constructors

Server

Fields

sendOutput :: ServerOutput tx -> m ()

Send some output to all connected clients.

type ServerCallback tx m = ClientInput tx -> m ()

Callback for receiving client inputs.

type ServerComponent tx m a = ServerCallback tx m -> (Server tx m -> m a) -> m a

A type tying both receiving input and sending output into a Component.

data APIServerConfig

Constructors

APIServerConfig

Fields

host :: IP

port :: PortNumber

tlsCertPath :: Maybe FilePath

tlsKeyPath :: Maybe FilePath

withAPIServer :: forall tx. IsChainState tx => APIServerConfig -> Environment -> Party -> PersistenceIncremental (TimedServerOutput tx) IO -> Tracer IO APIServerLog -> Chain tx IO -> PParams LedgerEra -> ServerOutputFilter tx -> ServerComponent tx IO ()

data RunServerException

An IOException with more IP and PortNumber added as context.

Constructors

RunServerException

Fields

ioException :: IOException

host :: IP

port :: PortNumber

type NotifyServerRunning = IO ()

type WaitForServer = IO ()

setupServerNotification :: IO (NotifyServerRunning, WaitForServer)

Setup notification and waiter to ensure that something only runs after the server is actually listening.