Hydra.API.Server

Documentation

newtype Server tx m

Handle to provide a means for sending server outputs to clients.

Constructors

Server

Fields

sendMessage :: ClientMessage tx -> m ()

Send some output to all connected clients.

data APIServerConfig

Constructors

APIServerConfig

Fields

host :: IP

port :: PortNumber

tlsCertPath :: Maybe FilePath

tlsKeyPath :: Maybe FilePath

withAPIServer :: forall tx. IsChainState tx => APIServerConfig -> Environment -> Party -> EventSource (StateEvent tx) IO -> Tracer IO APIServerLog -> Chain tx IO -> PParams LedgerEra -> ServerOutputFilter tx -> (ClientInput tx -> IO ()) -> ((EventSink (StateEvent tx) IO, Server tx IO) -> IO ()) -> IO ()

data RunServerException

An IOException with more IP and PortNumber added as context.

Constructors

RunServerException

Fields

ioException :: IOException

host :: IP

port :: PortNumber

type NotifyServerRunning = IO ()

type WaitForServer = IO ()

setupServerNotification :: IO (NotifyServerRunning, WaitForServer)

Setup notification and waiter to ensure that something only runs after the server is actually listening.

mkTimedServerOutputFromStateEvent :: IsTx tx => StateEvent tx -> Maybe (TimedServerOutput tx)

Defines the subset of StateEvent that should be sent as TimedServerOutput to clients.

projectPendingDeposits :: IsTx tx => [TxIdType tx] -> StateChanged tx -> [TxIdType tx]

Projection to obtain the list of pending deposits.

projectCommitInfo :: CommitInfo -> StateChanged tx -> CommitInfo

Projection to obtain CommitInfo needed to draft commit transactions. NOTE: We only want to project HeadId when the Head is in the Initializing state since this is when Head parties need to commit some funds.

projectInitializingHeadId :: Maybe HeadId -> StateChanged tx -> Maybe HeadId

Projection to obtain the HeadId needed to draft a commit transaction. NOTE: We only want to project HeadId when the Head is in the Initializing state since this is when Head parties need to commit some funds.

projectHeadStatus :: HeadStatus -> StateChanged tx -> HeadStatus

Projection function related to headStatus field in Greetings message.

projectSnapshotUtxo :: Monoid (UTxOType tx) => Maybe (UTxOType tx) -> StateChanged tx -> Maybe (UTxOType tx)

Projection of latest confirmed snapshot UTxO.

projectSeenSnapshot :: SeenSnapshot tx -> StateChanged tx -> SeenSnapshot tx

Projection of latest seen snapshot.

projectSnapshotConfirmed :: Maybe (ConfirmedSnapshot tx) -> StateChanged tx -> Maybe (ConfirmedSnapshot tx)

Projection of latest confirmed snapshot.