Hydra.API.Server

Documentation

newtype Server tx m

Handle to provide a means for sending server outputs to clients.

Constructors

Server

Fields

sendOutput :: ServerOutput tx -> m ()

Send some output to all connected clients.

type ServerCallback tx m = ClientInput tx -> m ()

Callback for receiving client inputs.

type ServerComponent tx m a = ServerCallback tx m -> (Server tx m -> m a) -> m a

A type tying both receiving input and sending output into a Component.

withAPIServer :: Tx tx => IP -> PortNumber -> Tracer IO APIServerLog -> ServerComponent tx IO ()

data APIServerLog