Hydra.API.ServerOutput

Documentation

data TimedServerOutput tx

The type of messages sent to clients by the Server.

Constructors

TimedServerOutput

Fields

output :: ServerOutput tx

seq :: Natural

time :: UTCTime

genTimedServerOutput :: ServerOutput tx -> Gen (TimedServerOutput tx)

Generate a random timed server output given a normal server output.

data ServerOutput tx

Individual server output messages as produced by the HeadLogic in the ClientEffect.

Constructors

PeerConnected

Fields

peer :: NodeId

PeerDisconnected

Fields

peer :: NodeId

HeadIsInitializing

Fields

headId :: HeadId

parties :: Set Party

Committed

Fields

headId :: HeadId

party :: Party

utxo :: UTxOType tx

HeadIsOpen

Fields

headId :: HeadId

utxo :: UTxOType tx

HeadIsClosed

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

Nominal deadline until which contest can be submitted and after which fanout is possible. NOTE: Use this only for informational purpose and wait for ReadyToFanout instead before sending Fanout as the ledger of our cardano-node might not have progressed sufficiently in time yet and we do not re-submit transactions (yet).

HeadIsContested

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

ReadyToFanout

Fields

headId :: HeadId

HeadIsAborted

Fields

headId :: HeadId

utxo :: UTxOType tx

HeadIsFinalized

Fields

headId :: HeadId

utxo :: UTxOType tx

CommandFailed

Fields

clientInput :: ClientInput tx

TxValid
Given transaction has been seen as valid in the Head. It is expected to eventually be part of a SnapshotConfirmed.

Fields

headId :: HeadId

transaction :: tx

TxInvalid
Given transaction was not not applicable to the given UTxO in time and has been dropped.

Fields

headId :: HeadId

utxo :: UTxOType tx

transaction :: tx

validationError :: ValidationError

SnapshotConfirmed
Given snapshot was confirmed and included transactions can be considered final.

Fields

headId :: HeadId

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

GetUTxOResponse

Fields

headId :: HeadId

utxo :: UTxOType tx

InvalidInput

Fields

reason :: String

input :: Text

Greetings
A friendly welcome message which tells a client something about the node. Currently used for knowing what signing key the server uses (it only knows one), HeadStatus and optionally (if HeadIsOpen or SnapshotConfirmed message is emitted) UTxO's present in the Hydra Head.

Fields

me :: Party

headStatus :: HeadStatus

snapshotUtxo :: Maybe (UTxOType tx)

hydraNodeVersion :: String

PostTxOnChainFailed

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

data OutputFormat

Possible transaction formats in the api server output

Constructors

OutputCBOR

OutputJSON

data WithUTxO

Whether or not to include full UTxO in server outputs.

Constructors

WithUTxO

WithoutUTxO

data ServerOutputConfig

Constructors

ServerOutputConfig

Fields

txOutputFormat :: OutputFormat

utxoInSnapshot :: WithUTxO

prepareServerOutput

Arguments

:: IsChainState tx

=> ServerOutputConfig
Decide on tx representation

-> TimedServerOutput tx
Server output

-> ByteString
Final output

Replaces the json encoded tx field with it's cbor representation.

NOTE: we deliberately pattern match on all ServerOutput constructors in handleTxOutput so that we don't forget to update this function if they change.

data HeadStatus

All possible Hydra states displayed in the API server outputs.

Constructors

Idle

Initializing

Open

Closed

FanoutPossible

Final

projectHeadStatus :: HeadStatus -> ServerOutput tx -> HeadStatus

Projection function related to headStatus field in Greetings message.

projectSnapshotUtxo :: Maybe (UTxOType tx) -> ServerOutput tx -> Maybe (UTxOType tx)

Projection function related to snapshotUtxo field in Greetings message.