Hydra.Chain

Documentation

data HeadParameters

Contains the head's parameters as established in the initial transaction.

Constructors

HeadParameters

Fields

contestationPeriod :: DiffTime

parties :: [Party]

type ContestationPeriod = DiffTime

data OnChainTx tx

Constructors

InitTx HeadParameters

CommitTx Party (Utxo tx)

AbortTx (Utxo tx)

CollectComTx (Utxo tx)

CloseTx (Snapshot tx)

ContestTx (Snapshot tx)

FanoutTx (Utxo tx)

data ChainError

Constructors

ChainError

newtype Chain tx m

Handle to interface with the main chain network

Constructors

Chain

Fields

postTx :: MonadThrow m => OnChainTx tx -> m ()

Construct and send a transaction to the main chain corresponding to the given OnChainTx event. Does at least throw ChainError.

type ChainCallback tx m = OnChainTx tx -> m ()

Handle to interface observed transactions.

type ChainComponent tx m a = ChainCallback tx m -> (Chain tx m -> m a) -> m a

A type tying both posting and observing transactions into a single Component.