Hydra.Chain

Description

Specifies the Head-Chain Interaction part of the protocol

Incoming and outgoing on-chain transactions are modelled respectively as OnChainTx and PostChainTx which are data type that abstracts away the details of the structure of the transaction.

Documentation

data HeadParameters

Contains the head's parameters as established in the initial transaction.

Constructors

HeadParameters

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

data PostChainTx tx

Data type used to post transactions on chain. It holds everything to construct corresponding Head protocol transactions.

Constructors

InitTx

Fields

headParameters :: HeadParameters

CommitTx

Fields

party :: Party

committed :: UTxOType tx

AbortTx

Fields

utxo :: UTxOType tx

CollectComTx

Fields

utxo :: UTxOType tx

CloseTx

Fields

confirmedSnapshot :: ConfirmedSnapshot tx

ContestTx

Fields

confirmedSnapshot :: ConfirmedSnapshot tx

FanoutTx

Fields

utxo :: UTxOType tx

newtype HeadId

Uniquely identifies a Hydra Head.

Constructors

HeadId ByteString

data OnChainTx tx

Describes transactions as seen on chain. Holds as minimal information as possible to simplify observing the chain.

Constructors

OnInitTx

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

OnCommitTx

Fields

party :: Party

committed :: UTxOType tx

OnAbortTx

OnCollectComTx

OnCloseTx

Fields

snapshotNumber :: SnapshotNumber

remainingContestationPeriod :: NominalDiffTime

The remaining contestation period in wall clock time calculated from the actual upper bound of the close transaction observed.

OnContestTx

Fields

snapshotNumber :: SnapshotNumber

OnFanoutTx

data PostTxError tx

Exceptions thrown by $sel:postTx:Chain.

Constructors

MoreThanOneUTxOCommitted

CannotSpendInput

Fields

input :: Text

walletUTxO :: UTxOType tx

headUTxO :: UTxOType tx

CannotCoverFees

Fields

walletUTxO :: UTxOType tx

headUTxO :: UTxOType tx

reason :: Text

tx :: tx

CannotFindOwnInitial

Fields

knownUTxO :: UTxOType tx

FailedToPostTx

Fields

failureReason :: Text

PlutusValidationFailed

Fields

plutusFailure :: Text

plutusDebugInfo :: Text

NoSeedInput

NoPaymentInput

InvalidStateToPost

Fields

txTried :: PostChainTx tx

UnsupportedLegacyOutput

Fields

byronAddress :: Address ByronAddr

newtype Chain tx m

Handle to interface with the main chain network

Constructors

Chain

Fields

postTx :: MonadThrow m => PostChainTx tx -> m ()

Construct and send a transaction to the main chain corresponding to the given OnChainTx event. This function is not expected to block, so it is only responsible for submitting, but it should validate the created transaction against a reasonable local view of the chain and throw an exception when invalid.

Does at least throw PostTxError.

data ChainEvent tx

Constructors

Observation (OnChainTx tx)

Rollback Word

type ChainCallback tx m = ChainEvent tx -> m ()

Handle to interface observed transactions.

type ChainComponent tx m a = ChainCallback tx m -> (Chain tx m -> m a) -> m a

A type tying both posting and observing transactions into a single Component.