Hydra.Chain.CardanoClient

Description

A basic cardano-node client that can talk to a local cardano-node.

The idea of this module is to provide a Haskell interface on top of cardano-cli's API, using cardano-api types.

Documentation

data QueryException

Constructors

QueryAcquireException AcquiringFailure

QueryEraMismatchException EraMismatch

QueryUnsupportedNtcVersionException UnsupportedNtcVersionError

QueryProtocolParamsConversionException ProtocolParametersConversionError

QueryProtocolParamsEraNotSupported AnyCardanoEra

QueryProtocolParamsEncodingFailureOnEra AnyCardanoEra Text

QueryEraNotInCardanoModeFailure AnyCardanoEra

QueryNotShelleyBasedEraException AnyCardanoEra

CardanoClient handle

data CardanoClient

Handle interface for abstract querying of a cardano node.

Constructors

CardanoClient

Fields

queryUTxOByAddress :: [Address ShelleyAddr] -> IO UTxO

networkId :: NetworkId

mkCardanoClient :: NetworkId -> SocketPath -> CardanoClient

Construct a CardanoClient handle.

Tx Construction / Submission

submitTransaction

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> Tx
A signed transaction.

-> IO ()

Submit a (signed) transaction to the node.

Throws SubmitTransactionException if submission fails.

data SubmitTransactionException

Exceptions that can occur during a transaction submission.

In principle, we can only encounter an EraMismatch at era boundaries, when we try to submit a "next era" transaction as a "current era" transaction, or vice-versa. Similarly, TxValidationError shouldn't occur given that the transaction was safely constructed through buildTransaction.

Constructors

SubmitEraMismatch EraMismatch

SubmitTxValidationError TxValidationErrorInCardanoMode

awaitTransaction

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> Tx
The transaction to watch / await

-> IO UTxO

Await until the given transaction is visible on-chain. Returns the UTxO set produced by that transaction.

Note that this function loops forever; hence, one probably wants to couple it with a surrounding timeout.

Local state query

data QueryPoint

Describes whether to query at the tip or at a specific point.

Constructors

QueryTip

QueryAt ChainPoint

queryTip :: NetworkId -> SocketPath -> IO ChainPoint

Query the latest chain point aka "the tip".

querySystemStart :: NetworkId -> SocketPath -> QueryPoint -> IO SystemStart

Query the system start parameter at given point.

Throws at least QueryException if query fails.

queryEraHistory :: NetworkId -> SocketPath -> QueryPoint -> IO EraHistory

Query the era history at given point.

Throws at least QueryException if query fails.

queryEpochNo :: NetworkId -> SocketPath -> QueryPoint -> IO EpochNo

Query the current epoch number.

Throws at least QueryException if query fails.

queryProtocolParameters

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> QueryPoint

-> IO (PParams LedgerEra)

Query the protocol parameters at given point and convert them to Babbage era protocol parameters.

Throws at least QueryException if query fails.

queryGenesisParameters

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> QueryPoint

-> IO (GenesisParameters ShelleyEra)

Query GenesisParameters at a given point.

Throws at least QueryException if query fails.

queryUTxO :: NetworkId -> SocketPath -> QueryPoint -> [Address ShelleyAddr] -> IO UTxO

Query UTxO for all given addresses at given point.

Throws at least QueryException if query fails.

queryUTxOExpr :: ShelleyBasedEra era -> [Address ShelleyAddr] -> LocalStateQueryExpr b p QueryInMode r IO UTxO

queryUTxOByTxIn

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> QueryPoint

-> [TxIn]

-> IO UTxO

Query UTxO for given tx inputs at given point.

Throws at least QueryException if query fails.

assumeShelleyBasedEraOrThrow :: MonadThrow m => CardanoEra era -> m (ShelleyBasedEra era)

queryUTxOWhole

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> QueryPoint

-> IO UTxO

Query the whole UTxO from node at given point. Useful for debugging, but should obviously not be used in production code.

Throws at least QueryException if query fails.

queryUTxOFor :: NetworkId -> SocketPath -> QueryPoint -> VerificationKey PaymentKey -> IO UTxO

Query UTxO for the address of given verification key at point.

Throws at least QueryException if query fails.

queryStakePools

Arguments

:: NetworkId
Current network discriminant

-> SocketPath
Filepath to the cardano-node's domain socket

-> QueryPoint

-> IO (Set PoolId)

Query the current set of registered stake pools.

Throws at least QueryException if query fails.

Helpers

queryCurrentEraExpr :: LocalStateQueryExpr b p QueryInMode r IO AnyCardanoEra

Monadic query expression to get current era.

queryInShelleyBasedEraExpr

Arguments

:: ShelleyBasedEra era
The current running era we can use to query the node

-> QueryInShelleyBasedEra era a

-> LocalStateQueryExpr b p QueryInMode r IO a

Monadic query expression for a QueryInShelleyBasedEra.

runQuery :: NetworkId -> SocketPath -> QueryPoint -> QueryInMode a -> IO a

Throws at least QueryException if query fails.

runQueryExpr :: NetworkId -> SocketPath -> QueryPoint -> LocalStateQueryExpr BlockInMode ChainPoint QueryInMode () IO a -> IO a

Throws at least QueryException if query fails.

throwOnEraMismatch :: MonadThrow m => Either EraMismatch a -> m a

throwOnUnsupportedNtcVersion :: MonadThrow m => Either UnsupportedNtcVersionError a -> m a

localNodeConnectInfo :: NetworkId -> SocketPath -> LocalNodeConnectInfo

cardanoModeParams :: ConsensusModeParams