Hydra.Chain.CardanoClient

Description

A basic cardano-node client that can talk to a local cardano-node.

The idea of this module is to provide a Haskell interface on top of cardano-cli's API, using cardano-api types.

Documentation

type NodeSocket = FilePath

newtype QueryException

Constructors

QueryException Text

CardanoClient handle

data CardanoClient

Handle interface for abstract querying of a cardano node.

Constructors

CardanoClient

Fields

queryUTxOByAddress :: [Address ShelleyAddr] -> IO UTxO

networkId :: NetworkId

mkCardanoClient :: NetworkId -> NodeSocket -> CardanoClient

Construct a CardanoClient handle.

Local state query

data QueryPoint

Describes whether to query at the tip or at a specific point.

Constructors

QueryTip

QueryAt ChainPoint

queryTip :: NetworkId -> FilePath -> IO ChainPoint

Query the latest chain point aka "the tip".

queryTipSlotNo :: NetworkId -> FilePath -> IO SlotNo

Query the latest chain point just for the slot number.

querySystemStart :: NetworkId -> FilePath -> QueryPoint -> IO SystemStart

Query the system start parameter at given point.

Throws at least QueryException if query fails.

queryEraHistory :: NetworkId -> FilePath -> QueryPoint -> IO (EraHistory CardanoMode)

Query the era history at given point.

Throws at least QueryException if query fails.

queryProtocolParameters :: NetworkId -> FilePath -> QueryPoint -> IO ProtocolParameters

Query the protocol parameters at given point.

Throws at least QueryException if query fails.

queryUTxO :: NetworkId -> FilePath -> QueryPoint -> [Address ShelleyAddr] -> IO UTxO

Query UTxO for all given addresses at given point.

Throws at least QueryException if query fails.

queryUTxOByTxIn :: NetworkId -> FilePath -> QueryPoint -> [TxIn] -> IO UTxO

Query UTxO for given tx inputs at given point.

Throws at least QueryException if query fails.

queryUTxOWhole :: NetworkId -> FilePath -> QueryPoint -> IO UTxO

Query the whole UTxO from node at given point. Useful for debugging, but should obviously not be used in production code.

Throws at least QueryException if query fails.

runQuery :: NetworkId -> FilePath -> QueryPoint -> QueryInMode CardanoMode a -> IO a

Throws at least QueryException if query fails.

Helpers

throwOnEraMismatch :: MonadThrow m => Either EraMismatch a -> m a

localNodeConnectInfo :: NetworkId -> FilePath -> LocalNodeConnectInfo CardanoMode

cardanoModeParams :: ConsensusModeParams CardanoMode