Hydra.Chain.Direct

Description

Chain component implementation which uses directly the Node-to-Client protocols to submit "hand-rolled" transactions.

Documentation

newtype NetworkMagic

Constructors

NetworkMagic Word32

newtype DirectBackend

Constructors

DirectBackend

Fields

options :: DirectOptions

type BlockType = BlockInMode

The block type used in the node-to-client protocols.

data DirectConnectException

Constructors

DirectConnectException

Fields

ioException :: IOException

nodeSocket :: SocketPath

networkId :: NetworkId

newtype IntersectionNotFoundException

Thrown when the user-provided custom point of intersection is unknown to the local node. This may happen if users shut down their node quickly after starting them and hold on a not-so-stable point of the chain. When they turn the node back on, that point may no longer exist on the network if a fork with deeper roots has been adopted in the meantime.

Constructors

IntersectionNotFound

Fields

requestedPoint :: ChainPoint

data EraNotSupportedException

Constructors

EraNotSupportedAnymore

Fields

otherEraName :: Text

EraNotSupportedYet

Fields

otherEraName :: Text

withDirectChain

Arguments

:: DirectBackend

-> Tracer IO CardanoChainLog

-> CardanoChainConfig

-> ChainContext

-> TinyWallet IO

-> ChainStateHistory Tx
Chain state loaded from persistence.

-> ChainComponent Tx IO a

chainSyncClient :: forall m. (MonadSTM m, MonadThrow m) => ChainSyncHandler m -> TinyWallet m -> ChainPoint -> ChainSyncClient BlockType ChainPoint ChainTip m ()

txSubmissionClient :: forall m. (MonadSTM m, MonadDelay m) => Tracer m CardanoChainLog -> TQueue m (Tx, TMVar m (Maybe (PostTxError Tx))) -> LocalTxSubmissionClient TxInMode TxValidationErrorInCardanoMode m ()