Hydra.Chain.Direct

Description

Chain component implementation which uses directly the Node-to-Client protocols to submit "hand-rolled" transactions including Plutus validators and observing the chain using it as well.

Documentation

newtype NetworkMagic

Constructors

NetworkMagic Word32

withIOManager :: WithIOManager

data DirectChainLog

Constructors

ToPost

Fields

toPost :: PostChainTx Tx

PostingTx

Fields

postedTx :: (TxId StandardCrypto, ValidatedTx Era)

PostedTx

Fields

postedTxId :: TxId StandardCrypto

ReceivedTxs

Fields

onChainTxs :: [OnChainTx Tx]

receivedTxs :: [(TxId StandardCrypto, ValidatedTx Era)]

RolledBackward

Fields

point :: SomePoint

Wallet TinyWalletLog

data IntersectionNotFoundException

Constructors

IntersectionNotFound

Fields

requestedPoint :: Point Block

data ConnectException

Constructors

ConnectException

Fields

ioException :: IOException

socketPath :: FilePath

networkId :: NetworkId

withDirectChain

Arguments

:: Tracer IO DirectChainLog
Tracer for logging

-> NetworkId
Network identifer to which we expect to connect.

-> IOManager
A cross-platform abstraction for managing I/O operations on local sockets

-> FilePath
Path to a domain socket used to connect to the server.

-> (VerificationKey PaymentKey, SigningKey PaymentKey)
Key pair for the wallet.

-> Party
Hydra party of our hydra node.

-> [VerificationKey PaymentKey]
Cardano keys of all Head participants (including our key pair).

-> Maybe ChainPoint
Point at which to start following the chain.

-> ChainComponent Tx IO a

client :: (MonadST m, MonadTimer m, MonadThrow m) => Tracer m DirectChainLog -> Maybe (Point Block) -> TQueue m (ValidatedTx Era, TMVar m (Maybe (PostTxError Tx))) -> TVar m SomeOnChainHeadState -> ChainCallback Tx m -> NodeToClientVersion -> OuroborosApplication 'InitiatorMode LocalAddress LByteString m () Void

chainSyncClient :: forall m. (MonadSTM m, MonadThrow m) => Tracer m DirectChainLog -> ChainCallback Tx m -> TVar m SomeOnChainHeadState -> Maybe (Point Block) -> ChainSyncClient Block (Point Block) (Tip Block) m ()

txSubmissionClient :: forall m. MonadSTM m => Tracer m DirectChainLog -> TQueue m (ValidatedTx Era, TMVar m (Maybe (PostTxError Tx))) -> LocalTxSubmissionClient (GenTx Block) (ApplyTxErr Block) m ()

finalizeTx :: (MonadSTM m, MonadThrow (STM m)) => TinyWallet m -> TVar m SomeOnChainHeadState -> ValidatedTx Era -> STM m (ValidatedTx Era)

Balance and sign the given partial transaction.

XXX(SN): This function does retry when no payment UTXO was found and thus might block. This is necessary in some situations when the wallet has not yet "seen" the change output although the direct chain client observed a transaction already. This is a smell and we should try to avoid having this race condition in the first place.

fromPostChainTx :: (MonadSTM m, MonadThrow (STM m)) => [VerificationKey PaymentKey] -> TinyWallet m -> TVar m SomeOnChainHeadState -> PostChainTx Tx -> STM m Tx

getAlonzoTxs :: Block -> StrictSeq (ValidatedTx Era)

This extract Alonzo transactions from a block. If the block wasn't produced in the Alonzo era, it returns a empty sequence.