Hydra.Chain.Direct

Description

Chain component implementation which uses directly the Node-to-Client protocols to submit "hand-rolled" transactions including Plutus validators and observing the chain using it as well.

Documentation

newtype NetworkMagic

Constructors

NetworkMagic Word32

data DirectChainLog

withDirectChain

Arguments

:: Tracer IO DirectChainLog
Tracer for logging

-> NetworkMagic
Network identifer to which we expect to connect.

-> FilePath
Path to a domain socket used to connect to the server.

-> ChainComponent tx IO ()

client :: (MonadST m, MonadTimer m) => TQueue m (PostChainTx tx) -> ChainCallback tx m -> NodeToClientVersion -> OuroborosApplication 'InitiatorMode LocalAddress LByteString m () Void

chainSyncClient :: forall m tx. Monad m => ChainCallback tx m -> ChainSyncClient Block (Point Block) (Tip Block) m ()

txSubmissionClient :: forall m tx. MonadSTM m => TQueue m (PostChainTx tx) -> LocalTxSubmissionClient (GenTx Block) (ApplyTxErr Block) m ()

withMockServer

Arguments

:: NetworkMagic
Network identifer to which we expect to connect.

-> FilePath
Path to a domain socket on which to listen.

-> IO a
Action to run in-between.

-> IO a

mockServer :: (MonadST m, MonadTimer m) => TVar m [ValidatedTx Era] -> NodeToClientVersion -> OuroborosApplication 'ResponderMode LocalAddress LByteString m Void ()

mockChainSyncServer :: forall m. MonadSTM m => TVar m [ValidatedTx Era] -> ChainSyncServer Block (Point Block) (Tip Block) m ()

mockTxSubmissionServer :: MonadSTM m => TVar m [ValidatedTx Era] -> LocalTxSubmissionServer (GenTx Block) reject m ()

getAlonzoTxs :: Block -> StrictSeq (ValidatedTx Era)

This extract Alonzo transactions from a block. If the block wasn't produced in the Alonzo era, it returns a empty sequence.