Hydra.Chain.Direct

Description

Chain component implementation which uses directly the Node-to-Client protocols to submit "hand-rolled" transactions.

Documentation

newtype NetworkMagic

Constructors

NetworkMagic Word32

withIOManager :: WithIOManager

newtype IntersectionNotFoundException

Thrown when the user-provided custom point of intersection is unknown to the local node. This may happen if users shut down their node quickly after starting them and hold on a not-so-stable point of the chain. When they turn the node back on, that point may no longer exist on the network if a fork with deeper roots has been adopted in the meantime.

Constructors

IntersectionNotFound

Fields

requestedPoint :: Point Block

data ConnectException

Constructors

ConnectException

Fields

ioException :: IOException

nodeSocket :: FilePath

networkId :: NetworkId

initialChainState :: ChainStateType Tx

Defines the starting state of the direct chain layer.

loadChainContext

Arguments

:: ChainConfig

-> Party
Hydra party of our hydra node.

-> TxId
Transaction id at which to look for Hydra scripts.

-> IO ChainContext

Build the ChainContext from a ChainConfig and additional information.

withDirectChain :: Tracer IO DirectChainLog -> ChainConfig -> ChainContext -> Maybe ChainPoint -> ChainComponent Tx IO a

ouroborosApplication :: (MonadST m, MonadTimer m, MonadThrow m) => Tracer m DirectChainLog -> Point Block -> TQueue m (ValidatedTx LedgerEra, TMVar m (Maybe (PostTxError Tx))) -> ChainSyncHandler m -> TinyWallet m -> NodeToClientVersion -> OuroborosApplication 'InitiatorMode LocalAddress LByteString m () Void

chainSyncClient :: forall m. (MonadSTM m, MonadThrow m) => ChainSyncHandler m -> TinyWallet m -> Point Block -> ChainSyncClient Block (Point Block) (Tip Block) m ()

txSubmissionClient :: forall m. MonadSTM m => Tracer m DirectChainLog -> TQueue m (ValidatedTx LedgerEra, TMVar m (Maybe (PostTxError Tx))) -> LocalTxSubmissionClient (GenTx Block) (ApplyTxErr Block) m ()