Hydra.Chain.Direct

Description

Chain component implementation which uses directly the Node-to-Client protocols to submit "hand-rolled" transactions.

Documentation

newtype NetworkMagic

Constructors

NetworkMagic Word32

withIOManager :: WithIOManager

newtype IntersectionNotFoundException

Thrown when the user-provided custom point of intersection is unknown to the local node. This may happen if users shut down their node quickly after starting them and hold on a not-so-stable point of the chain. When they turn the node back on, that point may no longer exist on the network if a fork with deeper roots has been adopted in the meantime.

Constructors

IntersectionNotFound

Fields

requestedPoint :: Point Block

data ConnectException

Constructors

ConnectException

Fields

ioException :: IOException

socketPath :: FilePath

networkId :: NetworkId

withDirectChain

Arguments

:: Tracer IO DirectChainLog
Tracer for logging

-> NetworkId
Network identifer to which we expect to connect.

-> IOManager
A cross-platform abstraction for managing I/O operations on local sockets

-> FilePath
Path to a domain socket used to connect to the server.

-> (VerificationKey PaymentKey, SigningKey PaymentKey)
Key pair for the wallet.

-> Party
Hydra party of our hydra node.

-> [VerificationKey PaymentKey]
Cardano keys of all Head participants (including our key pair).

-> Maybe ChainPoint
Point at which to start following the chain.

-> ChainComponent Tx IO a

queryTimeHandle :: NetworkId -> FilePath -> IO TimeHandle

Query ad-hoc epoch, system start and protocol parameters to determine current point in time.

ouroborosApplication :: (MonadST m, MonadTimer m, MonadThrow m) => Tracer m DirectChainLog -> Maybe (Point Block) -> TQueue m (ValidatedTx LedgerEra, TMVar m (Maybe (PostTxError Tx))) -> ChainSyncHandler m -> TinyWallet m -> NodeToClientVersion -> OuroborosApplication 'InitiatorMode LocalAddress LByteString m () Void

chainSyncClient :: forall m. (MonadSTM m, MonadThrow m) => ChainSyncHandler m -> TinyWallet m -> Maybe (Point Block) -> ChainSyncClient Block (Point Block) (Tip Block) m ()

txSubmissionClient :: forall m. MonadSTM m => Tracer m DirectChainLog -> TQueue m (ValidatedTx LedgerEra, TMVar m (Maybe (PostTxError Tx))) -> LocalTxSubmissionClient (GenTx Block) (ApplyTxErr Block) m ()