Hydra.Chain.Direct.State

Description

Contains the a state-ful interface to transaction construction and observation.

It defines the 'ChainStateType tx' to be used in the Direct layer and it's constituents.

Documentation

class HasKnownUTxO a where

A class for accessing the known UTxO set in a type. This is useful to get all the relevant UTxO for resolving transaction inputs.

Methods

getKnownUTxO :: a -> UTxO

States & transitions

data ChainStateAt

The chain state used by the Hydra.Chain.Direct implementation. It records the actual ChainState paired with a ChainSlot (used to know up to which point to rewind on rollbacks).

Constructors

ChainStateAt

Fields

chainState :: ChainState

recordedAt :: Maybe ChainPoint

chainSlotFromPoint :: ChainPoint -> ChainSlot

Get a generic ChainSlot from a Cardano ChainPoint. Slot 0 is used for the genesis point.

data ChainTransition

A definition of all transitions between ChainStates. Enumerable and bounded to be used as labels for checking coverage.

Constructors

Init

Commit

Collect

Close

Contest

Fanout

data ChainState

An enumeration of all possible on-chain states of a Hydra Head, where each case stores the relevant information to construct & observe transactions to other states.

Constructors

Idle
The idle state does not contain any head-specific information and exists to be used as a starting and terminal state.

Initial InitialState

Open OpenState

Closed ClosedState

data ChainContext

Read-only chain-specific data. This is different to HydraContext as it only provide contains data known to single peer.

Constructors

ChainContext

Fields

networkId :: NetworkId

peerVerificationKeys :: [VerificationKey PaymentKey]

ownVerificationKey :: VerificationKey PaymentKey

ownParty :: Party

scriptRegistry :: ScriptRegistry

allVerificationKeys :: ChainContext -> [VerificationKey PaymentKey]

Get all cardano verification keys available in the chain context.

data InitialState

Constructors

InitialState

Fields

initialThreadOutput :: InitialThreadOutput

initialInitials :: [UTxOWithScript]

initialCommits :: [UTxOWithScript]

initialHeadId :: HeadId

initialHeadTokenScript :: PlutusScript

data OpenState

Constructors

OpenState

Fields

openThreadOutput :: OpenThreadOutput

openHeadId :: HeadId

openHeadTokenScript :: PlutusScript

openUtxoHash :: UTxOHash

data ClosedState

Constructors

ClosedState

Fields

closedThreadOutput :: ClosedThreadOutput

closedHeadId :: HeadId

closedHeadTokenScript :: PlutusScript

Constructing transactions

initialize

Arguments

:: ChainContext

-> HeadParameters

-> TxIn
Seed input.

-> Tx

Construct an init transaction given some general ChainContext, the HeadParameters and a seed TxIn which will be spent.

commit :: ChainContext -> InitialState -> UTxO -> Either (PostTxError Tx) Tx

Construct a commit transaction based on the InitialState. This does look for "our initial output" to spend and check the given UTxO to be compatible. Hence, this function does fail if already committed.

abort :: HasCallStack => ChainContext -> InitialState -> Tx

Construct a collect transaction based on the InitialState. This will reimburse all the already committed outputs.

collect :: ChainContext -> InitialState -> Tx

Construct a collect transaction based on the InitialState. This will know collect all the committed outputs.

close :: ChainContext -> OpenState -> ConfirmedSnapshot Tx -> PointInTime -> Tx

Construct a close transaction based on the OpenState and a confirmed snapshot. The given PointInTime will be used as an upper validity bound and will define the start of the contestation period.

contest :: ChainContext -> ClosedState -> ConfirmedSnapshot Tx -> PointInTime -> Tx

Construct a contest transaction based on the ClosedState and a confirmed snapshot. The given PointInTime will be used as an upper validity bound and needs to be before the deadline.

fanout

Arguments

:: ClosedState

-> UTxO

-> SlotNo
Contestation deadline as SlotNo, used to set lower tx validity bound.

-> Tx

Construct a fanout transaction based on the ClosedState and off-chain agreed UTxO set to fan out.

Observing Transitions

observeSomeTx :: ChainContext -> ChainState -> Tx -> Maybe (OnChainTx Tx, ChainState)

Observe a transition without knowing the starting or ending state. This function should try to observe all relevant transitions given some ChainState.

IdleState transitions

observeInit :: ChainContext -> Tx -> Maybe (OnChainTx Tx, InitialState)

Observe an init transition using a InitialState and observeInitTx.

InitialState transitions

observeCommit :: ChainContext -> InitialState -> Tx -> Maybe (OnChainTx Tx, InitialState)

Observe an commit transition using a InitialState and observeCommitTx.

observeCollect :: InitialState -> Tx -> Maybe (OnChainTx Tx, OpenState)

Observe an collect transition using a InitialState and observeCollectComTx. This function checks the head id and ignores if not relevant.

observeAbort :: InitialState -> Tx -> Maybe (OnChainTx Tx)

Observe an abort transition using a InitialState and observeAbortTx.

OpenState transitions

observeClose :: OpenState -> Tx -> Maybe (OnChainTx Tx, ClosedState)

Observe a close transition using a OpenState and observeCloseTx. This function checks the head id and ignores if not relevant.

ClosedState transitions

observeContest :: ClosedState -> Tx -> Maybe (OnChainTx Tx, ClosedState)

Observe a fanout transition using a ClosedState and observeContestTx. This function checks the head id and ignores if not relevant.

observeFanout :: ClosedState -> Tx -> Maybe (OnChainTx Tx)

Observe a fanout transition using a ClosedState and observeFanoutTx.

Generators

maxGenParties :: Int

Maximum number of parties used in the generators.

maxGenAssets :: Int

Maximum number of assets (ADA or other tokens) used in the generators.

genChainState :: Gen ChainState

Generate a ChainState within known limits above.

genChainStateWithTx :: Gen (ChainContext, ChainState, Tx, ChainTransition)

Generate a ChainContext and ChainState within the known limits above, along with a transaction that results in a transition away from it.

Warning zone

data HydraContext

Define some global context from which generators can pick values for generation. This allows to write fairly independent generators which however still make sense with one another within the context of a head.

For example, one can generate a head's _party_ from that global list, whereas other functions may rely on all parties and thus, we need both generation to be coherent.

Do not use this in production code, but only for generating test data.

Constructors

HydraContext

Fields

ctxVerificationKeys :: [VerificationKey PaymentKey]

ctxHydraSigningKeys :: [SigningKey HydraKey]

ctxNetworkId :: NetworkId

ctxContestationPeriod :: ContestationPeriod

ctxParties :: HydraContext -> [Party]

ctxHeadParameters :: HydraContext -> HeadParameters

genHydraContext :: Int -> Gen HydraContext

Generate a HydraContext for a bounded arbitrary number of parties.

maxParties sets the upper bound in the number of parties in the Head.

genHydraContextFor :: Int -> Gen HydraContext

Generate a HydraContext for a given number of parties.

deriveChainContexts :: HydraContext -> Gen [ChainContext]

Get all peer-specific ChainContexts from a HydraContext. NOTE: This assumes that HydraContext has same length $sel:ctxVerificationKeys:HydraContext and $sel:ctxHydraSigningKeys:HydraContext.

pickChainContext :: HydraContext -> Gen ChainContext

Pick one of the participants and derive the peer-specific ChainContext from a HydraContext. NOTE: This assumes that HydraContext has same length $sel:ctxVerificationKeys:HydraContext and $sel:ctxHydraSigningKeys:HydraContext.

genStInitial :: HydraContext -> Gen (ChainContext, InitialState)

genInitTx :: HydraContext -> Gen Tx

genCommits :: HydraContext -> Tx -> Gen [Tx]

genCommit :: Gen UTxO

genCollectComTx :: Gen (ChainContext, [UTxO], InitialState, Tx)

genCloseTx :: Int -> Gen (ChainContext, OpenState, Tx, ConfirmedSnapshot Tx)

genContestTx :: Gen (HydraContext, PointInTime, ClosedState, Tx)

genFanoutTx :: Int -> Int -> Gen (HydraContext, ClosedState, Tx)

getContestationDeadline :: ClosedState -> UTCTime

genStOpen :: HydraContext -> Gen (UTxO, OpenState)

genStClosed :: HydraContext -> UTxO -> Gen (SnapshotNumber, UTxO, ClosedState)

Danger zone

unsafeCommit :: HasCallStack => ChainContext -> InitialState -> UTxO -> Tx

unsafeObserveInitAndCommits :: ChainContext -> Tx -> [Tx] -> ([UTxO], InitialState)

Orphan instances

IsChainState Tx

Instance details

Associated Types

type ChainStateType Tx = (c :: Type)

Methods

chainStateSlot :: ChainStateType Tx -> ChainSlot