Hydra.Chain.Direct.TimeHandle

Description

Module to deal with time in direct cardano chain layer. Defines the type for a PointInTime and a means to acquire one via a TimeHandle and queryTimeHandle.

Documentation

type PointInTime = (SlotNo, POSIXTime)

data TimeHandle

Constructors

TimeHandle

Fields

currentPointInTime :: Either Text PointInTime

Get the current PointInTime

adjustPointInTime :: SlotNo -> PointInTime -> Either Text PointInTime

Adjust a PointInTime by some number of slots, positively or negatively.

queryTimeHandle :: NetworkId -> FilePath -> IO TimeHandle

Compute current Query PointInTime from wall clock and by querying the node for era history, system start and protocol parameters. Current means within somewhere within

NOTE: We use wall clock because the slot in tip is from the last block and potentially too far in the past.