Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Documentation

type UTxOWithScript = (TxIn, TxOut CtxUTxO, HashableScriptData)

Needed on-chain data to create Head transactions.

newtype UTxOHash

Constructors

UTxOHash ByteString

data InitialThreadOutput

Representation of the Head output after an Init transaction.

Constructors

InitialThreadOutput

Fields

initialThreadUTxO :: (TxIn, TxOut CtxUTxO)

initialContestationPeriod :: ContestationPeriod

initialParties :: [Party]

getHydraHeadV1TxName :: Tx -> Maybe Text

Get the metadata entry to identify Hydra transactions (for informational purposes).

Observe Hydra Head transactions

data HeadObservation

Generalised type for arbitrary Head observations on-chain.

Constructors

NoHeadTx

Init InitObservation

Abort AbortObservation

Commit CommitObservation

CollectCom CollectComObservation

Decrement DecrementObservation

Close CloseObservation

Contest ContestObservation

Fanout FanoutObservation

observeHeadTx :: NetworkId -> UTxO -> Tx -> HeadObservation

Observe any Hydra head transaction.

data InitObservation

Data which can be observed from an initTx.

Constructors

InitObservation

Fields

initialThreadUTxO :: (TxIn, TxOut CtxUTxO)

initials :: [(TxIn, TxOut CtxUTxO)]

headId :: HeadId

seedTxIn :: TxIn

contestationPeriod :: ContestationPeriod

parties :: [Party]

participants :: [OnChainId]

data NotAnInitReason

Constructors

NoHeadOutput

NotAHeadDatum

NoSTFound

NotAHeadPolicy

observeInitTx :: Tx -> Either NotAnInitReason InitObservation

Identify a init tx by checking the output value for holding tokens that are valid head tokens (checked by seed + policy).

data CommitObservation

Full observation of a commit transaction.

Constructors

CommitObservation

Fields

commitOutput :: (TxIn, TxOut CtxUTxO)

party :: Party

Hydra participant who committed the UTxO.

committed :: UTxO

headId :: HeadId

observeCommitTx

Arguments

:: NetworkId

-> UTxO
A UTxO set to lookup tx inputs. Should at least contain the input spending from νInitial.

-> Tx

-> Maybe CommitObservation

Identify a commit tx by:

Check that its spending from the init validator,

Find the outputs which pays to the commit validator,

Using the datum of that output, deserialize the committed output,

Reconstruct the committed UTxO from both values (tx input and output).

data CollectComObservation

Constructors

CollectComObservation

Fields

threadOutput :: OpenThreadOutput

headId :: HeadId

utxoHash :: UTxOHash

observeCollectComTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CollectComObservation

Identify a collectCom tx by lookup up the input spending the Head output and decoding its redeemer.

data DecrementObservation

Constructors

DecrementObservation

Fields

headId :: HeadId

newVersion :: SnapshotVersion

distributedOutputs :: [TxOut CtxUTxO]

observeDecrementTx :: UTxO -> Tx -> Maybe DecrementObservation

data CloseObservation

Constructors

CloseObservation

Fields

threadOutput :: ClosedThreadOutput

headId :: HeadId

snapshotNumber :: SnapshotNumber

observeCloseTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CloseObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data ContestObservation

Constructors

ContestObservation

Fields

contestedThreadOutput :: (TxIn, TxOut CtxUTxO)

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

contesters :: [PubKeyHash]

observeContestTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe ContestObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

newtype FanoutObservation

Constructors

FanoutObservation

Fields

headId :: HeadId

observeFanoutTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe FanoutObservation

Identify a fanout tx by lookup up the input spending the Head output and decoding its redeemer.

newtype AbortObservation

Constructors

AbortObservation

Fields

headId :: HeadId

observeAbortTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe AbortObservation

Identify an abort tx by looking up the input spending the Head output and decoding its redeemer.

Cardano specific identifiers

currencySymbolToHeadId :: MonadFail m => CurrencySymbol -> m HeadId

headIdToPolicyId :: MonadFail m => HeadId -> m PolicyId

headSeedToTxIn :: MonadFail m => HeadSeed -> m TxIn

txInToHeadSeed :: TxIn -> HeadSeed

Helpers

findHeadAssetId :: TxOut ctx -> Maybe (PolicyId, AssetName)

findStateToken :: TxOut ctx -> Maybe HeadId

Find (if it exists) the head identifier contained in given TxOut.