Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Documentation

type UTxOWithScript = (TxIn, TxOut CtxUTxO, ScriptData)

headPolicyId :: TxIn -> PolicyId

mkHeadTokenScript :: TxIn -> PlutusScript

hydraHeadV1AssetName :: AssetName

headValue :: Value

Create Hydra Head transactions

initTx

Arguments

:: NetworkId

-> [VerificationKey PaymentKey]
Participant's cardano public keys.

-> HeadParameters

-> TxIn

-> Tx

Create the init transaction from some HeadParameters and a single TxIn which will be used as unique parameter for minting NFTs.

mkHeadOutput :: NetworkId -> PolicyId -> TxOutDatum ctx -> TxOut ctx

mkHeadOutputInitial :: NetworkId -> PolicyId -> HeadParameters -> TxOut CtxTx

mkInitialOutput :: NetworkId -> PolicyId -> VerificationKey PaymentKey -> TxOut CtxTx

commitTx

Arguments

:: NetworkId

-> Party

-> Maybe (TxIn, TxOut CtxUTxO)
A single UTxO to commit to the Head We currently limit committing one UTxO to the head because of size limitations.

-> (TxIn, TxOut CtxUTxO, Hash PaymentKey)
The initial output (sent to each party) which should contain the PT and is locked by initial script

-> Tx

Craft a commit transaction which includes the "committed" utxo as a datum.

TODO: Remove all the qualified Api once the refactoring is over and there's no more import clash with 'cardano-ledger'.

TODO: Get rid of Ledger types in the signature and fully rely on Cardano.Api

mkCommitDatum :: Party -> ValidatorHash -> Maybe (TxIn, TxOut CtxUTxO) -> Datum

collectComTx

Arguments

:: NetworkId

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> (TxIn, TxOut CtxUTxO, ScriptData, [Party])
Everything needed to spend the Head state-machine output.

-> Map TxIn (TxOut CtxUTxO, ScriptData)
Data needed to spend the commit output produced by each party. Should contain the PT and is locked by ν_commit script.

-> Tx

Create a transaction collecting all "committed" utxo and opening a Head, i.e. driving the Head script state.

closeTx

Arguments

:: VerificationKey PaymentKey
Party who's authorizing this transaction

-> Snapshot Tx

-> MultiSigned (Snapshot Tx)
Multi-signature of the whole snapshot

-> UTxOWithScript
Everything needed to spend the Head state-machine output. FIXME(SN): should also contain some Head identifier/address and stored Value (maybe the TxOut + Data?)

-> Tx

Create a transaction closing a head with given snapshot number and utxo.

fanoutTx

Arguments

:: UTxO
Snapshotted UTxO to fanout on layer 1

-> (TxIn, TxOut CtxUTxO, ScriptData)
Everything needed to spend the Head state-machine output.

-> PlutusScript
Minting Policy script, made from initial seed

-> Tx

data AbortTxError

Constructors

OverlappingInputs

abortTx

Arguments

:: VerificationKey PaymentKey
Party who's authorizing this transaction

-> (TxIn, TxOut CtxUTxO, ScriptData)
Everything needed to spend the Head state-machine output.

-> PlutusScript
Script for monetary policy to burn tokens

-> Map TxIn (TxOut CtxUTxO, ScriptData)
Data needed to spend the initial output sent to each party to the Head. Should contain the PT and is locked by initial script.

-> Map TxIn (TxOut CtxUTxO, ScriptData)
Data needed to spend commit outputs. Should contain the PT and is locked by commit script.

-> Either AbortTxError Tx

Create transaction which aborts a head by spending the Head output and all other "initial" outputs.

Observe Hydra Head transactions

data InitObservation

Constructors

InitObservation

Fields

threadOutput :: (TxIn, TxOut CtxUTxO, ScriptData, [Party])

The state machine UTxO produced by the Init transaction This output should always be present and threaded across all transactions. NOTE(SN): The Head's identifier is somewhat encoded in the TxOut's address XXX(SN): Data and [OnChain.Party] are overlapping

initials :: [UTxOWithScript]

commits :: [UTxOWithScript]

headId :: HeadId

headTokenScript :: PlutusScript

observeInitTx :: NetworkId -> [VerificationKey PaymentKey] -> Party -> Tx -> Maybe (OnChainTx Tx, InitObservation)

convertParty :: Party -> Party

type CommitObservation = UTxOWithScript

observeCommitTx

Arguments

:: NetworkId

-> [TxIn]
Known (remaining) initial tx inputs.

-> Tx

-> Maybe (OnChainTx Tx, CommitObservation)

Identify a commit tx by:

Find which initial tx input is being consumed.

Find the redeemer corresponding to that initial, which contains the tx input of the committed utxo.

Find the outputs which pays to the commit validator.

Using the datum of that output, deserialize the comitted output.

Reconstruct the committed UTxO from both values (tx input and output).

convertTxOut :: Maybe SerializedTxOut -> Maybe (TxOut CtxUTxO)

data CollectComObservation

Constructors

CollectComObservation

Fields

threadOutput :: (TxIn, TxOut CtxUTxO, ScriptData, [Party])

headId :: HeadId

observeCollectComTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe (OnChainTx Tx, CollectComObservation)

Identify a collectCom tx by lookup up the input spending the Head output and decoding its redeemer.

data CloseObservation

Constructors

CloseObservation

Fields

threadOutput :: (TxIn, TxOut CtxUTxO, ScriptData, [Party])

headId :: HeadId

observeCloseTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe (OnChainTx Tx, CloseObservation)

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

type FanoutObservation = ()

observeFanoutTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe (OnChainTx Tx, FanoutObservation)

Identify a fanout tx by lookup up the input spending the Head output and decoding its redeemer.

TODO: Ideally, the fanout does not produce any state-machine output. That means, to observe it, we need to look for a transaction with an input spent from a known script (the head state machine script) with a "fanout" redeemer.

type AbortObservation = ()

observeAbortTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe (OnChainTx Tx, AbortObservation)

Identify an abort tx by looking up the input spending the Head output and decoding its redeemer. FIXME(SN): make sure this is aborting "the right head / your head"

Functions related to OnChainHeadState

ownInitial :: PlutusScript -> VerificationKey PaymentKey -> [UTxOWithScript] -> Maybe (TxIn, TxOut CtxUTxO, Hash PaymentKey)

Look for the "initial" which corresponds to given cardano verification key.

mkHeadId :: PolicyId -> HeadId

Helpers

headTokensFromValue :: PlutusScript -> Value -> [(AssetName, Quantity)]

assetNameFromVerificationKey :: VerificationKey PaymentKey -> AssetName

findFirst :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Find first occurrence including a transformation.

findHeadAssetId :: TxOut ctx -> Maybe (PolicyId, AssetName)

findStateToken :: TxOut ctx -> Maybe HeadId

Find (if it exists) the head identifier contained in given TxOut.