Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Documentation

type UTxOWithScript = (TxIn, TxOut CtxUTxO, HashableScriptData)

Needed on-chain data to create Head transactions.

newtype UTxOHash

Constructors

UTxOHash ByteString

data InitialThreadOutput

Representation of the Head output after an Init transaction.

Constructors

InitialThreadOutput

Fields

initialThreadUTxO :: UTxOWithScript

initialContestationPeriod :: ContestationPeriod

initialParties :: [Party]

data OpenThreadOutput

Representation of the Head output after a CollectCom transaction.

Constructors

OpenThreadOutput

Fields

openThreadUTxO :: UTxOWithScript

openContestationPeriod :: ContestationPeriod

openParties :: [Party]

data ClosedThreadOutput

Constructors

ClosedThreadOutput

Fields

closedThreadUTxO :: UTxOWithScript

closedParties :: [Party]

closedContestationDeadline :: POSIXTime

closedContesters :: [PubKeyHash]

hydraHeadV1AssetName :: AssetName

headValue :: Value

Create Hydra Head transactions

initTx

Arguments

:: NetworkId

-> [VerificationKey PaymentKey]
All participants cardano keys.

-> HeadParameters

-> TxIn

-> Tx

Create the init transaction from some HeadParameters and a single TxIn which will be used as unique parameter for minting NFTs.

mkHeadOutput :: NetworkId -> PolicyId -> TxOutDatum ctx -> TxOut ctx

mkHeadOutputInitial :: NetworkId -> TxIn -> HeadParameters -> TxOut CtxTx

mkInitialOutput :: NetworkId -> TxIn -> VerificationKey PaymentKey -> TxOut CtxTx

commitTx

Arguments

:: NetworkId

-> ScriptRegistry
Published Hydra scripts to reference.

-> HeadId

-> Party

-> UTxO' (TxOut CtxUTxO, Witness WitCtxTxIn)
The UTxO to commit to the Head along with witnesses.

-> (TxIn, TxOut CtxUTxO, Hash PaymentKey)
The initial output (sent to each party) which should contain the PT and is locked by initial script

-> Tx

Craft a commit transaction which includes the "committed" utxo as a datum.

mkCommitDatum :: Party -> UTxO -> CurrencySymbol -> Datum

collectComTx

Arguments

:: NetworkId

-> ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> InitialThreadOutput
Everything needed to spend the Head state-machine output.

-> Map TxIn (TxOut CtxUTxO, HashableScriptData)
Data needed to spend the commit output produced by each party. Should contain the PT and is locked by ν_commit script.

-> HeadId
Head id

-> Tx

Create a transaction collecting all "committed" utxo and opening a Head, i.e. driving the Head script state.

data ClosingSnapshot

Low-level data type of a snapshot to close the head with. This is different to the ConfirmedSnasphot, which is provided to CloseTx as it also contains relevant chain state like the $sel:openUtxoHash:CloseWithInitialSnapshot.

Constructors

CloseWithInitialSnapshot

Fields

openUtxoHash :: UTxOHash

CloseWithConfirmedSnapshot

Fields

snapshotNumber :: SnapshotNumber

closeUtxoHash :: UTxOHash

signatures :: MultiSignature (Snapshot Tx)

closeTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> ClosingSnapshot
The snapshot to close with, can be either initial or confirmed one.

-> SlotNo
Lower validity slot number, usually a current or quite recent slot number.

-> PointInTime
Upper validity slot and UTC time to compute the contestation deadline time.

-> OpenThreadOutput
Everything needed to spend the Head state-machine output.

-> HeadId
Head identifier

-> Tx

Create a transaction closing a head with either the initial snapshot or with a multi-signed confirmed snapshot.

contestTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> Snapshot Tx
Contested snapshot number (i.e. the one we contest to)

-> MultiSignature (Snapshot Tx)
Multi-signature of the whole snapshot

-> PointInTime
Current slot and posix time to be used as the contestation time.

-> ClosedThreadOutput
Everything needed to spend the Head state-machine output.

-> HeadId

-> ContestationPeriod

-> Tx

fanoutTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> UTxO
Snapshotted UTxO to fanout on layer 1

-> UTxOWithScript
Everything needed to spend the Head state-machine output.

-> SlotNo
Contestation deadline as SlotNo, used to set lower tx validity bound.

-> PlutusScript
Minting Policy script, made from initial seed

-> Tx

Create the fanout transaction, which distributes the closed state accordingly. The head validator allows fanout only > deadline, so we need to set the lower bound to be deadline + 1 slot.

data AbortTxError

Constructors

OverlappingInputs

abortTx

Arguments

:: UTxO
Committed UTxOs to reimburse.

-> ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> (TxIn, TxOut CtxUTxO, HashableScriptData)
Everything needed to spend the Head state-machine output.

-> PlutusScript
Script for monetary policy to burn tokens

-> Map TxIn (TxOut CtxUTxO, HashableScriptData)
Data needed to spend the initial output sent to each party to the Head. Should contain the PT and is locked by initial script.

-> Map TxIn (TxOut CtxUTxO, HashableScriptData)
Data needed to spend commit outputs. Should contain the PT and is locked by commit script.

-> Either AbortTxError Tx

Create transaction which aborts a head by spending the Head output and all other "initial" outputs.

Observe Hydra Head transactions

data InitObservation

Constructors

InitObservation

Fields

threadOutput :: InitialThreadOutput

The state machine UTxO produced by the Init transaction This output should always be present and threaded across all transactions. NOTE(SN): The Head's identifier is somewhat encoded in the TxOut's address XXX(SN): Data and [OnChain.Party] are overlapping

initials :: [UTxOWithScript]

commits :: [UTxOWithScript]

headId :: HeadId

seedTxIn :: TxIn

contestationPeriod :: ContestationPeriod

parties :: [Party]

data NotAnInitReason

Constructors

NotAHeadPolicy

NoHeadOutput

NotAHeadDatum

NoSTFound

PartiesMismatch

OwnPartyMissing

CPMismatch

PTsNotMintedCorrectly

observeInitTx

Arguments

:: NetworkId

-> [VerificationKey PaymentKey]

-> ContestationPeriod
Our node's contestation period

-> Party

-> [Party]

-> Tx

-> Either NotAnInitReason InitObservation

data CommitObservation

Constructors

CommitObservation

Fields

commitOutput :: UTxOWithScript

party :: Party

committed :: UTxO

observeCommitTx

Arguments

:: NetworkId

-> [TxIn]
Known (remaining) initial tx inputs.

-> Tx

-> Maybe CommitObservation

Identify a commit tx by:

Find which initial tx input is being consumed,

Find the redeemer corresponding to that initial, which contains the tx input of the committed utxo,

Find the outputs which pays to the commit validator,

Using the datum of that output, deserialize the committed output,

Reconstruct the committed UTxO from both values (tx input and output).

data CollectComObservation

Constructors

CollectComObservation

Fields

threadOutput :: OpenThreadOutput

headId :: HeadId

utxoHash :: UTxOHash

observeCollectComTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CollectComObservation

Identify a collectCom tx by lookup up the input spending the Head output and decoding its redeemer.

data CloseObservation

Constructors

CloseObservation

Fields

threadOutput :: ClosedThreadOutput

headId :: HeadId

snapshotNumber :: SnapshotNumber

observeCloseTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CloseObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data ContestObservation

Constructors

ContestObservation

Fields

contestedThreadOutput :: (TxIn, TxOut CtxUTxO, HashableScriptData)

headId :: HeadId

snapshotNumber :: SnapshotNumber

contesters :: [PubKeyHash]

observeContestTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe ContestObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data FanoutObservation

Constructors

FanoutObservation

observeFanoutTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe FanoutObservation

Identify a fanout tx by lookup up the input spending the Head output and decoding its redeemer.

data AbortObservation

Constructors

AbortObservation

observeAbortTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe AbortObservation

Identify an abort tx by looking up the input spending the Head output and decoding its redeemer. FIXME: Add headId to AbortObservation to allow "upper layers" to determine we are seeing an abort of "our head"

Helpers

mkHeadId :: PolicyId -> HeadId

headIdToCurrencySymbol :: HeadId -> CurrencySymbol

headTokensFromValue :: PlutusScript -> Value -> [(AssetName, Quantity)]

assetNameFromVerificationKey :: VerificationKey PaymentKey -> AssetName

findFirst :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Find first occurrence including a transformation.

findHeadAssetId :: TxOut ctx -> Maybe (PolicyId, AssetName)

findStateToken :: TxOut ctx -> Maybe HeadId

Find (if it exists) the head identifier contained in given TxOut.