Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Hydra Head transactions

constructTx :: TxIn -> PostChainTx tx -> TxBodyContent BuildTx AlonzoEra

initTx :: HeadParameters -> TxIn -> TxBodyContent BuildTx AlonzoEra

Smart constructors for TxBodyContent which should be used with makeTransactionBodyAutoBalance to balance and account for fees accordingly.

Create the init transaction from some HeadParameters and a single UTXO which will be used for minting NFTs.