Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Documentation

network :: Network

Post Hydra Head transactions

constructTx :: TxIn StandardCrypto -> PostChainTx tx -> ValidatedTx (AlonzoEra StandardCrypto)

Construct the Head protocol transactions as Alonzo Tx. Note that ValidatedTx this produces an unbalanced, unsigned transaction and this type was used (in contrast to TxBody) to be able to express included datums, onto which at least the initTx relies on.

initTx :: HeadParameters -> TxIn StandardCrypto -> ValidatedTx (AlonzoEra StandardCrypto)

Create the init transaction from some HeadParameters and a single TxIn which will be used as unique parameter for minting NFTs.

Observe Hydra Head transactions

observeTx :: ValidatedTx (AlonzoEra StandardCrypto) -> Maybe (OnChainTx tx)

observeInitTx :: ValidatedTx (AlonzoEra StandardCrypto) -> Maybe (OnChainTx tx)

observeCommitTx :: ValidatedTx (AlonzoEra StandardCrypto) -> Maybe (OnChainTx tx)

Helpers

mkUnsignedTx :: TxBody (AlonzoEra StandardCrypto) -> TxDats (AlonzoEra StandardCrypto) -> ValidatedTx (AlonzoEra StandardCrypto)