Hydra.Chain.Direct.Tx

Description

Smart constructors for creating Hydra protocol transactions to be used in the Direct way of talking to the main-chain.

This module also encapsulates the transaction format used when talking to the cardano-node, which is currently different from the Cardano, thus we have not yet "reached" isomorphism.

Documentation

type UTxOWithScript = (TxIn, TxOut CtxUTxO, ScriptData)

Needed on-chain data to create Head transactions.

data InitialThreadOutput

Representation of the Head output after an Init transaction.

Constructors

InitialThreadOutput

Fields

initialThreadUTxO :: UTxOWithScript

initialContestationPeriod :: ContestationPeriod

initialParties :: [Party]

data OpenThreadOutput

Representation of the Head output after a CollectCom transaction.

Constructors

OpenThreadOutput

Fields

openThreadUTxO :: UTxOWithScript

openContestationPeriod :: ContestationPeriod

openParties :: [Party]

data ClosedThreadOutput

Constructors

ClosedThreadOutput

Fields

closedThreadUTxO :: UTxOWithScript

closedParties :: [Party]

closedContestationDeadline :: POSIXTime

headPolicyId :: TxIn -> PolicyId

mkHeadTokenScript :: TxIn -> PlutusScript

hydraHeadV1AssetName :: AssetName

headValue :: Value

Create Hydra Head transactions

initTx

Arguments

:: NetworkId

-> [VerificationKey PaymentKey]
Participant's cardano public keys.

-> HeadParameters

-> TxIn

-> Tx

Create the init transaction from some HeadParameters and a single TxIn which will be used as unique parameter for minting NFTs.

mkHeadOutput :: NetworkId -> PolicyId -> TxOutDatum ctx -> TxOut ctx

mkHeadOutputInitial :: NetworkId -> PolicyId -> HeadParameters -> TxOut CtxTx

mkInitialOutput :: NetworkId -> PolicyId -> VerificationKey PaymentKey -> TxOut CtxTx

commitTx

Arguments

:: NetworkId

-> Party

-> Maybe (TxIn, TxOut CtxUTxO)
A single UTxO to commit to the Head We currently limit committing one UTxO to the head because of size limitations.

-> (TxIn, TxOut CtxUTxO, Hash PaymentKey)
The initial output (sent to each party) which should contain the PT and is locked by initial script

-> Tx

Craft a commit transaction which includes the "committed" utxo as a datum.

mkCommitDatum :: Party -> ValidatorHash -> Maybe (TxIn, TxOut CtxUTxO) -> Datum

collectComTx

Arguments

:: NetworkId

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> InitialThreadOutput
Everything needed to spend the Head state-machine output.

-> Map TxIn (TxOut CtxUTxO, ScriptData)
Data needed to spend the commit output produced by each party. Should contain the PT and is locked by ν_commit script.

-> Tx

Create a transaction collecting all "committed" utxo and opening a Head, i.e. driving the Head script state.

data ClosingSnapshot

Low-level data type of a snapshot to close the head with. This is different to the ConfirmedSnasphot, which is provided to CloseTx as it also contains relevant chain state like the $sel:openUtxoHash:CloseWithInitialSnapshot.

Constructors

CloseWithInitialSnapshot

Fields

openUtxoHash :: ByteString

CloseWithConfirmedSnapshot

Fields

snapshotNumber :: SnapshotNumber

closeUtxoHash :: ByteString

signatures :: MultiSignature (Snapshot Tx)

type PointInTime = (SlotNo, POSIXTime)

closeTx

Arguments

:: VerificationKey PaymentKey
Party who's authorizing this transaction

-> ClosingSnapshot
The snapshot to close with, can be either initial or confirmed one.

-> PointInTime
Current slot and posix time to be recorded as the closing time.

-> OpenThreadOutput
Everything needed to spend the Head state-machine output.

-> Tx

Create a transaction closing a head with either the initial snapshot or with a multi-signed confirmed snapshot.

contestTx

Arguments

:: VerificationKey PaymentKey
Party who's authorizing this transaction

-> Snapshot Tx
Contested snapshot number (i.e. the one we contest to)

-> MultiSignature (Snapshot Tx)
Multi-signature of the whole snapshot

-> PointInTime
Current slot and posix time to be recorded as the closing time.

-> ClosedThreadOutput
Everything needed to spend the Head state-machine output.

-> Tx

fanoutTx

Arguments

:: UTxO
Snapshotted UTxO to fanout on layer 1

-> UTxOWithScript
Everything needed to spend the Head state-machine output.

-> PointInTime
Point in time at which this transaction is posted, used to set lower bound.

-> PlutusScript
Minting Policy script, made from initial seed

-> Tx

data AbortTxError

Constructors

OverlappingInputs

abortTx

Arguments

:: VerificationKey PaymentKey
Party who's authorizing this transaction

-> (TxIn, TxOut CtxUTxO, ScriptData)
Everything needed to spend the Head state-machine output.

-> PlutusScript
Script for monetary policy to burn tokens

-> Map TxIn (TxOut CtxUTxO, ScriptData)
Data needed to spend the initial output sent to each party to the Head. Should contain the PT and is locked by initial script.

-> Map TxIn (TxOut CtxUTxO, ScriptData)
Data needed to spend commit outputs. Should contain the PT and is locked by commit script.

-> Either AbortTxError Tx

Create transaction which aborts a head by spending the Head output and all other "initial" outputs.

Observe Hydra Head transactions

data InitObservation

Constructors

InitObservation

Fields

threadOutput :: InitialThreadOutput

The state machine UTxO produced by the Init transaction This output should always be present and threaded across all transactions. NOTE(SN): The Head's identifier is somewhat encoded in the TxOut's address XXX(SN): Data and [OnChain.Party] are overlapping

initials :: [UTxOWithScript]

commits :: [UTxOWithScript]

headId :: HeadId

headTokenScript :: PlutusScript

contestationPeriod :: ContestationPeriod

parties :: [Party]

observeInitTx :: NetworkId -> [VerificationKey PaymentKey] -> Party -> Tx -> Maybe InitObservation

data CommitObservation

Constructors

CommitObservation

Fields

commitOutput :: UTxOWithScript

party :: Party

committed :: UTxO

observeCommitTx

Arguments

:: NetworkId

-> [TxIn]
Known (remaining) initial tx inputs.

-> Tx

-> Maybe CommitObservation

Identify a commit tx by:

Find which initial tx input is being consumed,

Find the redeemer corresponding to that initial, which contains the tx input of the committed utxo,

Find the outputs which pays to the commit validator,

Using the datum of that output, deserialize the committed output,

Reconstruct the committed UTxO from both values (tx input and output).

convertTxOut :: Maybe Commit -> Maybe (TxOut CtxUTxO)

data CollectComObservation

Constructors

CollectComObservation

Fields

threadOutput :: OpenThreadOutput

headId :: HeadId

utxoHash :: ByteString

observeCollectComTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CollectComObservation

Identify a collectCom tx by lookup up the input spending the Head output and decoding its redeemer.

data CloseObservation

Constructors

CloseObservation

Fields

threadOutput :: ClosedThreadOutput

headId :: HeadId

snapshotNumber :: SnapshotNumber

observeCloseTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CloseObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data ContestObservation

Constructors

ContestObservation

Fields

contestedThreadOutput :: (TxIn, TxOut CtxUTxO, ScriptData)

headId :: HeadId

snapshotNumber :: SnapshotNumber

observeContestTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe ContestObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data FanoutObservation

Constructors

FanoutObservation

observeFanoutTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe FanoutObservation

Identify a fanout tx by lookup up the input spending the Head output and decoding its redeemer.

data AbortObservation

Constructors

AbortObservation

observeAbortTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe AbortObservation

Identify an abort tx by looking up the input spending the Head output and decoding its redeemer. FIXME: Add headId to AbortObservation to allow "upper layers" to determine we are seeing an abort of "our head"

Functions related to OnChainHeadState

ownInitial :: PlutusScript -> VerificationKey PaymentKey -> [UTxOWithScript] -> Maybe (TxIn, TxOut CtxUTxO, Hash PaymentKey)

Look for the "initial" which corresponds to given cardano verification key.

mkHeadId :: PolicyId -> HeadId

Helpers

headTokensFromValue :: PlutusScript -> Value -> [(AssetName, Quantity)]

assetNameFromVerificationKey :: VerificationKey PaymentKey -> AssetName

findFirst :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Find first occurrence including a transformation.

findHeadAssetId :: TxOut ctx -> Maybe (PolicyId, AssetName)

findStateToken :: TxOut ctx -> Maybe HeadId

Find (if it exists) the head identifier contained in given TxOut.