Hydra.Chain.Direct.Util

Documentation

type Block = CardanoBlock StandardCrypto

type VerificationKey = VerKeyDSIGN (DSIGN StandardCrypto)

type SigningKey = SignKeyDSIGN (DSIGN StandardCrypto)

nullConnectTracers :: NetworkConnectTracers LocalAddress NodeToClientVersion

nullServerTracers :: NetworkServerTracers LocalAddress NodeToClientVersion

nodeToClientVLatest :: NodeToClientVersion

versions :: NetworkId -> (NodeToClientVersion -> app) -> Versions NodeToClientVersion NodeToClientVersionData app

defaultCodecs :: MonadST m => NodeToClientVersion -> ClientCodecs Block m

readKeyPair :: FilePath -> IO (VerificationKey PaymentKey, SigningKey PaymentKey)

readFileTextEnvelopeThrow :: HasTextEnvelope a => AsType a -> FilePath -> IO a

readVerificationKey :: FilePath -> IO (VerificationKey PaymentKey)

retry :: forall e m a. (MonadCatch m, MonadDelay m, Exception e) => (e -> Bool) -> m a -> m a

A simple retrying function with a constant delay. Retries only if the given predicate evaluates to True.

Better coupled with a timeout function.

signWith :: SigningKey PaymentKey -> ValidatedTx LedgerEra -> ValidatedTx LedgerEra

markerDatum :: Data

Marker datum used to identify payment UTXO

markerDatumHash :: Hash ScriptData

Hash of the markerDatum

isMarkedOutput :: TxOut CtxUTxO -> Bool

Determine whether a TxOut is marked to be used for paying Hydra Head transactions