Hydra.Chain.Direct.Wallet

Description

Companion tiny-wallet for the direct chain component. This module provide some useful utilities to tracking the wallet's UTXO, and accessing it

Documentation

type Address = Addr StandardCrypto

type TxBody = TxBody Era

type VerificationKey = VerKeyDSIGN (DSIGN StandardCrypto)

type SigningKey = SignKeyDSIGN (DSIGN StandardCrypto)

type TxIn = TxIn StandardCrypto

type TxOut = TxOut Era

type VkWitness = WitVKey 'Witness StandardCrypto

type QueryResult result = Either (MismatchEraInfo (CardanoEras StandardCrypto)) result

type UtxoSet = UTxO Era

type AlonzoPoint = Point (ShelleyBlock Era)

data TinyWallet m

A TinyWallet is a small abstraction of a wallet with basic UTXO management. The wallet is assumed to have only one address, and only one UTXO at that address.

It can sign transactions and keeps track of its UTXO behind the scene.

Constructors

TinyWallet

Fields

getUtxo :: STM m (Map TxIn TxOut)

getAddress :: Address

sign :: TxBody -> VkWitness

coverFee :: TxBody -> STM m (Either ErrCoverFee TxBody)

withTinyWallet

Arguments

:: NetworkMagic
Network identifier to which we expect to connect.

-> (VerificationKey, SigningKey)
Credentials of the wallet.

-> IOManager
A cross-platform abstraction for managing I/O operations on local sockets

-> FilePath
Path to a domain socket used to connect to the server.

-> (TinyWallet IO -> IO ())

-> IO ()

applyBlock :: Block -> (Address -> Bool) -> Map TxIn TxOut -> Map TxIn TxOut

Apply a block to our wallet. Does nothing if the transaction does not modify the UTXO set, or else, remove consumed utxos and add produced ones.

To determine whether a produced output is ours, we compare it to our unique address.

data ErrCoverFee

Constructors

ErrNoAvailableUtxo

ErrNotEnoughFunds

Fields

missingDelta :: Coin

coverFee_ :: Map TxIn TxOut -> TxBody -> Either ErrCoverFee TxBody

Cover fee for a transaction body using the given UTXO set. This calculate necessary fees and augments inputs outputs collateral accordingly to cover for the transaction cost and get the change back.

TODO: The fee calculation is currently very dumb and static.

client :: (MonadST m, MonadTimer m) => TVar m (Point Block) -> TMVar m (Map TxIn TxOut) -> Address -> NodeToClientVersion -> OuroborosApplication 'InitiatorMode LocalAddress LByteString m () Void

The idea for this wallet client is rather simple:

We bootstrap the client using the Local State Query (abbrev. LSQ) protocol, and in particular the GetUTxO query to get the most recent UTxO at the tip.

This request is however quite costly (O(n) in the size of the ledger's UTXO). So we don't want to poll it too often and burn needless CPU resources. Plus, whatever poll value we choose, we'll always end up report outdated information. Thus,

Once the most recent tip and UTXO acquired, we give the baton from the LSQ to the chain-sync protocol, which can continue watching the chain from that point on such that, if the UTXO is ever spent, it'll replace it with its new UTXO (reminder: the wallet has only one address).

Then, there's a bit of error-handling because, it's possible that between the moment we acquire the tip with the LSQ and pass the relay to the chain-sync, the chain has rolled back and the tip is no longer available. This can be gracefully done via waiting on MVars. See below.

The MVar also provides a convenient interface for the `TinyWallet m` abstraction which is used by consumers downstreams.

chainSyncClient :: forall m. MonadSTM m => TVar m (Point Block) -> TMVar m (Map TxIn TxOut) -> Address -> ChainSyncClient Block (Point Block) (Tip Block) m ()

stateQueryClient :: forall m. (MonadSTM m, MonadTimer m) => TVar m (Point Block) -> TMVar m (Map TxIn TxOut) -> Address -> LocalStateQueryClient Block (Point Block) (Query Block) m ()