Hydra.Chain.Direct.Wallet

Description

Companion tiny-wallet for the direct chain component. This module provide some useful utilities to tracking the wallet's UTXO, and accessing it

Documentation

type Address = Addr StandardCrypto

type TxIn = TxIn StandardCrypto

type TxOut = TxOut LedgerEra

data TinyWallet m

A TinyWallet is a small abstraction of a wallet with basic UTXO management. The wallet is assumed to have only one address, and only one UTXO at that address. It can sign transactions and keeps track of its UTXO behind the scene.

This wallet is not connecting to the node initially and when asked to reset. Otherwise it can be fed blocks via update as the chain rolls forward.

Constructors

TinyWallet

Fields

getUTxO :: STM m (Map TxIn TxOut)

Return all known UTxO addressed to this wallet.

sign :: ValidatedTx LedgerEra -> ValidatedTx LedgerEra

coverFee :: Map TxIn TxOut -> ValidatedTx LedgerEra -> STM m (Either ErrCoverFee (ValidatedTx LedgerEra))

reset :: QueryPoint -> m ()

Reset the wallet state to some point.

update :: Block -> m ()

Update the wallet state given some Block.

type ChainQuery m = QueryPoint -> Address ShelleyAddr -> m (Map TxIn TxOut, PParams LedgerEra, SystemStart, EpochInfo (Either Text))

getFuelUTxO :: MonadSTM m => TinyWallet m -> STM m (Maybe (TxIn, TxOut))

Get a single, marked as "fuel" UTxO.

watchUTxOUntil :: (Map TxIn TxOut -> Bool) -> TinyWallet IO -> IO (Map TxIn TxOut)

newTinyWallet

Arguments

:: Tracer IO TinyWalletLog
A tracer for logging

-> NetworkId
Network identifier to generate our address.

-> (VerificationKey PaymentKey, SigningKey PaymentKey)
Credentials of the wallet.

-> ChainQuery IO
A function to query UTxO, pparams, system start and epoch info from the node. Initially and on demand later.

-> IO (TinyWallet IO)

Create a new tiny wallet handle.

applyBlock :: Block -> (Address -> Bool) -> Map TxIn TxOut -> Map TxIn TxOut

Apply a block to our wallet. Does nothing if the transaction does not modify the UTXO set, or else, remove consumed utxos and add produced ones.

To determine whether a produced output is ours, we apply the given function checking the output's address.

getTxId :: (HashAlgorithm (HASH crypto), HashAnnotated (TxBody (era crypto)) EraIndependentTxBody crypto) => ValidatedTx (era crypto) -> TxId crypto

data ErrCoverFee

Constructors

ErrNoAvailableUTxO

ErrNotEnoughFunds ChangeError

ErrUnknownInput

Fields

input :: TxIn

ErrNoPaymentUTxOFound

ErrScriptExecutionFailed (RdmrPtr, TransactionScriptFailure StandardCrypto)

data ChangeError

Constructors

ChangeError

Fields

inputBalance :: Coin

outputBalance :: Coin

coverFee_ :: PParams LedgerEra -> SystemStart -> EpochInfo (Either Text) -> Map TxIn TxOut -> Map TxIn TxOut -> ValidatedTx LedgerEra -> Either ErrCoverFee (Map TxIn TxOut, ValidatedTx LedgerEra)

Cover fee for a transaction body using the given UTXO set. This calculate necessary fees and augments inputs outputs collateral accordingly to cover for the transaction cost and get the change back.

TODO: The fee calculation is currently very dumb and static.

findFuelUTxO :: Map TxIn TxOut -> Maybe (TxIn, TxOut)

estimateScriptsCost

Arguments

:: PParams LedgerEra
Protocol parameters

-> SystemStart
Start of the blockchain, for converting slots to UTC times

-> EpochInfo (Either Text)
Information about epoch sizes, for converting slots to UTC times

-> Map TxIn TxOut
A UTXO needed to resolve inputs

-> ValidatedTx LedgerEra
The pre-constructed transaction

-> Either (RdmrPtr, TransactionScriptFailure StandardCrypto) (Map RdmrPtr ExUnits)

Estimate cost of script executions on the transaction. This is only an estimates because the transaction isn't sealed at this point and adding new elements to it like change outputs or script integrity hash may increase that cost a little.

newtype BadTranslationException

Constructors

BadTranslationException (TranslationError StandardCrypto)

data TinyWalletLog

Constructors

InitializingWallet QueryPoint (Map TxIn TxOut)

ApplyBlock (Map TxIn TxOut)

LedgerEraMismatchError

Fields

expected :: Text

actual :: Text