Hydra.Crypto

Description

Hydra multi-signature credentials and cryptographic primitives used to sign and verify snapshots (or any messages) within the Hydra protocol.

Currently this interface is only supporting naiive, concatenated multi-signatures and will change when we adopt aggregated multi-signatures including aggregate keys.

It is recommended to import this module qualified to avoid confusion with Cardano keys & signatures.

Documentation

type SignAlg = Ed25519DSIGN

The used signature algorithm

Hydra keys

newtype SigningKey

Hydra signing key which can be used to sign messages and aggregate multi-signatures or deriveVerificationKey.

REVIEW: Maybe rewrite Show instance to not expose secret, eg. 8 bytes from the hash of the key? Although both, cardano-api and cardano-crypto-class are both deriving this and thus showing secret key material as well.

Constructors

HydraSigningKey (SignKeyDSIGN SignAlg)

serialiseSigningKeyToRawBytes :: SigningKey -> ByteString

Serialise the signing key material as raw bytes.

deserialiseSigningKeyFromRawBytes :: MonadFail m => ByteString -> m SigningKey

Deserialise a signing key from raw bytes.

deriveVerificationKey :: SigningKey -> VerificationKey

Get the VerificationKey for a given SigningKey.

generateSigningKey :: ByteString -> SigningKey

Create a new SigningKey from a ByteString seed. The created keys are not random and insecure, so don't use this in production code!

newtype VerificationKey

Hydra verification key, which can be used to verify signed messages.

Constructors

HydraVerificationKey (VerKeyDSIGN SignAlg)

serialiseVerificationKeyToRawBytes :: VerificationKey -> ByteString

Serialise the verification key material as raw bytes.

deserialiseVerificationKeyFromRawBytes :: MonadFail m => ByteString -> m VerificationKey

Deserialise a verirfication key from raw bytes.

hashVerificationKey :: VerificationKey -> Hash Blake2b_256 VerificationKey

Get the Blake2b hash of a VerificationKey.

Signatures

newtype Signature a

Signature of a, not containing the actual payload.

Constructors

HydraSignature (SigDSIGN SignAlg)

sign :: SignableRepresentation a => SigningKey -> a -> Signature a

Sign some value a with the provided SigningKey.

verify :: SignableRepresentation a => VerificationKey -> Signature a -> a -> Bool

Verify a given 'Signature a' and value a using provided VerificationKey.

Multi-signatures

newtype MultiSignature a

Naiively aggregated multi-signatures.

Constructors

HydraMultiSignature

Fields

multiSignature :: [Signature a]

aggregate :: [Signature a] -> MultiSignature a

Combine multiple signatures of a into a 'MultiSignature a'.

aggregateInOrder :: Ord k => Map k (Signature a) -> [k] -> MultiSignature a

toPlutusSignatures :: MultiSignature a -> [Signature]