Hydra.Events.FileBased

Description

A file-based event source and sink using JSON encoding.

This is currently used as the main event source and sink in the hydra-node.

Documentation

mkFileBasedEventStore :: (ToJSON e, FromJSON e, HasEventId e) => FilePath -> PersistenceIncremental e IO -> IO (EventStore e IO)

A basic file based event source and sink defined using a rotated PersistenceIncremental handle.