Hydra.HeadLogic

Description

Implements the Head Protocol's state machine as pure functions in an event sourced manner.

More specifically, the update will handle Inputs (or rather "commands" in event sourcing speak) and convert that into a list of side-Effects and StateChanged events, which in turn are aggregated into a single HeadState.

As the specification is using a more imperative way of specifying the protocol behavior, one would find the decision logic in update while state updates can be found in the corresponding aggregate branch.

Documentation

update

Arguments

:: IsChainState tx

=> Environment

-> Ledger tx

-> NodeState tx
Current NodeState to validate the command against.

-> Input tx
Input to be processed.

-> Outcome tx

Handles inputs and converts them into StateChanged events along with Effects, in case it is processed successfully. Later, the Node will aggregate the events, resulting in a new HeadState.

aggregateNodeState :: IsChainState tx => NodeState tx -> StateChanged tx -> NodeState tx

Reflect StateChanged events onto the NodeState aggregateNodeState.

onConnectionEvent :: Text -> Connectivity -> Outcome tx

aggregateChainStateHistory :: IsChainState tx => ChainStateHistory tx -> StateChanged tx -> ChainStateHistory tx

aggregateState :: IsChainState tx => NodeState tx -> Outcome tx -> NodeState tx

aggregate :: IsChainState tx => HeadState tx -> StateChanged tx -> HeadState tx

Reflect StateChanged events onto the HeadState aggregate.

isLeader :: HeadParameters -> Party -> SnapshotNumber -> Bool

onIdleClientInit :: Environment -> Outcome tx

Client request to init the head. This leads to an init transaction on chain, containing the head parameters.

Transition: IdleState → IdleState

onIdleChainInitTx

Arguments

:: Environment

-> ChainStateType tx
New chain state.

-> HeadId

-> HeadSeed

-> HeadParameters

-> [OnChainId]

-> Outcome tx

Observe an init transaction, initialize parameters in an InitialState and notify clients that they can now commit.

Transition: IdleState → InitialState

onInitialChainCommitTx

Arguments

:: Monoid (UTxOType tx)

=> InitialState tx

-> ChainStateType tx
New chain state

-> Party
Committing party

-> UTxOType tx
Committed UTxO

-> Outcome tx

Observe a commit transaction and record the committed UTxO in the state. Also, if this is the last commit to be observed, post a collect-com transaction on-chain.

Transition: InitialState → InitialState

onInitialClientAbort :: Monoid (UTxOType tx) => InitialState tx -> Outcome tx

Client request to abort the head. This leads to an abort transaction on chain, reimbursing already committed UTxOs.

Transition: InitialState → InitialState

onInitialChainAbortTx

Arguments

:: Monoid (UTxOType tx)

=> ChainStateType tx
New chain state

-> Committed tx

-> HeadId

-> Outcome tx

Observe an abort transaction by switching the state and notifying clients about it.

Transition: InitialState → IdleState

onInitialChainCollectTx

Arguments

:: IsChainState tx

=> InitialState tx

-> ChainStateType tx
New chain state

-> Outcome tx

Observe a collectCom transaction. We initialize the OpenState using the head parameters from IdleState and construct an InitialSnapshot holding u0 from the committed UTxOs.

Transition: InitialState → OpenState

onOpenClientNewTx

Arguments

:: tx
The transaction to be submitted to the head.

-> Outcome tx

Client request to ingest a new transaction into the head.

Transition: OpenState → OpenState

onOpenNetworkReqTx

Arguments

:: IsTx tx

=> Environment

-> Ledger tx

-> ChainSlot

-> OpenState tx

-> TTL

-> tx
The transaction to be submitted to the head.

-> Outcome tx

Process a transaction request (ReqTx) from a party.

We apply this transaction to the seen utxo (ledger state). If not applicable, we wait and retry later. If it applies, this yields an updated seen ledger state. Then, we check whether we are the leader for the next snapshot and emit a snapshot request ReqSn including this transaction if needed.

Transition: OpenState → OpenState

onOpenNetworkReqSn

Arguments

:: IsTx tx

=> Environment

-> Ledger tx

-> PendingDeposits tx

-> ChainSlot

-> OpenState tx

-> Party
Party which sent the ReqSn.

-> SnapshotVersion
Requested snapshot version.

-> SnapshotNumber
Requested snapshot number.

-> [TxIdType tx]
List of transactions to snapshot.

-> Maybe tx
Optional decommit transaction of removing funds from the head.

-> Maybe (TxIdType tx)
Optional commit of additional funds into the head.

-> Outcome tx

Process a snapshot request (ReqSn) from party.

This checks that s is the next snapshot number and that the party is responsible for leading that snapshot. Then, we potentially wait until the previous snapshot is confirmed (no snapshot is in flight), before we apply (or wait until applicable) the requested transactions to the last confirmed snapshot. Only then, we start tracking this new "seen" snapshot, compute a signature of it and send the corresponding AckSn to all parties. Finally, the pending transaction set gets pruned to only contain still applicable transactions.

Transition: OpenState → OpenState

onOpenNetworkAckSn

Arguments

:: IsTx tx

=> Environment

-> PendingDeposits tx

-> OpenState tx

-> Party
Party which sent the AckSn.

-> Signature (Snapshot tx)
Signature from other party.

-> SnapshotNumber
Snapshot number of this AckSn.

-> Outcome tx

Process a snapshot acknowledgement (AckSn) from a party.

We do require that the is from the last seen or next expected snapshot, and potentially wait wait for the corresponding ReqSn before proceeding. If the party hasn't sent us a signature yet, we store it. Once a signature from each party has been collected, we aggregate a multi-signature and verify it is correct. If everything is fine, the snapshot can be considered as the latest confirmed one. Similar to processing a ReqTx, we check whether we are leading the next snapshot and craft a corresponding ReqSn if needed.

Transition: OpenState → OpenState

onClientRecover :: IsTx tx => ChainSlot -> PendingDeposits tx -> TxIdType tx -> Outcome tx

Client request to recover deposited UTxO.

Transition: OpenState → OpenState

onOpenClientDecommit

Arguments

:: IsTx tx

=> HeadId

-> Ledger tx

-> ChainSlot

-> CoordinatedHeadState tx

-> tx
Decommit transaction.

-> Outcome tx

Client request to decommit UTxO from the head.

Only possible if there is no decommit _in flight_ and if the tx applies cleanly to the local ledger state.

Transition: OpenState → OpenState

onOpenNetworkReqDec :: IsTx tx => Environment -> Ledger tx -> TTL -> ChainSlot -> OpenState tx -> tx -> Outcome tx

Process the request ReqDec to decommit something from the Open head.

Transition: OpenState → OpenState

When node receives ReqDec network message it should: - Check there is no decommit in flight: - Alter it's state to record what is to be decommitted - Issue a server output DecommitRequested with the relevant utxo - Issue a ReqSn since all parties need to agree in order for decommit to be taken out of a Head. - Check if we are the leader

onOpenChainTick :: IsTx tx => Environment -> PendingDeposits tx -> OpenState tx -> UTCTime -> Outcome tx

Process the chain (and time) advancing in an open head.

Transition: OpenState → OpenState

This is primarily used to track deposits and either drop them or request snapshots for inclusion.

onOpenChainIncrementTx

Arguments

:: OpenState tx

-> ChainStateType tx

-> SnapshotVersion
New open state version

-> TxIdType tx
Deposit TxId

-> Outcome tx

Observe a increment transaction. If the outputs match the ones of the pending commit UTxO, then we consider the deposit/increment finalized, and remove the increment UTxO from $sel:pendingDeposits:NodeState from the local state.

Transition: OpenState → OpenState

onOpenChainDecrementTx

Arguments

:: OpenState tx

-> ChainStateType tx

-> SnapshotVersion
New open state version

-> UTxOType tx
Outputs removed by the decrement

-> Outcome tx

Observe a decrement transaction. If the outputs match the ones of the pending decommit tx, then we consider the decommit finalized, and remove the decommit tx in flight.

Finally, if the client observing happens to be the leader, then a new ReqSn is broadcasted.

Transition: OpenState → OpenState

onOpenClientClose :: OpenState tx -> Outcome tx

Client request to close the head. This leads to a close transaction on chain using the latest confirmed snaphshot of the OpenState.

Transition: OpenState → OpenState

onOpenChainCloseTx

Arguments

:: OpenState tx

-> ChainStateType tx
New chain state.

-> SnapshotNumber
Closed snapshot number.

-> UTCTime
Contestation deadline.

-> Outcome tx

Observe a close transaction. If the closed snapshot number is smaller than our last confirmed, we post a contest transaction. Also, we do schedule a notification for clients to fanout at the deadline.

Transition: OpenState → ClosedState

onOpenClientSideLoadSnapshot :: IsTx tx => OpenState tx -> ConfirmedSnapshot tx -> Outcome tx

Client request to side load confirmed snapshot.

Note this is not covered by the spec as it is not reachable from an organic use of the protocol.

It must not have any effects outside of a neutral modification of the state to: * something it was before (in the case of the initial snapshot). * something it would be using side communication (in the case of a confirmed snapshot).

Besides the above, it is expected to work very much like the confirmed snapshot.

Transition: OpenState → OpenState

onClosedChainContestTx

Arguments

:: ClosedState tx

-> ChainStateType tx
New chain state.

-> SnapshotNumber

-> UTCTime
Contestation deadline.

-> Outcome tx

Observe a contest transaction. If the contested snapshot number is smaller than our last confirmed snapshot, we post a contest transaction.

Transition: ClosedState → ClosedState

onClosedClientFanout :: ClosedState tx -> Outcome tx

Client request to fanout leads to a fanout transaction on chain using the latest confirmed snapshot from ClosedState.

Transition: ClosedState → ClosedState

onClosedChainFanoutTx

Arguments

:: ClosedState tx

-> ChainStateType tx
New chain state

-> UTxOType tx

-> Outcome tx

Observe a fanout transaction by finalize the head state and notifying clients about it.

Transition: ClosedState → IdleState

data Input tx

Inputs that are processed by the head logic (the "core"). Corresponding to each of the "shell" layers, we distinguish between inputs from the client, the network and the chain.

Constructors

ClientInput
Input received from clients via the Hydra.API.

Fields

clientInput :: ClientInput tx

NetworkInput
Input received from peers via a Hydra.Network.

$sel:ttl:ClientInput is a simple counter that's decreased every time the event is reenqueued due to a wait. It's default value is defaultTTL

Fields

ttl :: TTL

networkEvent :: NetworkEvent (Message tx)

ChainInput
Input received from the chain via a Hydra.Chain.

Fields

chainEvent :: ChainEvent tx

type TTL = Natural

data LogicError tx

Constructors

UnhandledInput

Fields

input :: Input tx

currentHeadState :: HeadState tx

RequireFailed

Fields

requirementFailure :: RequirementFailure tx

AssertionFailed

Fields

message :: Text

NotOurHead

Fields

ourHeadId :: HeadId

otherHeadId :: HeadId

SideLoadSnapshotFailed

Fields

sideLoadRequirementFailure :: SideLoadRequirementFailure tx

data RequirementFailure tx

Constructors

ReqSnNumberInvalid

Fields

requestedSn :: SnapshotNumber

lastSeenSn :: SnapshotNumber

ReqSvNumberInvalid

Fields

requestedSv :: SnapshotVersion

lastSeenSv :: SnapshotVersion

ReqSnNotLeader

Fields

requestedSn :: SnapshotNumber

leader :: Party

ReqSnDecommitNotSettled

ReqSnCommitNotSettled

InvalidMultisignature

Fields

multisig :: Text

vkeys :: [VerificationKey HydraKey]

SnapshotAlreadySigned

Fields

knownSignatures :: [Party]

receivedSignature :: Party

AckSnNumberInvalid

Fields

requestedSn :: SnapshotNumber

lastSeenSn :: SnapshotNumber

SnapshotDoesNotApply

Fields

requestedSn :: SnapshotNumber

txid :: TxIdType tx

error :: ValidationError

NoMatchingDeposit

RequestedDepositExpired

Fields

depositTxId :: TxIdType tx

data SideLoadRequirementFailure tx

Constructors

SideLoadInitialSnapshotMismatch

SideLoadSnNumberInvalid

Fields

requestedSn :: SnapshotNumber

lastSeenSn :: SnapshotNumber

SideLoadSvNumberInvalid

Fields

requestedSv :: SnapshotVersion

lastSeenSv :: SnapshotVersion

SideLoadUTxOToCommitInvalid

Fields

requestedSc :: Maybe (UTxOType tx)

lastSeenSc :: Maybe (UTxOType tx)

SideLoadUTxOToDecommitInvalid

Fields

requestedSd :: Maybe (UTxOType tx)

lastSeenSd :: Maybe (UTxOType tx)

SideLoadInvalidMultisignature

Fields

multisig :: Text

vkeys :: [VerificationKey HydraKey]

data NodeState tx

Constructors

NodeState

Fields

headState :: HeadState tx

pendingDeposits :: PendingDeposits tx

Pending deposits as observed on chain. TODO: could even move the chain state here (also see todo below) , chainState :: ChainStateType tx

currentSlot :: ChainSlot

data Deposit tx

A deposit tracked by the protocol. The DepositStatus determines whether it may be used for an incremental commit or not.

Constructors

Deposit

Fields

headId :: HeadId

deposited :: UTxOType tx

created :: UTCTime

deadline :: UTCTime

status :: DepositStatus

data HeadState tx

The main state of the Hydra protocol state machine. It holds both, the overall protocol state, but also the off-chain CoordinatedHeadState.

Each of the sub-types (InitialState, OpenState, etc.) contain a black-box IdleState corresponding to the ChainEvent that has been observed leading to the state.

Note that rollbacks are currently not fully handled in the head logic and only this internal chain state gets replaced with the "rolled back to" version.

TODO: chainState would actually not be needed in the HeadState anymore as we do not persist the HeadState and not access it in the HeadLogic either.

Constructors

Idle (IdleState tx)

Initial (InitialState tx)

Open (OpenState tx)

Closed (ClosedState tx)

data InitialState tx

An Initial head which already has an identity and is collecting commits.

Constructors

InitialState

Fields

parameters :: HeadParameters

pendingCommits :: PendingCommits

committed :: Committed tx

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

data OpenState tx

An Open head with a CoordinatedHeadState tracking off-chain transactions.

Constructors

OpenState

Fields

parameters :: HeadParameters

coordinatedHeadState :: CoordinatedHeadState tx

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

data ClosedState tx

An Closed head with an current candidate ConfirmedSnapshot, which may be contested before the $sel:contestationDeadline:ClosedState.

Constructors

ClosedState

Fields

parameters :: HeadParameters

confirmedSnapshot :: ConfirmedSnapshot tx

contestationDeadline :: UTCTime

readyToFanoutSent :: Bool

Tracks whether we have informed clients already about being ReadyToFanout.

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

version :: SnapshotVersion

data SeenSnapshot tx

Data structure to help in tracking whether we have seen or requested a ReqSn already and if seen, the signatures we collected already.

Constructors

NoSeenSnapshot
Never saw a ReqSn.

LastSeenSnapshot
No snapshot in flight with last seen snapshot number as given.

Fields

lastSeen :: SnapshotNumber

RequestedSnapshot
ReqSn was sent out and it should be considered already in flight.

Fields

lastSeen :: SnapshotNumber

requested :: SnapshotNumber

SeenSnapshot
ReqSn for given snapshot was received.

Fields

snapshot :: Snapshot tx

signatories :: Map Party (Signature (Snapshot tx))

Collected signatures and so far.

type Committed tx = Map Party (UTxOType tx)

data CoordinatedHeadState tx

Off-chain state of the Coordinated Head protocol.

Constructors

CoordinatedHeadState

Fields

localUTxO :: UTxOType tx

The latest UTxO resulting from applying $sel:localTxs:CoordinatedHeadState to CoordinatedHeadState. Spec: L̂

localTxs :: [tx]

List of transactions applied locally and pending inclusion in a snapshot. Ordering in this list is important as transactions are added in order of application. Spec: T̂

allTxs :: !(Map (TxIdType tx) tx)

Map containing all the transactions ever seen by this node and not yet included in a snapshot. Spec: Tall

confirmedSnapshot :: ConfirmedSnapshot tx

The latest confirmed snapshot. Spec: S̅

seenSnapshot :: SeenSnapshot tx

Last seen snapshot and signatures accumulator. Spec: Û, ŝ and Σ̂

currentDepositTxId :: Maybe (TxIdType tx)

Current/next deposit to incrementally commit. Spec: Uα TODO: update in spec: Uα -> tx^#α

decommitTx :: Maybe tx

Pending decommit transaction. Spec: txω

version :: SnapshotVersion

Last open state version as observed on chain. Spec: ̂v

data DepositStatus

Constructors

Inactive

Active

Expired

newtype IdleState tx

An Idle head only having a chain state with things seen on chain so far.

Constructors

IdleState

Fields

chainState :: ChainStateType tx

type PendingCommits = Set Party

type PendingDeposits tx = Map (TxIdType tx) (Deposit tx)

getChainState :: HeadState tx -> ChainStateType tx

Get the chain state in any HeadState.

seenSnapshotNumber :: SeenSnapshot tx -> SnapshotNumber

Get the last seen snapshot number given a SeenSnapshot.

setChainState :: ChainStateType tx -> HeadState tx -> HeadState tx

Update the chain state in any HeadState.

data Effect tx

Analogous to inputs, the pure head logic "core" can have effects emitted to the "shell" layers and we distinguish the same: effects onto the client, the network and the chain.

Constructors

ClientEffect
Effect to be handled by the Hydra.API, results in sending this ClientMessage.

Fields

clientMessage :: ClientMessage tx

NetworkEffect
Effect to be handled by a Hydra.Network, results in a broadcast.

Fields

message :: Message tx

OnChainEffect
Effect to be handled by a Hydra.Chain, results in a postTx.

Fields

postChainTx :: PostChainTx tx

data Outcome tx

Constructors

Continue
Continue with the given state updates and side effects.

Fields

stateChanges :: [StateChanged tx]

effects :: [Effect tx]

Wait
Wait for some condition to be met with optional state updates.

Fields

reason :: WaitReason tx

stateChanges :: [StateChanged tx]

Error
Processing resulted in an error.

Fields

error :: LogicError tx

data StateChanged tx

Head state changed event. These events represent all the internal state changes, get persisted and processed in an event sourcing manner.

Constructors

NetworkConnected

NetworkDisconnected

PeerConnected

Fields

peer :: Host

PeerDisconnected

Fields

peer :: Host

NetworkVersionMismatch

Fields

ourVersion :: ProtocolVersion

theirVersion :: Maybe ProtocolVersion

NetworkClusterIDMismatch

Fields

clusterPeers :: Text

misconfiguredPeers :: Text

HeadInitialized

Fields

parameters :: HeadParameters

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

parties :: [Party]

CommittedUTxO

Fields

headId :: HeadId

party :: Party

committedUTxO :: UTxOType tx

chainState :: ChainStateType tx

HeadAborted

Fields

headId :: HeadId

utxo :: UTxOType tx

chainState :: ChainStateType tx

HeadOpened

Fields

headId :: HeadId

chainState :: ChainStateType tx

initialUTxO :: UTxOType tx

TransactionReceived

Fields

tx :: tx

TransactionAppliedToLocalUTxO

Fields

headId :: HeadId

tx :: tx

newLocalUTxO :: UTxOType tx

SnapshotRequestDecided

Fields

snapshotNumber :: SnapshotNumber

SnapshotRequested
A snapshot was requested by some party. NOTE: We deliberately already include an updated local ledger state to not need a ledger to interpret this event.

Fields

snapshot :: Snapshot tx

requestedTxIds :: [TxIdType tx]

newLocalUTxO :: UTxOType tx

newLocalTxs :: [tx]

newCurrentDepositTxId :: Maybe (TxIdType tx)

PartySignedSnapshot

Fields

snapshot :: Snapshot tx

party :: Party

signature :: Signature (Snapshot tx)

SnapshotConfirmed

Fields

headId :: HeadId

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

DepositRecorded

Fields

chainState :: ChainStateType tx

headId :: HeadId

depositTxId :: TxIdType tx

deposited :: UTxOType tx

created :: UTCTime

deadline :: UTCTime

DepositActivated

Fields

depositTxId :: TxIdType tx

chainTime :: UTCTime

deposit :: Deposit tx

DepositExpired

Fields

depositTxId :: TxIdType tx

chainTime :: UTCTime

deposit :: Deposit tx

DepositRecovered

Fields

chainState :: ChainStateType tx

headId :: HeadId

depositTxId :: TxIdType tx

recovered :: UTxOType tx

CommitApproved

Fields

headId :: HeadId

utxoToCommit :: UTxOType tx

CommitFinalized

Fields

chainState :: ChainStateType tx

headId :: HeadId

newVersion :: SnapshotVersion

depositTxId :: TxIdType tx

DecommitRecorded

Fields

headId :: HeadId

decommitTx :: tx

newLocalUTxO :: UTxOType tx

utxoToDecommit :: UTxOType tx

DecommitApproved

Fields

headId :: HeadId

decommitTxId :: TxIdType tx

utxoToDecommit :: UTxOType tx

DecommitInvalid

Fields

headId :: HeadId

decommitTx :: tx

decommitInvalidReason :: DecommitInvalidReason tx

DecommitFinalized

Fields

chainState :: ChainStateType tx

headId :: HeadId

distributedUTxO :: UTxOType tx

newVersion :: SnapshotVersion

HeadClosed

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

chainState :: ChainStateType tx

contestationDeadline :: UTCTime

HeadContested

Fields

headId :: HeadId

chainState :: ChainStateType tx

contestationDeadline :: UTCTime

snapshotNumber :: SnapshotNumber

HeadIsReadyToFanout

Fields

headId :: HeadId

HeadFannedOut

Fields

headId :: HeadId

utxo :: UTxOType tx

chainState :: ChainStateType tx

ChainRolledBack

Fields

chainState :: ChainStateType tx

TickObserved

Fields

chainSlot :: ChainSlot

IgnoredHeadInitializing

Fields

headId :: HeadId

contestationPeriod :: ContestationPeriod

parties :: [Party]

participants :: [OnChainId]

TxInvalid

Fields

headId :: HeadId

utxo :: UTxOType tx

transaction :: tx

validationError :: ValidationError

LocalStateCleared

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

Checkpoint

Fields

state :: NodeState tx

data WaitReason tx

Constructors

WaitOnNotApplicableTx

Fields

validationError :: ValidationError

WaitOnSnapshotNumber

Fields

waitingForNumber :: SnapshotNumber

WaitOnSnapshotVersion

Fields

waitingForVersion :: SnapshotVersion

WaitOnSeenSnapshot

WaitOnTxs

Fields

waitingForTxIds :: [TxIdType tx]

WaitOnContestationDeadline

WaitOnNotApplicableDecommitTx

Fields

notApplicableReason :: DecommitInvalidReason tx

WaitOnUnresolvedCommit

Fields

commitUTxO :: UTxOType tx

WaitOnUnresolvedDecommit

Fields

decommitTx :: tx

WaitOnDepositObserved

Fields

depositTxId :: TxIdType tx

WaitOnDepositActivation

Fields

depositTxId :: TxIdType tx

cause :: Effect tx -> Outcome tx

causes :: [Effect tx] -> Outcome tx

changes :: [StateChanged tx] -> Outcome tx

newState :: StateChanged tx -> Outcome tx

noop :: Outcome tx

wait :: WaitReason tx -> Outcome tx