Hydra.HeadLogic

Documentation

data Event tx

Constructors

ClientEvent (ClientRequest tx)

NetworkEvent (HydraMessage tx)

OnChainEvent (OnChainTx tx)

ShouldPostFanout

data Effect tx

Constructors

ClientEffect (ClientResponse tx)

NetworkEffect (HydraMessage tx)

OnChainEffect (OnChainTx tx)

Delay DiffTime (Event tx)

data ClientRequest tx

Constructors

Init [Party]

Commit (UTxO tx)

NewTx tx

Close

Contest

type SnapshotNumber = Natural

data Snapshot tx

Constructors

Snapshot

Fields

number :: SnapshotNumber

utxo :: UTxO tx

confirmed :: [tx]

The set of transactions that lead to utxo

data ClientResponse tx

Constructors

PeerConnected Party

ReadyToCommit

HeadIsOpen (UTxO tx)

HeadIsClosed DiffTime (Snapshot tx) [tx]

HeadIsFinalized (UTxO tx)

CommandFailed

TxConfirmed tx

TxInvalid tx

SnapshotConfirmed SnapshotNumber

data HydraMessage tx

Constructors

ReqTx tx

AckTx Party tx

ReqSn Party SnapshotNumber [tx]

AckSn Party SnapshotNumber

Ping Party

data OnChainTx tx

Constructors

InitTx (Set Party)

CommitTx Party (UTxO tx)

CollectComTx (UTxO tx)

CloseTx (Snapshot tx) [tx]

ContestTx (Snapshot tx) [tx]

FanoutTx (UTxO tx)

data HeadState tx

Constructors

HeadState

Fields

headParameters :: HeadParameters

headStatus :: HeadStatus tx

data HeadStatus tx

Constructors

InitState

CollectingState PendingCommits (Committed tx)

OpenState (SimpleHeadState tx)

ClosedState (UTxO tx)

FinalState

data SimpleHeadState tx

Constructors

SimpleHeadState

Fields

confirmedUTxO :: UTxO tx

unconfirmedTxs :: Map tx (Set Party)

confirmedTxs :: [tx]

confirmedSnapshot :: Snapshot tx

seenSnapshot :: Maybe (Snapshot tx, Set Party)

type PendingCommits = Set Party

data HeadParameters

Contains at least the contestation period and other things.

Constructors

HeadParameters

Fields

contestationPeriod :: DiffTime

parties :: Set Party

data SnapshotStrategy

Decides if snapshots should be done, or not.

Constructors

NoSnapshots

SnapshotAfterEachTx

createHeadState :: [Party] -> HeadParameters -> HeadState tx

Assume: We know the party members and their verification keys. These need to be exchanged somehow, eventually.

data LogicError tx

Preliminary type for collecting errors occurring during update. Might make sense to merge this (back) into Outcome.

Constructors

InvalidEvent (Event tx) (HeadState tx)

InvalidState (HeadState tx)

LedgerError ValidationError

data Outcome tx

Constructors

NewState (HeadState tx) [Effect tx]

Wait

Error (LogicError tx)

data Environment

Constructors

Environment

Fields

party :: Party

This is the p_i from the paper

snapshotStrategy :: SnapshotStrategy

update :: (Tx tx, Ord tx) => Environment -> Ledger tx -> HeadState tx -> Event tx -> Outcome tx

The heart of the Hydra head logic, a handler of all kinds of Event in the Hydra head. This may also be split into multiple handlers, i.e. one for hydra network events, one for client events and one for main chain events, or by sub-State.