Hydra.HeadLogic

Documentation

data Event tx

Constructors

ClientEvent

Fields

clientInput :: ClientInput tx

NetworkEvent

Fields

message :: Message tx

OnChainEvent

Fields

onChainTx :: OnChainTx tx

ShouldPostFanout

DoSnapshot

data Effect tx

Constructors

ClientEffect

Fields

serverOutput :: ServerOutput tx

NetworkEffect

Fields

message :: Message tx

OnChainEffect

Fields

onChainTx :: PostChainTx tx

Delay

Fields

delay :: DiffTime

event :: Event tx

data HeadState tx

Constructors

ReadyState

InitialState

Fields

parameters :: HeadParameters

pendingCommits :: PendingCommits

committed :: Committed tx

OpenState

Fields

parameters :: HeadParameters

coordinatedHeadState :: CoordinatedHeadState tx

ClosedState

Fields

parameters :: HeadParameters

utxos :: Utxo tx

data CoordinatedHeadState tx

Constructors

CoordinatedHeadState

Fields

seenUtxo :: Utxo tx

seenTxs :: [tx]

confirmedSnapshot :: Snapshot tx

seenSnapshot :: SeenSnapshot tx

data SeenSnapshot tx

Constructors

NoSeenSnapshot

RequestedSnapshot

SeenSnapshot

Fields

snapshot :: Snapshot tx

signatories :: Set Party

type PendingCommits = Set Party

data SnapshotStrategy

Decides if snapshots should be done, or not.

Constructors

NoSnapshots

SnapshotAfterEachTx

data LogicError tx

Preliminary type for collecting errors occurring during update. Might make sense to merge this (back) into Outcome.

Constructors

InvalidEvent (Event tx) (HeadState tx)

InvalidState (HeadState tx)

InvalidSnapshot

Fields

expected :: SnapshotNumber

actual :: SnapshotNumber

LedgerError ValidationError

data Outcome tx

Constructors

NewState (HeadState tx) [Effect tx]

Wait

Error (LogicError tx)

data Environment

Constructors

Environment

Fields

party :: Party

This is the p_i from the paper

signingKey :: SigningKey

otherParties :: [Party]

snapshotStrategy :: SnapshotStrategy

update :: Tx tx => Environment -> Ledger tx -> HeadState tx -> Event tx -> Outcome tx

The heart of the Hydra head logic, a handler of all kinds of Event in the Hydra head. This may also be split into multiple handlers, i.e. one for hydra network events, one for client events and one for main chain events, or by sub-State.