Hydra.HeadLogic

Description

Implements the Head Protocol's state machine as a pure function.

The protocol is described in two parts in the Hydra paper

One part detailing how the Head deals with client input.

Another part describing how the Head reacts to network messages from peers.

A third part detailing the On-Chain Verification (OCV) protocol, i.e. the abstract "smart contracts" that are need to provide on-chain security.

This module is about the first two parts, while the Hydra.Contract.Head module in 'hydra-plutus' covers the third part.

Types

data Event tx

The different events which are processed by the head logic (the "core"). Corresponding to each of the "shell" layers, we distinguish between events from the client, the network and the chain.

Constructors

ClientEvent
Event received from clients via the Hydra.API.

Fields

clientInput :: ClientInput tx

NetworkEvent
Event received from peers via a Hydra.Network.

$sel:ttl:ClientEvent is a simple counter that's decreased every time the event is reenqueued due to a wait. It's default value is defaultTTL

Fields

ttl :: TTL

message :: Message tx

OnChainEvent
Event received from the chain via a Hydra.Chain.

Fields

chainEvent :: ChainEvent tx

PostTxError
Event to re-ingest errors from postTx for further processing.

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

data Effect tx

Analogous to events, the pure head logic "core" can have effects emited to the "shell" layers and we distinguish the same: effects onto the client, the network and the chain.

Constructors

ClientEffect
Effect to be handled by the Hydra.API, results in sending this ServerOutput.

Fields

serverOutput :: ServerOutput tx

NetworkEffect
Effect to be handled by a Hydra.Network, results in a broadcast.

Fields

message :: Message tx

OnChainEffect
Effect to be handled by a Hydra.Chain, results in a postTx.

Fields

chainState :: ChainStateType tx

postChainTx :: PostChainTx tx

data HeadState tx

The main state of the Hydra protocol state machine. It holds both, the overall protocol state, but also the off-chain CoordinatedHeadState.

It is a recursive data structure, where $sel:previousRecoverableState:IdleState fields record the state before the latest OnChainEvent that has been observed. On-Chain events are indeed only eventually immutable and the application state may be rolled back at any time (with a decreasing probability as the time pass).

Thus, leverage functional immutable data-structure, we build a recursive structure of states which we can easily navigate backwards when needed (see Rollback and rollback).

Note that currently, rolling back to a previous recoverable state eliminates any off-chain events (e.g. transactions) that happened after that state. This is particularly important for anything following the transition to OpenState since this is where clients may start submitting transactions. In practice, clients should not send transactions right way but wait for a certain grace period to minimize the risk.

Constructors

IdleState

Fields

chainState :: ChainStateType tx

InitialState

Fields

parameters :: HeadParameters

pendingCommits :: PendingCommits

committed :: Committed tx

previousRecoverableState :: HeadState tx

chainState :: ChainStateType tx

OpenState

Fields

parameters :: HeadParameters

coordinatedHeadState :: CoordinatedHeadState tx

previousRecoverableState :: HeadState tx

chainState :: ChainStateType tx

ClosedState

Fields

parameters :: HeadParameters

confirmedSnapshot :: ConfirmedSnapshot tx

previousRecoverableState :: HeadState tx

contestationDeadline :: UTCTime

readyToFanoutSent :: Bool

Tracks whether we have informed clients already about being ReadyToFanout.

chainState :: ChainStateType tx

getChainState :: HeadState tx -> ChainStateType tx

type Committed tx = Map Party (UTxOType tx)

data CoordinatedHeadState tx

Off-chain state of the Coordinated Head protocol.

Constructors

CoordinatedHeadState

Fields

seenUTxO :: UTxOType tx

The latest UTxO of the "seen ledger".

seenTxs :: [tx]

List of seen transactions.

confirmedSnapshot :: ConfirmedSnapshot tx

The latest confirmed snapshot, representing the "confirmed ledger".

seenSnapshot :: SeenSnapshot tx

Whether we are currently collecting signatures for a snapshot.

data SeenSnapshot tx

Data structure to help in tracking whether we are currently collecting signatures for a snapshot.

Constructors

NoSeenSnapshot

RequestedSnapshot

SeenSnapshot

Fields

snapshot :: Snapshot tx

signatories :: Map Party (Signature (Snapshot tx))

type PendingCommits = Set Party

type TTL = Natural

defaultTTL :: TTL

data LogicError tx

Preliminary type for collecting errors occurring during update. TODO: Try to merge this (back) into Outcome.

Constructors

InvalidEvent (Event tx) (HeadState tx)

InvalidState (HeadState tx)

InvalidSnapshot

Fields

expected :: SnapshotNumber

actual :: SnapshotNumber

LedgerError ValidationError

data Outcome tx

Constructors

OnlyEffects [Effect tx]

NewState (HeadState tx) [Effect tx]

Wait WaitReason

Error (LogicError tx)

data WaitReason

Constructors

WaitOnNotApplicableTx

Fields

validationError :: ValidationError

WaitOnSnapshotNumber

Fields

waitingFor :: SnapshotNumber

WaitOnSeenSnapshot

WaitOnContestationDeadline

data Environment

Constructors

Environment

Fields

party :: Party

This is the p_i from the paper

signingKey :: SigningKey HydraKey

otherParties :: [Party]

The Coordinated Head protocol

Opening the Head

onIdleClientInit

Arguments

:: ChainStateType tx
Current chain state

-> Party
Us

-> [Party]
Others

-> ContestationPeriod

-> Outcome tx

Client request to init the head. This leads to an init transaction on chain, containing the head parameters.

TODO: maybe change signature so it takes [Party] instead (all parties)?

Transition: IdleState → IdleState

onIdleChainInitTx

Arguments

:: HeadState tx
Current head state.

-> ChainStateType tx
New chain state.

-> [Party]

-> ContestationPeriod

-> Outcome tx

Observe an init transaction, initialize parameters in an InitialState and notify clients that they can now commit.

Transition: IdleState → InitialState

onInitialClientCommit

Arguments

:: ChainStateType tx
Current chain state

-> Party

-> PendingCommits

-> ClientInput tx

-> Outcome tx

Client request to commit a UTxO entry to the head. Provided the client hasn't committed yet, this leads to a commit transaction on-chain containing that UTxO entry.

Transition: InitialState → InitialState

onInitialChainCommitTx

Arguments

:: Monoid (UTxOType tx)

=> HeadState tx
Current head state; recorded as previous recoverable state

-> ChainStateType tx
New chain state

-> Party
Us

-> HeadParameters

-> PendingCommits

-> Committed tx

-> Party
Comitting party

-> UTxOType tx
Committed UTxO

-> Outcome tx

Observe a commit transaction and record the committed UTxO in the state. Also, if this is the last commit to be observed, post a collect-com transaction on-chain.

Transition: InitialState → InitialState

onInitialClientAbort

Arguments

:: Monoid (UTxOType tx)

=> ChainStateType tx
Current chain state

-> Committed tx

-> Outcome tx

Client request to abort the head. This leads to an abort transaction on chain, reimbursing already committed UTxOs.

Transition: InitialState → InitialState

onInitialChainAbortTx

Arguments

:: Monoid (UTxOType tx)

=> ChainStateType tx
New chain state

-> Committed tx

-> Outcome tx

Observe an abort transaction by switching the state and notifying clients about it.

Transition: InitialState → IdleState

onInitialChainCollectTx

Arguments

:: (Foldable t, Monoid (UTxOType tx))

=> HeadState tx
Current head state; recorded as previous recoverable state

-> ChainStateType tx
New chain state

-> HeadParameters

-> t (UTxOType tx)

-> Outcome tx

Observe a collectCom transaction. We initialize the OpenState using the head parameters from IdleState and construct an InitialSnapshot holding u0 from the committed UTxOs.

Transition: InitialState → OpenState

Off-chain protocol

onOpenClientNewTx

Arguments

:: Ledger tx

-> Party
Us

-> UTxOType tx
UTxO from the last confirmed snapshot, a.k.a the confirmed ledger state.

-> tx
The transaction to be submitted to the head.

-> Outcome tx

Client request to ingest a new transaction into the head. We do check whether the given transaction can be applied against the confirmed ledger state and yield a corresponding TxValid or TxInvalid client response.

Transition: OpenState → OpenState

onOpenNetworkReqTx

Arguments

:: Ledger tx

-> HeadParameters

-> HeadState tx
Previous recoverable state (to re-create OpenState)

-> ChainStateType tx
Current chain state (to re-create OpenState)

-> CoordinatedHeadState tx
The offchain coordinated head state

-> tx
The transaction to be submitted to the head.

-> Outcome tx

Receive network message about a new transaction request (ReqTx) from a peer. We apply this transaction to the seen utxo (ledger state), resulting in an updated seen ledger state. If it is not applicable, then we wait to retry later.

Transition: OpenState → OpenState

onOpenNetworkReqSn

Arguments

:: IsTx tx

=> Ledger tx

-> Party

-> SigningKey HydraKey

-> HeadState tx
Previous recoverable state (to re-create OpenState)

-> ChainStateType tx
Current chain state (to re-create OpenState)

-> HeadParameters

-> CoordinatedHeadState tx
The offchain coordinated head state

-> Party

-> SnapshotNumber

-> [tx]

-> HeadState tx

-> Event tx

-> Outcome tx

Receive network message about a snapshot request (ReqSn) from a peer. We do distinguish two cases:

Case 1:

The peer is the leader for requested snapshot number.

Snapshot number is the next expected (based on the last confirmed)

There is no snapshot pending, i.e. we are not collecting any signatures for a snapshot.

We try to apply the transactions of the requested snapshot to the confirmed utxo:

If that succeeds, we do sign the snapshot, yield a snapshot acknowledgment (AckSn) and start tracking this snapshot.

Else, we wait until the transactions become applicable.

Case 2:

The peer is the leader for requested snapshot number.

Snapshot number is greater than the next expected.

We wait for the snapshots in between, i.e. until this ReqSn is the next.

Transition: OpenState → OpenState

onOpenNetworkAckSn

Arguments

:: IsTx tx

=> [Party]

-> Party

-> HeadParameters

-> HeadState tx
Previous recoverable state (to re-create OpenState)

-> ChainStateType tx
Current chain state (to re-create OpenState)

-> Signature (Snapshot tx)

-> CoordinatedHeadState tx
The offchain coordinated head state

-> SnapshotNumber

-> Outcome tx

Receive network message about a snapshot acknowledgement (AckSn) from a peer. We do distinguish two cases:

Case 1: we received an AckSn request we did not expect

respective AckSn and ReqSn out of order.

multiple AckSns out of order.

In this case we simply wait to see the expected AckSn and we reenqueue the event.

The reason this can happen is because we don't make any assumptions on the network packet delivery, and therefore the messages can arrive in any order.

Case 2: we received the expected Ack

provided that the signature is valid, we add it to the set of signatories we have

when we have gather all the signatures then we confirm the snapshot.

when the signature is not valid then nothing changes.

Transition: OpenState → OpenState

Closing the Head

onOpenClientClose

Arguments

:: ChainStateType tx
Current chain state

-> ConfirmedSnapshot tx

-> Outcome tx

Client request to close the head. This leads to a close transaction on chain using the latest confirmed snaphshot of the OpenState.

Transition: OpenState → OpenState

onOpenChainCloseTx

Arguments

:: HeadParameters

-> HeadState tx
Current head state; recorded as previous recoverable state

-> ChainStateType tx
New chain state

-> CoordinatedHeadState tx
The offchain coordinated head state

-> SnapshotNumber

-> UTCTime

-> Outcome tx

Observe a close transaction. If the closed snapshot number is smaller than our last confirmed, we post a contest transaction. Also, we do schedule a notification for clients to fanout at the deadline.

Transition: OpenState → ClosedState

onClosedChainContestTx

Arguments

:: ChainStateType tx
Current chain state

-> ConfirmedSnapshot tx

-> SnapshotNumber

-> Outcome tx

Observe a contest transaction. If the contested snapshot number is smaller than our last confirmed snapshot, we post a contest transaction.

Transition: ClosedState → ClosedState

onClosedClientFanout

Arguments

:: ChainStateType tx
Current chain state

-> ConfirmedSnapshot tx

-> UTCTime

-> Outcome tx

Client request to fanout leads to a fanout transaction on chain using the latest confirmed snapshot from ClosedState.

Transition: ClosedState → ClosedState

onClosedChainFanoutTx

Arguments

:: ChainStateType tx
New chain state

-> ConfirmedSnapshot tx

-> Outcome tx

Observe a fanout transaction by finalize the head state and notifying clients about it.

Transition: ClosedState → IdleState

onCurrentChainRollback :: IsChainState tx => HeadState tx -> ChainSlot -> Outcome tx

Observe a chain rollback and transition to corresponding previous recoverable state.

Transition: OpenState → HeadState

update :: (IsTx tx, IsChainState tx) => Environment -> Ledger tx -> HeadState tx -> Event tx -> Outcome tx

The "pure core" of the Hydra node, which handles the Event against a current HeadState. Resulting new HeadStates are retained and Effect outcomes handled by the Hydra.Node.

data SnapshotOutcome tx

Constructors

ShouldSnapshot SnapshotNumber [tx]

ShouldNotSnapshot NoSnapshotReason

data NoSnapshotReason

Constructors

NotLeader SnapshotNumber

SnapshotInFlight SnapshotNumber

NoTransactionsToSnapshot

isLeader :: HeadParameters -> Party -> SnapshotNumber -> Bool

newSn :: IsTx tx => Environment -> HeadParameters -> CoordinatedHeadState tx -> SnapshotOutcome tx

Snapshot emission decider

emitSnapshot :: IsTx tx => Environment -> [Effect tx] -> HeadState tx -> (HeadState tx, [Effect tx])