Hydra.HeadLogic

Description

Implements the Head Protocol's _state machine_ as a _pure function_.

The protocol is described in two parts in the Hydra paper:

One part detailing how the Head deals with _clients input_, ClientInput and ServerOutput

Another part detailing how the Head reacts to _peers input_ provided by the network, Message

Documentation

data Event tx

Constructors

ClientEvent

Fields

clientInput :: ClientInput tx

NetworkEvent

Fields

message :: Message tx

OnChainEvent

Fields

chainEvent :: ChainEvent tx

ShouldPostFanout

PostTxError

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

data Effect tx

Constructors

ClientEffect

Fields

serverOutput :: ServerOutput tx

NetworkEffect

Fields

message :: Message tx

OnChainEffect

Fields

onChainTx :: PostChainTx tx

Delay

Fields

delay :: NominalDiffTime

reason :: WaitReason

event :: Event tx

data HeadState tx

A recursive data-structure which represent the application state as a state-machine. The $sel:previousRecoverableState:IdleState records the state of the application before the latest OnChainEvent that has been observed. On-Chain events are indeed only eventually immutable and the application state may be rolled back at any time (with a decreasing probability as the time pass).

Thus, leverage functional immutable data-structure, we build a recursive structure of states which we can easily navigate backwards when needed (see Rollback and rollback).

Note that currently, rolling back to a previous recoverable state eliminates any off-chain events (e.g. transactions) that happened after that state. This is particularly important for anything following the transition to OpenState since this is where clients may start submitting transactions. In practice, clients should not send transactions right way but wait for a certain grace period to minimize the risk.

Constructors

IdleState

InitialState

Fields

parameters :: HeadParameters

pendingCommits :: PendingCommits

committed :: Committed tx

previousRecoverableState :: HeadState tx

OpenState

Fields

parameters :: HeadParameters

coordinatedHeadState :: CoordinatedHeadState tx

previousRecoverableState :: HeadState tx

ClosedState

Fields

parameters :: HeadParameters

confirmedSnapshot :: ConfirmedSnapshot tx

previousRecoverableState :: HeadState tx

type Committed tx = Map Party (UTxOType tx)

data CoordinatedHeadState tx

Constructors

CoordinatedHeadState

Fields

seenUTxO :: UTxOType tx

seenTxs :: [tx]

confirmedSnapshot :: ConfirmedSnapshot tx

seenSnapshot :: SeenSnapshot tx

data SeenSnapshot tx

Constructors

NoSeenSnapshot

RequestedSnapshot

SeenSnapshot

Fields

snapshot :: Snapshot tx

signatories :: Map Party (Signature (Snapshot tx))

type PendingCommits = Set Party

data LogicError tx

Preliminary type for collecting errors occurring during update. Might make sense to merge this (back) into Outcome.

Constructors

InvalidEvent (Event tx) (HeadState tx)

InvalidState (HeadState tx)

InvalidSnapshot

Fields

expected :: SnapshotNumber

actual :: SnapshotNumber

LedgerError ValidationError

data Outcome tx

Constructors

NewState (HeadState tx) [Effect tx]

Wait WaitReason

Error (LogicError tx)

data WaitReason

Constructors

WaitOnNotApplicableTx

Fields

validationError :: ValidationError

WaitOnSnapshotNumber

Fields

waitingFor :: SnapshotNumber

WaitOnSeenSnapshot

WaitOnContestationPeriod

data Environment

Constructors

Environment

Fields

party :: Party

This is the p_i from the paper

signingKey :: SigningKey

otherParties :: [Party]

update :: IsTx tx => Environment -> Ledger tx -> HeadState tx -> Event tx -> Outcome tx

The heart of the Hydra head logic, a handler of all kinds of Event in the Hydra head. This may also be split into multiple handlers, i.e. one for hydra network events, one for client events and one for main chain events, or by sub-State.

data SnapshotOutcome tx

Constructors

ShouldSnapshot SnapshotNumber [tx]

ShouldNotSnapshot NoSnapshotReason

data NoSnapshotReason

Constructors

NotLeader SnapshotNumber

SnapshotInFlight SnapshotNumber

NoTransactionsToSnapshot

isLeader :: HeadParameters -> Party -> SnapshotNumber -> Bool

newSn :: IsTx tx => Environment -> HeadParameters -> CoordinatedHeadState tx -> SnapshotOutcome tx

Snapshot emission decider

emitSnapshot :: IsTx tx => Environment -> [Effect tx] -> HeadState tx -> (HeadState tx, [Effect tx])

rollback :: HasCallStack => Word -> HeadState tx -> HeadState tx

Unwind the HeadState to some depth.

The HeadState is rolled back a number of times to some previous state. It's an error to call this function with a depth that's larger than the current state depth. See rollback for the on-chain counterpart to this function.