Hydra.HeadLogic.Outcome

Documentation

data Effect tx

Analogous to inputs, the pure head logic "core" can have effects emited to the "shell" layers and we distinguish the same: effects onto the client, the network and the chain.

Constructors

ClientEffect
Effect to be handled by the Hydra.API, results in sending this ServerOutput.

Fields

serverOutput :: ServerOutput tx

NetworkEffect
Effect to be handled by a Hydra.Network, results in a broadcast.

Fields

message :: Message tx

OnChainEffect
Effect to be handled by a Hydra.Chain, results in a postTx.

Fields

postChainTx :: PostChainTx tx

data StateChanged tx

Head state changed event. These events represent all the internal state changes, get persisted and processed in an event sourcing manner.

Constructors

HeadInitialized

Fields

parameters :: HeadParameters

chainState :: ChainStateType tx

headId :: HeadId

headSeed :: HeadSeed

CommittedUTxO

Fields

party :: Party

committedUTxO :: UTxOType tx

chainState :: ChainStateType tx

HeadAborted

Fields

chainState :: ChainStateType tx

HeadOpened

Fields

chainState :: ChainStateType tx

initialUTxO :: UTxOType tx

TransactionReceived

Fields

tx :: tx

TransactionAppliedToLocalUTxO

Fields

tx :: tx

newLocalUTxO :: UTxOType tx

DecommitRecorded

Fields

decommitTx :: tx

newLocalUTxO :: UTxOType tx

SnapshotRequestDecided

Fields

snapshotNumber :: SnapshotNumber

SnapshotRequested
A snapshot was requested by some party. NOTE: We deliberately already include an updated local ledger state to not need a ledger to interpret this event.

Fields

snapshot :: Snapshot tx

requestedTxIds :: [TxIdType tx]

newLocalUTxO :: UTxOType tx

newLocalTxs :: [tx]

DecommitFinalized

Fields

newVersion :: SnapshotVersion

PartySignedSnapshot

Fields

snapshot :: Snapshot tx

party :: Party

signature :: Signature (Snapshot tx)

SnapshotConfirmed

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

HeadClosed

Fields

chainState :: ChainStateType tx

contestationDeadline :: UTCTime

HeadContested

Fields

chainState :: ChainStateType tx

contestationDeadline :: UTCTime

HeadIsReadyToFanout

HeadFannedOut

Fields

chainState :: ChainStateType tx

ChainRolledBack

Fields

chainState :: ChainStateType tx

TickObserved

Fields

chainSlot :: ChainSlot

genStateChanged :: (ArbitraryIsTx tx, IsChainState tx) => Environment -> Gen (StateChanged tx)

data Outcome tx

Constructors

Continue
Continue with the given state updates and side effects.

Fields

stateChanges :: [StateChanged tx]

effects :: [Effect tx]

Wait
Wait for some condition to be met with optional state updates.

Fields

reason :: WaitReason tx

stateChanges :: [StateChanged tx]

Error
Processing resulted in an error.

Fields

error :: LogicError tx

noop :: Outcome tx

wait :: WaitReason tx -> Outcome tx

newState :: StateChanged tx -> Outcome tx

cause :: Effect tx -> Outcome tx

causes :: [Effect tx] -> Outcome tx

data WaitReason tx

Constructors

WaitOnNotApplicableTx

Fields

validationError :: ValidationError

WaitOnSnapshotNumber

Fields

waitingForNumber :: SnapshotNumber

WaitOnSnapshotVersion

Fields

waitingForVersion :: SnapshotVersion

WaitOnSeenSnapshot

WaitOnTxs

Fields

waitingForTxIds :: [TxIdType tx]

WaitOnContestationDeadline

WaitOnNotApplicableDecommitTx

Fields

notApplicableReason :: DecommitInvalidReason tx