Hydra.Ledger.Cardano

Documentation

adjustUTxO :: Tx -> UTxO -> UTxO

Utility function to "adjust" a UTxO' set given a Tx

The inputs from the Tx are removed from the internal map of the UTxO' and the outputs added, correctly indexed by the TxIn. This function is useful to manually maintain a UTxO' set without caring too much about the Ledger rules.

fromChainSlot :: ChainSlot -> SlotNo

Simple conversion from a generic slot to a specific local one.

cardanoLedger :: Globals -> LedgerEnv LedgerEra -> Ledger Tx

Use the cardano-ledger as an in-hydra Ledger.

newLedgerEnv :: PParams LedgerEra -> LedgerEnv LedgerEra

Create a new ledger env from given protocol parameters.

mkTransferTx :: MonadFail m => NetworkId -> UTxO -> SigningKey PaymentKey -> VerificationKey PaymentKey -> m Tx

Build a zero-fee transaction which spends the first output owned by given signing key and transfers it in full to given verification key.

mkSimpleTx

Arguments

:: (TxIn, TxOut CtxUTxO)

-> (AddressInEra, Value)
Recipient address and amount.

-> SigningKey PaymentKey
Sender's signing key.

-> Either TxBodyError Tx

Build a zero-fee payment transaction.

mkRangedTx

Arguments

:: (TxIn, TxOut CtxUTxO)

-> (AddressInEra, Value)
Recipient address and amount.

-> SigningKey PaymentKey
Sender's signing key.

-> (Maybe TxValidityLowerBound, Maybe TxValidityUpperBound)

-> Either TxBodyError Tx

Create a zero-fee, payment cardano transaction with validity range.

genSequenceOfSimplePaymentTransactions :: Gen (UTxO, [Tx])

Generates a sequence of simple "transfer" transactions for a single key.

genFixedSizeSequenceOfSimplePaymentTransactions :: Int -> Gen (UTxO, [Tx])

data ShelleyGenesis

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !(NonZero Word64)

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams ShelleyEra)

sgGenDelegs :: !(Map (KeyHash 'Genesis) GenDelegPair)

sgInitialFunds :: ListMap Addr Coin

sgStaking :: ShelleyGenesisStaking

data Globals

data LedgerEnv era

type Tx = Tx Era

Orphan instances

Arbitrary (Hash PaymentKey)

Instance details

Methods

arbitrary :: Gen (Hash PaymentKey)

shrink :: Hash PaymentKey -> [Hash PaymentKey]