Hydra.Ledger.Cardano

Documentation

cardanoLedger :: Globals -> LedgerEnv LedgerEra -> Ledger Tx

Use the cardano-ledger as an in-hydra Ledger.

mkSimpleTx

Arguments

:: (TxIn, TxOut CtxUTxO)

-> (AddressInEra, Value)
Recipient address and amount.

-> SigningKey PaymentKey
Sender's signing key.

-> Either TxBodyError Tx

Create a zero-fee, payment cardano transaction.

renderTx :: IsString str => Tx -> str

Obtain a human-readable pretty text representation of a transaction.

renderTxs :: IsString str => [Tx] -> str

renderTxWithUTxO :: IsString str => UTxO -> Tx -> str

Like renderTx, but uses the given UTxO to resolve inputs.

genSigningKey :: Gen (SigningKey PaymentKey)

genVerificationKey :: Gen (VerificationKey PaymentKey)

genKeyPair :: Gen (VerificationKey PaymentKey, SigningKey PaymentKey)

genSequenceOfSimplePaymentTransactions :: Gen (UTxO, [Tx])

Generates a sequence of simple "transfer" transactions for a single key. The kind of transactions produced by this generator is very limited, see generateOneTransfer.

genFixedSizeSequenceOfSimplePaymentTransactions :: Int -> Gen (UTxO, [Tx])

generateOneTransfer :: NetworkId -> (UTxO, (VerificationKey PaymentKey, SigningKey PaymentKey), [Tx]) -> Int -> Gen (UTxO, (VerificationKey PaymentKey, SigningKey PaymentKey), [Tx])

genOutput :: forall ctx. VerificationKey PaymentKey -> Gen (TxOut ctx)

genTxOutAdaOnly :: Gen (TxOut ctx)

Generate an ada-only TxOut payed to an arbitrary public key.

genTxIn :: Gen TxIn

A more random generator than the 'Arbitrary TxIn' from cardano-ledger.

genUTxOAdaOnlyOfSize :: Int -> Gen UTxO

Generate a fixed size UTxO with ada-only outputs.

genUTxOAlonzo :: Gen UTxO

Generate Alonzo era UTxO, which may contain arbitrary assets in TxOuts addressed to public keys *and* scripts. NOTE: This is not reducing size when generating assets in TxOuts, so will end up regularly with 300+ assets with generator size 30. NOTE: The Arbitrary TxIn instance from the ledger is producing colliding values, so we replace them.

genUTxOFor :: VerificationKey PaymentKey -> Gen UTxO

Generate utxos owned by the given cardano key.

genOneUTxOFor :: VerificationKey PaymentKey -> Gen UTxO

Generate a single UTXO owned by vk.

genAddressInEra :: NetworkId -> Gen AddressInEra

NOTE: See note on mkVkAddress about NetworkId.

genValue :: Gen Value

genAdaValue :: Gen Value

genAdaOnlyUTxO :: Gen UTxO

Generate UTXO entries that do not contain any assets. Useful to test / measure cases where

adaOnly :: TxOut CtxUTxO -> TxOut CtxUTxO

genUTxOWithSimplifiedAddresses :: Gen UTxO

Generate "simplified" UTXO, ie. without some of the complexities required for backward-compatibility and obscure features.

simplifyUTxO :: UTxO -> UTxO

Rewrite given UTXO to remove some corner cases.

Remove Byron addresses. Those are unnecessarily complicated and deprecated so they should not be used either on- or off-chain

Replace stake pointers with StakeRefNull. Stake pointers is an obscure feature of Cardano addresses that's very rarely used in practice.

shrinkUTxO :: UTxO -> [UTxO]

shrinkValue :: Value -> [Value]

module Hydra.Ledger.Cardano.Builder

data ShelleyGenesis era

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTime

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams era)

sgGenDelegs :: !(Map (KeyHash 'Genesis (Crypto era)) (GenDelegPair (Crypto era)))

sgInitialFunds :: !(Map (Addr (Crypto era)) Coin)

sgStaking :: !(ShelleyGenesisStaking (Crypto era))

type Tx = Tx Era

Orphan instances

FromCBOR Tx

Instance details

Methods

fromCBOR :: Decoder s Tx

label :: Proxy Tx -> Text

FromCBOR UTxO

Instance details

Methods

fromCBOR :: Decoder s UTxO

label :: Proxy UTxO -> Text

ToCBOR Tx

Instance details

Methods

toCBOR :: Tx -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy Tx -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [Tx] -> Size

ToCBOR UTxO

Instance details

Methods

toCBOR :: UTxO -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy UTxO -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [UTxO] -> Size

FromJSON Tx

Instance details

Methods

parseJSON :: Value -> Parser Tx

parseJSONList :: Value -> Parser [Tx]

FromJSON UTxO

Instance details

Methods

parseJSON :: Value -> Parser UTxO

parseJSONList :: Value -> Parser [UTxO]

ToJSON Tx

Instance details

Methods

toJSON :: Tx -> Value

toEncoding :: Tx -> Encoding

toJSONList :: [Tx] -> Value

toEncodingList :: [Tx] -> Encoding

ToJSON UTxO

Instance details

Methods

toJSON :: UTxO -> Value

toEncoding :: UTxO -> Encoding

toJSONList :: [UTxO] -> Value

toEncodingList :: [UTxO] -> Encoding

Arbitrary Tx

Instance details

Methods

arbitrary :: Gen Tx

shrink :: Tx -> [Tx]

Arbitrary UTxO

Instance details

Methods

arbitrary :: Gen UTxO

shrink :: UTxO -> [UTxO]

Arbitrary AssetName

Instance details

Methods

arbitrary :: Gen AssetName

shrink :: AssetName -> [AssetName]

Arbitrary TxIn

Instance details

Methods

arbitrary :: Gen TxIn

shrink :: TxIn -> [TxIn]

Arbitrary TxId

Instance details

Methods

arbitrary :: Gen TxId

shrink :: TxId -> [TxId]

IsTx Tx

Instance details

Associated Types

type UTxOType Tx

type TxIdType Tx

type ValueType Tx

Methods

txId :: Tx -> TxIdType Tx

balance :: UTxOType Tx -> ValueType Tx

hashUTxO :: UTxOType Tx -> ByteString

Arbitrary (TxOut CtxUTxO)

Instance details

Methods

arbitrary :: Gen (TxOut CtxUTxO)

shrink :: TxOut CtxUTxO -> [TxOut CtxUTxO]

Arbitrary (VerificationKey PaymentKey)

Instance details

Methods

arbitrary :: Gen (VerificationKey PaymentKey)

shrink :: VerificationKey PaymentKey -> [VerificationKey PaymentKey]

Arbitrary (Hash PaymentKey)

Instance details

Methods

arbitrary :: Gen (Hash PaymentKey)

shrink :: Hash PaymentKey -> [Hash PaymentKey]