Hydra.Ledger.Cardano.Evaluate

Description

Simplified interface to phase-2 validation of transaction, eg. evaluation of Plutus scripts.

The evaluateTx function simplifies the call to underlying Plutus providing execution report using pre-canned PParams. This should only be used for testing or benchmarking purpose as the real evaluation parameters are set when the Hydra node starts.

NOTE: The reason this module is here instead of part of `test/` directory is to be used in tx-cost executable.

Documentation

type RedeemerReport = Map RdmrPtr (Either (ScriptFailure StandardCrypto) ExUnits)

evaluateTx :: Tx -> UTxO -> Either (BasicFailure StandardCrypto) RedeemerReport

evaluateTx'

Arguments

:: ExecutionUnits
Max tx execution units.

-> Tx

-> UTxO

-> Either (BasicFailure StandardCrypto) RedeemerReport

pparams :: PParams Era

Current mainchain cost parameters.

epochInfo :: Monad m => EpochInfo m

systemStart :: SystemStart

costModels :: Array Language CostModel