Hydra.Ledger.Cardano.Evaluate

Description

Simplified interface to phase-2 validation of transactions, eg. evaluation of Plutus scripts.

The evaluateTx function simplifies the call to ledger and plutus providing an EvaluationReport using pre-canned ProtocolParameters. This should only be used for testing or benchmarking purpose as the real evaluation parameters are set when the Hydra node starts.

NOTE: The reason this module is here instead of part of `test/` directory is to be used in tx-cost executable.

Documentation

evaluateTx :: Tx -> UTxO -> Either TransactionValidityError (Map ScriptWitnessIndex (Either ScriptExecutionError ExecutionUnits))

Thin wrapper around evaluateTransactionExecutionUnits, which uses fixtures for system start, era history and protocol parameters. See pparams.

evaluateTx'

Arguments

:: ExecutionUnits
Max tx execution units.

-> Tx

-> UTxO

-> Either TransactionValidityError (Map ScriptWitnessIndex (Either ScriptExecutionError ExecutionUnits))

Thin wrapper around evaluateTransactionExecutionUnits, which uses fixtures for system start, era history, protocol parameters, but allows to configure max ExecutionUnits. See pparams.

type EvaluationReport = Map ScriptWitnessIndex (Either ScriptExecutionError ExecutionUnits)

renderEvaluationReportFailures :: EvaluationReport -> Text

estimateMinFee :: Tx -> EvaluationReport -> Lovelace

Estimate minimum fee for given transaction and evaluated redeemers. Instead of using the budgets from the transaction (which might are usually set to 0 until balancing), this directly computes the fee from transaction size and the units of the EvaluationReport. Note that this function only provides a rough estimate using this modules' pparams and likely under-estimates cost as we have no witnesses on this Tx.

renderScriptExecutionError :: ScriptExecutionError -> Text

Fixtures

pparams :: ProtocolParameters

Current mainchain protocol parameters.

maxTxSize :: Natural

Max transaction size of the current pparams.

maxTxExecutionUnits :: ExecutionUnits

Max transaction execution unit budget of the current params.

maxMem :: Natural

Max memory and cpu units of the current pparams.

maxCpu :: Natural

Max memory and cpu units of the current pparams.

eraHistory :: EraHistory CardanoMode

An artifical era history comprised by a single never ending (forking) era, with fixed epochSize and slotLength.

eraHistoryWithHorizonAt :: SlotNo -> EraHistory CardanoMode

An era history with a single era which will end at some point.

A "real" EraHistory received from the cardano-node will have the eraEnd at a known or earliest possible end of the current era + a safe zone.

See EraParams for details.

NOTE: This era is using not so realistic epoch sizes of 1 and sets a slot length of 1

epochInfo :: Monad m => EpochInfo m

epochSize :: EpochSize

slotLength :: SlotLength

systemStart :: SystemStart

genPointInTime :: Gen (SlotNo, UTCTime)

genPointInTimeBefore :: UTCTime -> Gen (SlotNo, UTCTime)

genPointInTimeAfter :: UTCTime -> Gen (SlotNo, UTCTime)

slotNoFromUTCTime :: UTCTime -> SlotNo

Using hard-coded systemStart and slotLength, do not use in production!

slotNoToUTCTime :: HasCallStack => SlotNo -> UTCTime

Using hard-coded defaults above. Fails for slots past epoch boundaries.