Hydra.Ledger.Cardano.Evaluate

Description

Simplified interface to phase-2 validation of transaction, eg. evaluation of Plutus scripts.

The evaluateTx function simplifies the call to underlying Plutus providing execution report using pre-canned PParams. This should only be used for testing or benchmarking purpose as the real evaluation parameters are set when the Hydra node starts.

NOTE: The reason this module is here instead of part of `test/` directory is to be used in tx-cost executable.

Documentation

type RedeemerReport = Map RdmrPtr (Either (TransactionScriptFailure StandardCrypto) ExUnits)

evaluateTx :: Tx -> UTxO -> Either (TranslationError StandardCrypto) RedeemerReport

evaluateTx'

Arguments

:: ExecutionUnits
Max tx execution units.

-> Tx

-> UTxO

-> Either (TranslationError StandardCrypto) RedeemerReport

costModels :: Array Language CostModel

Cost models used in evaluateTx.

pparams :: PParams LedgerEra

Current mainchain cost parameters.

maxTxSize :: Natural

Max transaction size of the current pparams.

maxTxExecutionUnits :: ExecutionUnits

Max transaction execution unit budget of the current params.

maxMem :: Fixed E2

Max memory and cpu units of the current pparams.

maxCpu :: Fixed E2

Max memory and cpu units of the current pparams.

epochInfo :: Monad m => EpochInfo m

slotLength :: SlotLength

systemStart :: SystemStart

genPointInTime :: Gen (SlotNo, POSIXTime)

genPointInTimeBefore :: POSIXTime -> Gen (SlotNo, POSIXTime)

genPointInTimeAfter :: POSIXTime -> Gen (SlotNo, POSIXTime)

slotNoFromPOSIXTime :: POSIXTime -> SlotNo

Using hard-coded systemStart and slotLength, do not use in production!

slotNoToPOSIXTime :: HasCallStack => SlotNo -> POSIXTime

Using hard-coded defaults above. Fails for slots past epoch boundaries.