Hydra.Ledger.Simple

Description

A mock implementation of a ledger slightly less dumb than Mock.

This implementation of a Ledger adds a bit more logic in order to:

Be able to have a representation of UTxO closer to what a real-life eUTxO would be, so that we can distinguish it from other components of the ledger,

Be able to have transactions validated against the current state of the ledger, so that one transaction could at some point be invalid, then becomes valid because some inputs it consumes is now available.

Documentation

data SimpleTx

Simple transaction. A transaction is a SimpleId, a list of inputs and a list of outputs.

Constructors

SimpleTx

Fields

txSimpleId :: SimpleId

txInputs :: Utxo SimpleTx

txOutputs :: Utxo SimpleTx

type SimpleId = Integer

newtype TxIn

An identifier for a single output of a SimpleTx.

Constructors

TxIn

Fields

unTxIn :: Integer

simpleLedger :: Ledger SimpleTx

Builders

utxoRef :: Integer -> Utxo SimpleTx

utxoRefs :: [Integer] -> Utxo SimpleTx

aValidTx :: Integer -> SimpleTx

Generators

listOfCommittedUtxos :: Integer -> Gen [Utxo SimpleTx]

genSequenceOfValidTransactions :: Utxo SimpleTx -> Gen [SimpleTx]