Hydra.Logging.Messages

Description

Aggregates all tracing messages in a single type.

This module provides a central point where top-level traced messages are grouped. This is useful for traces consumers that will need to do something specific depending on various tracing messages, eg. monitoring and metrics collection.

Documentation

data HydraLog tx net

Constructors

DirectChain

Fields

directChain :: DirectChainLog

APIServer

Fields

api :: APIServerLog

Network

Fields

network :: net

Node

Fields

node :: HydraNodeLog tx