Hydra.Logging.Monitoring

Description

Provides Prometheus-based metrics server based on Tracer collection.

Documentation

withMonitoring :: (MonadIO m, MonadAsync m) => Maybe PortNumber -> Tracer m (HydraLog tx net) -> (Tracer m (HydraLog tx net) -> m ()) -> m ()

Wraps a monadic action using a Tracer and capture metrics based on traces. Given a portNumber, this wrapper starts a Prometheus-compliant server on this port. This is a no-op if given Nothing. This function is not polymorphic over the type of messages because it needs to understand them in order to provide meaningful metrics.