Hydra.Network.Heartbeat

Description

A naive implementation of an application-level Heartbeat This module exposes a Component withHeartbeat than can be used to wrap another NetworkComponent and piggy-back on it to send and propagate HeartbeatMessages.

Its current behavior is very simple: When it starts, it sends a Heartbeat message with its own identifier every 500ms, until the wrapped component sends another message. Heartbeat messages received from other components are simply propagated to the wrapped component.

Documentation

data HeartbeatState

Constructors

SendHeartbeat

StopHeartbeat

withHeartbeat :: (MonadAsync m, MonadDelay m) => Party -> NetworkComponent m (HydraMessage tx) -> NetworkComponent m (HydraMessage tx)

Wrap a NetworkComponent and handle sending/receiving of heartbeats.

checkMessages :: MonadSTM m => Network m (HydraMessage tx) -> TVar m HeartbeatState -> Network m (HydraMessage tx)

sendHeartbeatFor :: (MonadDelay m, MonadSTM m) => Party -> TVar m HeartbeatState -> Network m (HydraMessage tx) -> m ()