Hydra.Network.Heartbeat

Description

An implementation of an application-level failure detector. This module exposes a Component withHeartbeat than can be used to wrap another NetworkComponent and piggy-back on it to send and propagate Heartbeat messages and detect other parties' liveness.

It is inspired by the Increasing timeout algorithms from the book Introduction to Reliable and Secure Distributed Programming by Cachin et al. which is an Eventually Perfect Failure Detector suitable for partially synchronous network settings. It has the following behaviour:

It broadcasts a Ping to other parties through the underlying Network implementation if the last message has been sent more than 3s ago

When receiving messages from other parties, it records reception time and notifies underlying node with a Connected message

If new messages are received from alive parties before 3s timeout expires no new Connected message is sent * If main thread detects that a formerly alive party has not been seen for more than 3s, it is marked as suspected and a Disconnected message is sent to the node.

Documentation

data HeartbeatState

Constructors

HeartbeatState

Fields

alive :: Map Host Time

The map of known Connected parties with the last time they've been "seen". This is updated when we see a message from another Host

suspected :: Set Host

The set of known parties which might be Disconnected This is updated after some time no message has been received from a Host.

lastSent :: Maybe Time

The timestamp of the last sent message.

initialHeartbeatState :: HeartbeatState

data Heartbeat msg

Constructors

Data Host msg

Ping Host

heartbeatDelay :: DiffTime

Delay between each heartbeat check.

livenessDelay :: DiffTime

Maximal delay between expected and sent heartbeats.

withHeartbeat :: (MonadAsync m, MonadDelay m, MonadMonotonicTime m) => Host -> NetworkComponent m (Heartbeat (Message tx)) a -> NetworkComponent m (Message tx) a

Wrap a NetworkComponent and handle sending/receiving of heartbeats.

updateStateFromIncomingMessages :: (MonadSTM m, MonadMonotonicTime m) => TVar m HeartbeatState -> NetworkCallback (Message tx) m -> NetworkCallback (Heartbeat (Message tx)) m

updateStateFromOutgoingMessages :: (MonadSTM m, MonadMonotonicTime m) => Host -> TVar m HeartbeatState -> Network m (Heartbeat (Message msg)) -> Network m (Message msg)

updateLastSent :: MonadSTM m => TVar m HeartbeatState -> Time -> m ()

checkHeartbeatState :: (MonadDelay m, MonadSTM m, MonadMonotonicTime m) => Host -> TVar m HeartbeatState -> Network m (Heartbeat (Message msg)) -> m ()

shouldSendHeartbeat :: Time -> HeartbeatState -> Bool

checkRemoteParties :: (MonadDelay m, MonadSTM m, MonadMonotonicTime m) => TVar m HeartbeatState -> NetworkCallback (Message msg) m -> m ()

updateSuspected :: MonadSTM m => TVar m HeartbeatState -> Time -> m (Set Host)

checkTimeout :: (DiffTime -> DiffTime) -> Time -> Time -> Bool