Hydra.Network.Ouroboros

Description

Ouroboros-based implementation of Network interface. This implements a dumb FireForget protocol and maintains one connection to each peer. Contrary to other protocols implemented in Ouroboros, this is a push-based protocol.

Documentation

withOuroborosNetwork :: forall msg. (ToCBOR msg, FromCBOR msg) => Tracer IO (WithHost (TraceOuroborosNetwork msg)) -> Host -> [Host] -> NetworkCallback msg IO -> (Network IO msg -> IO ()) -> IO ()

withIOManager :: WithIOManager

data TraceOuroborosNetwork msg

data PortNumber

data Host

Constructors

Host

Fields

hostname :: Text

port :: PortNumber

type NetworkComponent m msg a = NetworkCallback msg m -> (Network m msg -> m a) -> m a

A type tying both inbound and outbound messages sending in a single Component.

type NetworkCallback msg m = msg -> m ()

Handle to interface for inbound messages.

newtype Network m msg

Handle to interface with the hydra network and send messages "off chain".

Constructors

Network

Fields

broadcast :: msg -> m ()

Send a msg to the whole hydra network.