Hydra.Node

Description

Top-level module to run a single Hydra node.

Checkout Hydra Documentation for some details about the overall architecture of the Node.

Environment Handling

initEnvironment :: RunOptions -> IO Environment

Intialize the Environment from command line options.

checkHeadState :: MonadThrow m => Tracer m (HydraNodeLog tx) -> Environment -> HeadState tx -> m ()

Checks that command line options match a given HeadState. This function takes Environment because it is derived from RunOptions via initEnvironment.

Throws: ParameterMismatch when state not matching the environment.

Create and run a hydra node

data DraftHydraNode tx m

A draft version of the HydraNode that holds state, but is not yet connected (see connect). This is commonly created by the hydrate smart constructor.

Constructors

DraftHydraNode

Fields

tracer :: Tracer m (HydraNodeLog tx)

env :: Environment

ledger :: Ledger tx

nodeState :: NodeState tx m

inputQueue :: InputQueue m (Input tx)

eventSource :: EventSource (StateEvent tx) m

eventSinks :: [EventSink (StateEvent tx) m]

chainStateHistory :: ChainStateHistory tx

hydrate :: (MonadDelay m, MonadLabelledSTM m, MonadAsync m, MonadThrow m, IsChainState tx) => Tracer m (HydraNodeLog tx) -> Environment -> Ledger tx -> ChainStateType tx -> EventSource (StateEvent tx) m -> [EventSink (StateEvent tx) m] -> m (DraftHydraNode tx m)

Hydrate a DraftHydraNode by loading events from source, re-aggregate node state and sending events to sinks while doing so.

wireChainInput :: DraftHydraNode tx m -> ChainEvent tx -> m ()

wireClientInput :: DraftHydraNode tx m -> ClientInput tx -> m ()

wireNetworkInput :: DraftHydraNode tx m -> NetworkCallback (NetworkEvent (Message tx)) m

connect :: Monad m => Chain tx m -> Network m (Message tx) -> Server tx m -> DraftHydraNode tx m -> m (HydraNode tx m)

Connect chain, network and API to a hydrated DraftHydraNode to get a fully connected HydraNode.

data HydraNode tx m

Fully connected hydra node with everything wired in.

Constructors

HydraNode

Fields

tracer :: Tracer m (HydraNodeLog tx)

env :: Environment

ledger :: Ledger tx

nodeState :: NodeState tx m

inputQueue :: InputQueue m (Input tx)

eventSource :: EventSource (StateEvent tx) m

eventSinks :: [EventSink (StateEvent tx) m]

oc :: Chain tx m

hn :: Network m (Message tx)

server :: Server tx m

runHydraNode :: (MonadCatch m, MonadAsync m, IsChainState tx) => HydraNode tx m -> m ()

stepHydraNode :: (MonadCatch m, MonadAsync m, IsChainState tx) => HydraNode tx m -> m ()

waitDelay :: DiffTime

The time to wait between re-enqueuing a Wait outcome from HeadLogic.

processNextInput :: IsChainState tx => HydraNode tx m -> Input tx -> STM m (Outcome tx)

Monadic interface around update.

processStateChanges :: MonadSTM m => HydraNode tx m -> [StateChanged tx] -> m ()

processEffects :: (MonadAsync m, MonadCatch m, IsChainState tx) => HydraNode tx m -> Tracer m (HydraNodeLog tx) -> Word64 -> [Effect tx] -> m ()

Manage state

data NodeState tx m

Handle to access and modify the state in the Hydra Node.

Constructors

NodeState

Fields

modifyHeadState :: forall a. (HeadState tx -> (a, HeadState tx)) -> STM m a

queryHeadState :: STM m (HeadState tx)

getNextEventId :: STM m EventId

createNodeState

Arguments

:: MonadLabelledSTM m

=> Maybe EventId
Last seen EventId.

-> HeadState tx

-> m (NodeState tx m)

Initialize a new NodeState.

Logging

data HydraNodeLog tx

Constructors

BeginInput

Fields

by :: Party

inputId :: Word64

input :: Input tx

EndInput

Fields

by :: Party

inputId :: Word64

BeginEffect

Fields

by :: Party

inputId :: Word64

effectId :: Word32

effect :: Effect tx

EndEffect

Fields

by :: Party

inputId :: Word64

effectId :: Word32

LogicOutcome

Fields

by :: Party

outcome :: Outcome tx

DroppedFromQueue

Fields

inputId :: Word64

input :: Input tx

LoadedState

Fields

numberOfEvents :: Word64

Misconfiguration

Fields

misconfigurationErrors :: [ParamMismatch]