Hydra.Node

Description

Top-level module to run a single Hydra node.

Checkout Hydra Documentation for some details about the overall architecture of the Node.

Environment Handling

initEnvironment :: RunOptions -> IO Environment

Create and run a hydra node

data HydraNode tx m

Main handle of a hydra node where all layers are tied together.

Constructors

HydraNode

Fields

eq :: EventQueue m (Event tx)

hn :: Network m (Message tx)

nodeState :: NodeState tx m

oc :: Chain tx m

server :: Server tx m

ledger :: Ledger tx

env :: Environment

persistence :: Persistence (HeadState tx) m

data HydraNodeLog tx

Constructors

BeginEvent

Fields

by :: Party

eventId :: Word64

event :: Event tx

EndEvent

Fields

by :: Party

eventId :: Word64

BeginEffect

Fields

by :: Party

eventId :: Word64

effectId :: Word32

effect :: Effect tx

EndEffect

Fields

by :: Party

eventId :: Word64

effectId :: Word32

LogicOutcome

Fields

by :: Party

outcome :: Outcome tx

runHydraNode :: (MonadCatch m, MonadAsync m, IsChainState tx) => Tracer m (HydraNodeLog tx) -> HydraNode tx m -> m ()

stepHydraNode :: (MonadCatch m, MonadAsync m, IsChainState tx) => Tracer m (HydraNodeLog tx) -> HydraNode tx m -> m ()

waitDelay :: DiffTime

The time to wait between re-enqueuing a Wait outcome from HeadLogic.

processNextEvent :: IsChainState tx => HydraNode tx m -> Event tx -> STM m (Outcome tx)

Monadic interface around update.

processEffect :: (MonadAsync m, MonadCatch m, IsChainState tx) => HydraNode tx m -> Tracer m (HydraNodeLog tx) -> (Word64, Word32) -> Effect tx -> m ()

Manage state

data NodeState tx m

Handle to access and modify the state in the Hydra Node.

Constructors

NodeState

Fields

modifyHeadState :: forall a. (HeadState tx -> (a, HeadState tx)) -> STM m a

queryHeadState :: STM m (HeadState tx)

createNodeState :: MonadLabelledSTM m => HeadState tx -> m (NodeState tx m)

Initialize a new NodeState.