Hydra.Node

Description

Top-level module to run a single Hydra node.

Create and run a hydra node

data HydraNode tx m

Constructors

HydraNode

Fields

eq :: EventQueue m (Event tx)

hn :: Network m (HydraMessage tx)

hh :: HydraHead tx m

oc :: OnChain tx m

sendResponse :: ClientResponse tx -> m ()

env :: Environment

data HydraNodeLog tx

Constructors

ErrorHandlingEvent Party (Event tx) (LogicError tx)

ProcessingEvent Party (Event tx)

ProcessedEvent Party (Event tx)

ProcessingEffect Party (Effect tx)

ProcessedEffect Party (Effect tx)

handleClientRequest :: HydraNode tx m -> ClientRequest tx -> m ()

handleChainTx :: HydraNode tx m -> OnChainTx tx -> m ()

handleMessage :: HydraNode tx m -> HydraMessage tx -> m ()

runHydraNode :: (MonadThrow m, MonadAsync m, MonadTimer m, Tx tx, Ord tx) => Tracer m (HydraNodeLog tx) -> HydraNode tx m -> m ()

processNextEvent :: (Tx tx, Ord tx, MonadSTM m) => HydraNode tx m -> Event tx -> m (Either (LogicError tx) [Effect tx])

Monadic interface around update.

processEffect :: (MonadAsync m, MonadTimer m, MonadThrow m) => HydraNode tx m -> Tracer m (HydraNodeLog tx) -> Effect tx -> m ()

Some general event queue from which the Hydra head is "fed"

data EventQueue m e

The single, required queue in the system from which a hydra head is "fed". NOTE(SN): this probably should be bounded and include proper logging NOTE(SN): handle pattern, but likely not required as there is no need for an alternative implementation

Constructors

EventQueue

Fields

putEvent :: e -> m ()

nextEvent :: m e

createEventQueue :: MonadSTM m => m (EventQueue m e)

HydraHead handle to manage a single hydra head state concurrently

data HydraHead tx m

Handle to access and modify a Hydra Head's state.

Constructors

HydraHead

Fields

modifyHeadState :: forall a. (HeadState tx -> (a, HeadState tx)) -> STM m a

ledger :: Ledger tx

queryHeadState :: HydraHead tx m -> STM m (HeadState tx)

putState :: HydraHead tx m -> HeadState tx -> STM m ()

createHydraHead :: MonadSTM m => HeadState tx -> Ledger tx -> m (HydraHead tx m)

OnChain handle to abstract over chain access

data ChainError

Constructors

ChainError

newtype OnChain tx m

Handle to interface with the main chain network

Constructors

OnChain

Fields

postTx :: MonadThrow m => OnChainTx tx -> m ()

Construct and send a transaction to the main chain corresponding to the given OnChainTx event. Does at least throw ChainError.