Hydra.Node

Description

Top-level module to run a single Hydra node.

Environment Handling

initEnvironment :: Options -> IO Environment

Create and run a hydra node

data HydraNode tx m

Constructors

HydraNode

Fields

eq :: EventQueue m (Event tx)

hn :: Network m (Message tx)

hh :: HydraHead tx m

oc :: Chain tx m

server :: Server tx m

env :: Environment

data HydraNodeLog tx

Constructors

ErrorHandlingEvent

Fields

emitter :: Party

event :: Event tx

reason :: LogicError tx

ProcessingEvent

Fields

emitter :: Party

event :: Event tx

ProcessedEvent

Fields

emitter :: Party

event :: Event tx

ProcessingEffect

Fields

emitter :: Party

effect :: Effect tx

ProcessedEffect

Fields

emitter :: Party

effect :: Effect tx

handleClientInput :: HydraNode tx m -> ClientInput tx -> m ()

handleChainTx :: HydraNode tx m -> OnChainTx tx -> m ()

handleMessage :: HydraNode tx m -> Message tx -> m ()

runHydraNode :: (MonadThrow m, MonadAsync m, MonadTimer m, Tx tx) => Tracer m (HydraNodeLog tx) -> HydraNode tx m -> m ()

processNextEvent :: (Tx tx, MonadSTM m) => HydraNode tx m -> Event tx -> m (Either (LogicError tx) [Effect tx])

Monadic interface around update.

processEffect :: (MonadAsync m, MonadTimer m, MonadThrow m) => HydraNode tx m -> Tracer m (HydraNodeLog tx) -> Effect tx -> m ()

Some general event queue from which the Hydra head is "fed"

data EventQueue m e

The single, required queue in the system from which a hydra head is "fed". NOTE(SN): this probably should be bounded and include proper logging NOTE(SN): handle pattern, but likely not required as there is no need for an alternative implementation

Constructors

EventQueue

Fields

putEvent :: e -> m ()

nextEvent :: m e

createEventQueue :: MonadSTM m => m (EventQueue m e)

HydraHead handle to manage a single hydra head state concurrently

data HydraHead tx m

Handle to access and modify a Hydra Head's state.

Constructors

HydraHead

Fields

modifyHeadState :: forall a. (HeadState tx -> (a, HeadState tx)) -> STM m a

ledger :: Ledger tx

queryHeadState :: HydraHead tx m -> STM m (HeadState tx)

putState :: HydraHead tx m -> HeadState tx -> STM m ()

createHydraHead :: MonadSTM m => HeadState tx -> Ledger tx -> m (HydraHead tx m)