Hydra.Node.Network

Description

Concrete Network stack used in a hydra-node.

This module provides a withNetwork function which is the composition of several layers in order to provide various capabilities:

withAuthentication handles messages' authentication and signature verification

withEtcdNetwork uses an etcd cluster to implement reliable broadcast

The following diagram details the various types of messages each layer is exchanging with its predecessors and successors.

          ▲
          │                        │
      Authenticated msg           msg
          │                        │
          │                        │
┌─────────┼────────────────────────▼──────┐
│                                         │
│               Authenticate              │
│                                         │
└─────────▲────────────────────────┼──────┘
          │                        │
          │                        │
         msg                      msg
          │                        │
┌─────────┼────────────────────────▼──────┐
│                                         │
│                   Etcd                  │
│                                         │
└─────────▲────────────────────────┼──────┘
          │                        │
          │        (bytes)         │
          │                        ▼

Documentation

data NetworkConfiguration

Configuration for a Node network layer.

Constructors

NetworkConfiguration

Fields

persistenceDir :: FilePath

Persistence directory

signingKey :: SigningKey HydraKey

This node's signing key. This is used to sign messages sent to peers.

otherParties :: [Party]

The list of peers Party known to this node.

listen :: Host

Address to listen on for incoming connections.

advertise :: Host

Address to advertise to peers.

peers :: [Host]

Addresses and ports of remote peers.

nodeId :: NodeId

This node's id.

whichEtcd :: WhichEtcd

Whether to use the system etcd (on the path) or the embedded one.

withNetwork

Arguments

:: forall tx. IsTx tx

=> Tracer IO NetworkLog
Tracer to use for logging messages.

-> NetworkConfiguration
The network configuration

-> NetworkComponent IO (Authenticated (Message tx)) (Message tx) ()
Produces a NetworkComponent that can send msg and consumes Authenticated msg.

Starts the network layer of a node, passing configured Network to its continuation.

data NetworkLog