Hydra.Options

Documentation

data Command

Constructors

Run RunOptions

Publish PublishOptions

GenHydraKey GenerateKeyPair

data DirectChainConfig

Constructors

DirectChainConfig

Fields

networkId :: NetworkId

Network identifer to which we expect to connect.

nodeSocket :: SocketPath

Path to a domain socket used to connect to the server.

hydraScriptsTxId :: [TxId]

Identifier of transaction holding the hydra scripts to use.

cardanoSigningKey :: FilePath

Path to the cardano signing key of the internal wallet.

cardanoVerificationKeys :: [FilePath]

Paths to other node's verification keys.

startChainFrom :: Maybe ChainPoint

Point at which to start following the chain.

contestationPeriod :: ContestationPeriod

depositDeadline :: DepositDeadline

Deadline to detect deposit tx on-chain.

data OfflineChainConfig

Constructors

OfflineChainConfig

Fields

offlineHeadSeed :: HeadSeed

Manually provided seed of the offline head.

initialUTxOFile :: FilePath

Path to a json encoded starting UTxO for the offline-mode head.

ledgerGenesisFile :: Maybe FilePath

Path to a shelley genesis file with slot lengths used by the offline-mode chain.

data RunOptions

Constructors

RunOptions

Fields

verbosity :: Verbosity

nodeId :: NodeId

listen :: Host

advertise :: Maybe Host

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

tlsCertPath :: Maybe FilePath

tlsKeyPath :: Maybe FilePath

monitoringPort :: Maybe PortNumber

hydraSigningKey :: FilePath

hydraVerificationKeys :: [FilePath]

persistenceDir :: FilePath

chainConfig :: ChainConfig

ledgerConfig :: LedgerConfig

data ChainConfig

Constructors

Offline OfflineChainConfig

Direct DirectChainConfig

data InvalidOptions

Constructors

MaximumNumberOfPartiesExceeded

CardanoAndHydraKeysMissmatch

newtype LedgerConfig

Constructors

CardanoLedgerConfig

Fields

cardanoLedgerProtocolParametersFile :: FilePath

newtype GenerateKeyPair

Constructors

GenerateKeyPair

Fields

outputFile :: FilePath

data PublishOptions

Constructors

PublishOptions

Fields

chainBackend :: ChainBackend

publishSigningKey :: FilePath

data ChainBackend

Constructors

DirectBackend

Fields

publishNetworkId :: NetworkId

publishNodeSocket :: SocketPath

BlockfrostBackend

Fields

projectPath :: FilePath

toArgs :: RunOptions -> [String]

Convert an Options instance into the corresponding list of command-line arguments.

This is useful in situations where one wants to programatically define Options, providing some measure of type safety, without having to juggle with strings.

defaultContestationPeriod :: ContestationPeriod

defaultDepositDeadline :: DepositDeadline

validateRunOptions :: RunOptions -> Either InvalidOptions ()

Validate cmd line arguments for hydra-node and check if they make sense before actually running the node. Rules we apply: - Check if number of parties is bigger than our hardcoded limit (by looking at loaded hydra or cardano keys and comparing it to the maximumNumberOfParties) - Check that number of loaded hydra keys match with the number of loaded cardano keys (by comparing lengths of the two lists)

genChainPoint :: Gen ChainPoint

hydraNodeVersion :: Version

commandParser :: Parser Command

runOptionsParser :: Parser RunOptions

Parser for running the cardano-node with all its RunOptions.

publishOptionsParser :: Parser PublishOptions

outputFileParser :: Parser FilePath

defaultPublishOptions :: PublishOptions

Default options as they should also be provided by runOptionsParser.

defaultDirectBackend :: ChainBackend

chainBackendParser :: Parser ChainBackend

cardanoSigningKeyFileParser :: Parser FilePath

genFilePath :: String -> Gen FilePath

genDirPath :: Gen FilePath

defaultRunOptions :: RunOptions

Default options as they should also be provided by runOptionsParser.

defaultDirectChainConfig :: DirectChainConfig

defaultLedgerConfig :: LedgerConfig

verbosityParser :: Parser Verbosity

nodeIdParser :: Parser NodeId

listenParser :: Parser Host

advertiseParser :: Parser Host

peerParser :: Parser Host

apiHostParser :: Parser IP

apiPortParser :: Parser PortNumber

tlsCertPathParser :: Parser FilePath

tlsKeyPathParser :: Parser FilePath

monitoringPortParser :: Parser PortNumber

hydraSigningKeyFileParser :: Parser FilePath

hydraVerificationKeyFileParser :: Parser FilePath

persistenceDirParser :: Parser FilePath

chainConfigParser :: Parser ChainConfig

ledgerConfigParser :: Parser LedgerConfig

directChainConfigParser :: Parser DirectChainConfig

offlineChainConfigParser :: Parser OfflineChainConfig

networkIdParser :: Parser NetworkId

nodeSocketParser :: Parser SocketPath

blockfrostProjectPathParser :: Parser FilePath

cardanoLedgerProtocolParametersParser :: Parser FilePath

offlineHeadSeedParser :: Parser HeadSeed

initialUTxOFileParser :: Parser FilePath

ledgerGenesisFileParser :: Parser (Maybe FilePath)

hydraScriptsTxIdsParser :: Parser [TxId]

hydraScriptsTxIdParser :: Parser TxId

cardanoVerificationKeyFileParser :: Parser FilePath

startChainFromParser :: Parser ChainPoint

contestationPeriodParser :: Parser ContestationPeriod

depositDeadlineParser :: Parser DepositDeadline

hydraNodeCommand :: ParserInfo Command

parseHydraCommand :: IO Command

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraCommandFromArgs :: [String] -> ParserResult Command

Pure parsing of Option from a list of arguments.

toArgApiPort :: PortNumber -> [String]

toArgNetworkId :: NetworkId -> [String]

toArgNodeSocket :: SocketPath -> [String]

data ParserResult a

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult

renderFailure :: ParserFailure ParserHelp -> String -> (String, ExitCode)

Orphan instances

Arbitrary IP

Instance details

Methods

arbitrary :: Gen IP

shrink :: IP -> [IP]