Hydra.Options

Documentation

data Options

Constructors

Options

Fields

verbosity :: Verbosity

nodeId :: Natural

host :: IP

port :: PortNumber

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

monitoringPort :: Maybe PortNumber

hydraSigningKey :: FilePath

hydraVerificationKeys :: [FilePath]

chainConfig :: ChainConfig

ledgerConfig :: LedgerConfig

data ChainConfig

Constructors

DirectChainConfig

Fields

networkId :: NetworkId

nodeSocket :: FilePath

cardanoSigningKey :: FilePath

cardanoVerificationKeys :: [FilePath]

data LedgerConfig

Constructors

CardanoLedgerConfig

Fields

cardanoLedgerGenesisFile :: FilePath

cardanoLedgerProtocolParametersFile :: FilePath

parseHydraOptions :: IO Options

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraOptionsFromString :: [String] -> ParserResult Options

Pure parsing of Option from a list of arguments.

getParseResult :: ParserResult a -> Maybe a

Extract the actual result from a ParserResult value.

This function returns Nothing in case of errors. Possible error messages or completion actions are simply discarded.

If you want to display error messages and invoke completion actions appropriately, use handleParseResult instead.

data ParserResult a

Result of execParserPure.

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult