Hydra.Options

Documentation

data Options

Constructors

Options

Fields

verbosity :: Verbosity

nodeId :: Natural

host :: IP

port :: PortNumber

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

monitoringPort :: Maybe PortNumber

me :: FilePath

parties :: [FilePath]

mockChain :: MockChain

parseHydraOptions :: IO Options

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraOptionsFromString :: [String] -> ParserResult Options

Pure parsing of Option from a list of arguments.

getParseResult :: ParserResult a -> Maybe a

defaultOptions :: Options

data ParserResult a

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult