Hydra.Options

Documentation

data Options

Constructors

Options

Fields

verbosity :: Verbosity

nodeId :: Natural

host :: IP

port :: PortNumber

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

monitoringPort :: Maybe PortNumber

hydraSigningKey :: FilePath

hydraVerificationKeys :: [FilePath]

chainConfig :: ChainConfig

ledgerConfig :: LedgerConfig

data ChainConfig

Constructors

DirectChainConfig

Fields

networkId :: NetworkId

nodeSocket :: FilePath

cardanoSigningKey :: FilePath

cardanoVerificationKeys :: [FilePath]

startChainFrom :: Maybe ChainPoint

data LedgerConfig

Constructors

CardanoLedgerConfig

Fields

cardanoLedgerGenesisFile :: FilePath

cardanoLedgerProtocolParametersFile :: FilePath

parseHydraOptions :: IO Options

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraOptionsFromString :: [String] -> ParserResult Options

Pure parsing of Option from a list of arguments.

getParseResult :: ParserResult a -> Maybe a

data ParserResult a

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult

toArgs :: Options -> [String]

Convert an Options instance into the corresponding list of command-line arguments.

This is useful in situations where one wants to programatically define Options, providing some measure of type safety, without having to juggle with strings.

defaultOptions :: Options

defaultLedgerConfig :: LedgerConfig

defaultChainConfig :: ChainConfig