Hydra.Party

Description

Types and functions revolving around a Hydra Party. That is, a participant in a Hydra Head, which signs transactions or snapshots in the Hydra protocol. Hence, this module also includes functions to sign and verify data given a Party.

Documentation

newtype Party

Identifies a party in a Hydra head by it's VerificationKey.

Constructors

Party

Fields

vkey :: VerificationKey

deriveParty :: SigningKey -> Party

Hydra keys

type VerificationKey = VerKeyDSIGN MockDSIGN

showVerificationKey :: VerificationKey -> Text

type SigningKey = SignKeyDSIGN MockDSIGN

generateKey :: Integer -> SigningKey

sign :: SignableRepresentation a => SigningKey -> a -> Signed a

verify :: SignableRepresentation a => Signed a -> Party -> a -> Bool

newtype MultiSigned a

Naiive mult-signatures.

Constructors

MultiSigned

Fields

multiSignature :: [Signed a]

aggregate :: [Signed a] -> MultiSigned a

toPlutusSignatures :: MultiSigned a -> [Signature]

aggregateInOrder :: Map Party (Signed a) -> [Party] -> MultiSigned a

newtype Signed a

Signature of a

Constructors

UnsafeSigned ByteString

toPlutusSignature :: Signed a -> Signature

Helpers

decodeBase16' :: MonadFail f => Text -> f ByteString

Orphan instances

ToJSON VerificationKey

Instance details

Methods

toJSON :: VerificationKey -> Value

toEncoding :: VerificationKey -> Encoding

toJSONList :: [VerificationKey] -> Value

toEncodingList :: [VerificationKey] -> Encoding

FromJSON VerificationKey

Instance details

Methods

parseJSON :: Value -> Parser VerificationKey

parseJSONList :: Value -> Parser [VerificationKey]