Hydra.Party

Description

Types and functions revolving around a Hydra Party. That is, a participant in a Hydra Head, which signs transactions or snapshots in the Hydra protocol. Hence, this module also includes functions to sign and verify data given a Party.

Documentation

data Party

Identifies a party in a Hydra head by it's VerificationKey.

Constructors

Party

Fields

alias :: Maybe Text

vkey :: VerificationKey

type VerificationKey = VerKeyDSIGN MockDSIGN

showVerificationKey :: VerificationKey -> Text

type SigningKey = SignKeyDSIGN MockDSIGN

deriveParty :: SigningKey -> Party

generateKey :: Integer -> SigningKey

sign :: SignableRepresentation a => SigningKey -> a -> Signed a

verify :: SignableRepresentation a => Signed a -> Party -> a -> Bool

newtype Signed a

Signature of a

Constructors

UnsafeSigned (SigDSIGN MockDSIGN)

decodeBase16' :: MonadFail f => Text -> f ByteString

Orphan instances

ToJSON VerificationKey

Instance details

Methods

toJSON :: VerificationKey -> Value

toEncoding :: VerificationKey -> Encoding

toJSONList :: [VerificationKey] -> Value

toEncodingList :: [VerificationKey] -> Encoding

FromJSON VerificationKey

Instance details

Methods

parseJSON :: Value -> Parser VerificationKey

parseJSONList :: Value -> Parser [VerificationKey]