Hydra.Party

Description

Types and functions revolving around a Hydra Party. That is, a participant in a Hydra Head, which signs transactions or snapshots in the Hydra protocol.

Documentation

newtype Party

Identifies a party in a Hydra head by it's VerificationKey.

Constructors

Party

Fields

vkey :: VerificationKey

deriveParty :: SigningKey -> Party

Get the Party given some Hydra SigningKey.

partyToChain :: Party -> Party

Convert "high-level" Party to the "low-level" representation as used on-chain. See Party for an explanation why this is a distinct type.

partyFromChain :: MonadFail m => Party -> m Party

Retrieve the "high-level" 'Party from the "low-level" on-chain representation. This can fail because of the lower type-safety used on-chain and a non-guaranteed verification key length. See Party for an explanation why this is a distinct type.