Hydra.Persistence

Description

Handles to save/load files across the hydra-node. We use a simple JSON encoding and two modes of operation to store things: Full and Incremental.

Documentation

newtype PersistenceException

Constructors

PersistenceException String

data Persistence a m

Handle to save and load files to/from disk using JSON encoding.

Constructors

Persistence

Fields

save :: ToJSON a => a -> m ()

load :: FromJSON a => m (Maybe a)

createPersistence :: (MonadIO m, MonadThrow m) => FilePath -> m (Persistence a m)

Initialize persistence handle for given type a at given file path.

data PersistenceIncremental a m

Handle to save incrementally and load files to/from disk using JSON encoding.

Constructors

PersistenceIncremental

Fields

append :: ToJSON a => a -> m ()

source :: FromJSON a => ConduitT () a (ResourceT m) ()

Stream all elements from the file.

loadAll :: (FromJSON a, MonadUnliftIO m) => PersistenceIncremental a m -> m [a]

Load all elements from persistence into a list. XXX: Deprecate this to avoid large memory usage.

createPersistenceIncremental :: forall a m. (MonadUnliftIO m, MonadThrow m, FromJSON a) => FilePath -> m (PersistenceIncremental a m)

Initialize persistence handle for given type a at given file path.

This instance of PersistenceIncremental is "thread-safe" in the sense that it prevents appending while a source is still running (while ResourceT is still not fully unwrapped.